% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{aux_kurtosis}
\alias{aux_kurtosis}
\title{title aux_kurtosis
description calculates the kurtosis
param trial number of trial
param prob probability of success
return corresponding value from the computed kurtosis}
\usage{
aux_kurtosis(trials, prob)
}
\description{
title aux_kurtosis
description calculates the kurtosis
param trial number of trial
param prob probability of success
return corresponding value from the computed kurtosis
}
