% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{aux_mean}
\alias{aux_mean}
\title{title aux_mean
description calculates the mean
param trial number of trials
param prob probability of success
return corresponding value from the computed mean}
\usage{
aux_mean(trials, prob)
}
\description{
title aux_mean
description calculates the mean
param trial number of trials
param prob probability of success
return corresponding value from the computed mean
}
