% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{aux_mode}
\alias{aux_mode}
\title{title aux_mode
description calculates the mode
param trial number of trial
param prob probability of success
return corresponding value from the computed mode}
\usage{
aux_mode(trials, prob)
}
\description{
title aux_mode
description calculates the mode
param trial number of trial
param prob probability of success
return corresponding value from the computed mode
}
