% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{aux_skewness}
\alias{aux_skewness}
\title{title aux_skewness
description calculates the skewness
param trial number of trial
param prob probability of success
return corresponding value from the computed skewness}
\usage{
aux_skewness(trials, prob)
}
\description{
title aux_skewness
description calculates the skewness
param trial number of trial
param prob probability of success
return corresponding value from the computed skewness
}
