% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{aux_variance}
\alias{aux_variance}
\title{title aux_variance
description calculates the variance
param trial number of trials
param prob probability of success
return corresponding value from the computed variance}
\usage{
aux_variance(trials, prob)
}
\description{
title aux_variance
description calculates the variance
param trial number of trials
param prob probability of success
return corresponding value from the computed variance
}
