% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n:}{number of trials total}

\item{k:}{number of successes}
}
\value{
number of combinations for k successes within n trials
}
\description{
calculates the number of combinations in which k successes can occur in n trials according to the formula n choose k = n! / (k! (n - k)!)
}
