% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative probabilities}
\usage{
bin_cumulative(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a data frame with the probability distribution
}
\description{
to create a dataframe of the binomial distribution
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
