% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials:}{number of total trials}

\item{prob:}{probability of success}
}
\value{
a data.frame with two classes -- c("bindis", "data.frame") -- with the probability distribution: successes in the 1st column, probability in the 2nd column
}
\description{
calculates using bin_probability the chance of getting k successes within n trials for all possible values of k (0 to n)
}
