% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of event}
}
\value{
returns a numeric kurtosis of the binomial distribution
}
\description{
The kurtosis of a binomial distribution given trials and prob
}
\examples{
# kurtosis of a binomial with 10 trials and a prob of 0.3
bin_kurtosis(10, 0.3)

}
