% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials:}{number of total trials}

\item{prob:}{probability of success}
}
\value{
numeric value representing mean given number of trials and prob of success
}
\description{
calculates mean of binomial scenario of n trials and p prob of success using aux_mean
}
