% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{bin_mode}
\alias{bin_mode}
\title{Mode of binomial distribution}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{(numeric) the number of trials}

\item{prob}{(numeric) the probability of success on each trial}
}
\value{
Mode of the binomial distribution with the given trials and probability parameters.
}
\description{
Compute mode of binomial distribution with the given trials and probability parameters.
}
