% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Probability of Success}
\usage{
bin_probability(k, n, p)
}
\arguments{
\item{k}{number of successes}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
probability of a certian number of successes in n trials
}
\description{
to calculate the probability of a certian number of successes in n trials
}
\examples{
bin_probability(2, 5, 0.5)
bin_probability(0:2, 5, 0.5)
bin_probability(55, 100, 0.45)
}
