% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_var}
\alias{bin_var}
\title{Variable}
\usage{
bin_var(trials, prob)
}
\arguments{
\item{prob}{probability of a success}

\item{n}{number of trials}
}
\value{
an object of class "binvar", a binomial random variable object
}
\description{
Classifies object as a binomial random variable with given parameters, also returns summary of the random variable
}
\examples{
#
}
