% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials:}{number of total trials}

\item{prob:}{probability of success}
}
\value{
an object of class "binvar" -- a binomial variable object
}
\description{
creates a binomial variable list of named elements -- trials for the number of trials and prob for the probability of success
}
