% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of event}
}
\value{
returns a numeric variance of the binomial distribution
}
\description{
The variance of a binomial distribution given trials and prob
}
\examples{
# mean of a binomial with 10 trials and a prob of 0.3
bin_variance(10, 0.3)

}
