% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bincum}
\alias{bincum}
\title{Cumulative Binomial Distribution}
\usage{
bincum(trials, prob)
}
\arguments{
\item{trials, }{the number of trials}

\item{prob, }{the probability of a success}
}
\value{
returns a data frame of class "bincum" of probability density
}
\description{
gives the area under the probability density function
}
\examples{
bincum(trials = 5L, prob = 0.5)
bincum(3L, .3)
plot.bincum(bincum(2,.4))
}
