% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bindis}
\alias{bindis}
\title{Binomial Distribution}
\usage{
bindis(trials, prob)
}
\arguments{
\item{trials, }{the number of trials}

\item{prob, }{the probability of a success}
}
\value{
returns a data frame of class "bindis" of probabilities
}
\description{
finds the probability per success for a number of trials
}
\examples{
bindis(trials = 5L, prob = 0.5)
bindis(trials = 3L, prob = 0.1)
plot.bindis(bindis(4,.1))
}
