% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{binvar}
\alias{binvar}
\title{Binomial Random Variable}
\usage{
binvar(trials, prob)
}
\arguments{
\item{trials, }{the number of trials}

\item{prob, }{the probability of a success}
}
\value{
returns a binvar object
}
\description{
Binomial Random Variable
}
\examples{
###ADD examples
}
