% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_prob}
\alias{check_prob}
\title{title check_prob
description checks whether input probability is a valid probability value in the range between 0 and 1
param prob probability of success
return True if prob input is valid and an error if the prob input is not in the valid range}
\usage{
check_prob(prob)
}
\description{
title check_prob
description checks whether input probability is a valid probability value in the range between 0 and 1
param prob probability of success
return True if prob input is valid and an error if the prob input is not in the valid range
}
