% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_success}
\alias{check_success}
\title{title check_success
description checks whether success is a valid value for number of success, which has to be in between bigger than or equal to 0 and less than or equal to trials
param success number of success
param trials number of trials
return True if number of successes are valid and an error if its not a valid value for number of successes}
\usage{
check_success(success, trials)
}
\description{
title check_success
description checks whether success is a valid value for number of success, which has to be in between bigger than or equal to 0 and less than or equal to trials
param success number of success
param trials number of trials
return True if number of successes are valid and an error if its not a valid value for number of successes
}
