% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_trials}
\alias{check_trials}
\title{title check_trial
description checks whether trials is a valid value for number of trials of being a non-negative integer
param trials number of trials
return True if trials are valid and an error if trial is invalid}
\usage{
check_trials(trials)
}
\description{
title check_trial
description checks whether trials is a valid value for number of trials of being a non-negative integer
param trials number of trials
return True if trials are valid and an error if trial is invalid
}
