% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{plot.bincum}
\alias{plot.bincum}
\title{plot.bincum()}
\usage{
\method{plot}{bincum}(x, ...)
}
\description{
generate a scatter plot of the cumulative distribution
}
\examples{
dis2 <- bin_cumulative(trials = 5, prob = 0.5)
plot(dis2)

}
