% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_dis_and_related.R
\name{plot.bindis}
\alias{plot.bindis}
\title{title plot.bindis
description plots a bar graph of a binomial distribution
param dist an object of class "bindis"
return a bar plot}
\usage{
\method{plot}{bindis}(dist)
}
\description{
title plot.bindis
description plots a bar graph of a binomial distribution
param dist an object of class "bindis"
return a bar plot
}
