% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{print.binvar}
\alias{print.binvar}
\title{print.binvar()}
\usage{
\method{print}{binvar}(x, ...)
}
\description{
create a method function to be able to nicely print the content of and object
}
\examples{
bin1 <- bin_variable(trials = 10, p = 0.3)
bin1

}
