% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_var_and_related.R
\name{print.summary.binvar}
\alias{print.summary.binvar}
\title{title print.summary.binvar
description prints list that contains statistical calculations (mean, mode, etc.) of a given random variable
param var an object of class "summary.binvar"
return concatenated strings}
\usage{
\method{print}{summary.binvar}(sumvar)
}
\description{
title print.summary.binvar
description prints list that contains statistical calculations (mean, mode, etc.) of a given random variable
param var an object of class "summary.binvar"
return concatenated strings
}
