% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic Functions.R
\docType{data}
\name{probability}
\alias{probability}
\title{Binomial Distribution}
\format{An object of class \code{NULL} of length 0.}
\usage{
probability
}
\arguments{
\item{trials}{Number of trials (numeric)}

\item{prob}{Probability of success (numeric)}
}
\value{
A dataframe of binomial distribution with probability
}
\description{
Calculates a binomial distribution with probability
}
\examples{
plot.bindis(bin_distribution(trials = 5, prob = 0.5))
 @export
}
\keyword{datasets}
