% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.binvar}
\alias{summary.binvar}
\title{Binomial Variable Summary}
\usage{
\method{summary}{binvar}(binvar)
}
\arguments{
\item{binvar, }{a binomial random variable object}
}
\value{
a summary.binvar object that is a list with above information
}
\description{
Creates a list with trials, prob, mean, variance, mode, skewness, and kurtosis information
}
