% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{aux_kurtosis}
\alias{aux_kurtosis}
\title{name: aux_kurtosis
description: computes the kurtosis
param: trials the number of trials
param: prob a probability
return: double: the kurtosis}
\usage{
aux_kurtosis(trials, prob)
}
\description{
name: aux_kurtosis
description: computes the kurtosis
param: trials the number of trials
param: prob a probability
return: double: the kurtosis
}
