% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary_functions.R
\name{aux_kurtosis}
\alias{aux_kurtosis}
\title{title aux_kurtosis
description determines kurtosis of given probability and given number of trials
param trials representing # of trials (numeric value)
param prob representing probability (numeric value)
return kurtosis (numeric value)}
\usage{
aux_kurtosis(trials, prob)
}
\description{
title aux_kurtosis
description determines kurtosis of given probability and given number of trials
param trials representing # of trials (numeric value)
param prob representing probability (numeric value)
return kurtosis (numeric value)
}
