% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{aux_mean}
\alias{aux_mean}
\title{name: aux_mean
description: computes the mean
param: trials the number of trials
param: prob a probability
return: double: the mean}
\usage{
aux_mean(trials, prob)
}
\description{
name: aux_mean
description: computes the mean
param: trials the number of trials
param: prob a probability
return: double: the mean
}
