% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary_functions.R
\name{aux_mean}
\alias{aux_mean}
\title{title aux_mean
description determines mean of given probability and given number of trials
param trials representing # of trials (numeric value)
param prob representing probability (numeric value)
return: mean (numeric value)}
\usage{
aux_mean(trials, prob)
}
\description{
title aux_mean
description determines mean of given probability and given number of trials
param trials representing # of trials (numeric value)
param prob representing probability (numeric value)
return: mean (numeric value)
}
