% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{aux_mode}
\alias{aux_mode}
\title{name: aux_mode
description: computes the mode
param: trials the number of trials
param: prob a probability
return: int: the mean}
\usage{
aux_mode(trials, prob)
}
\description{
name: aux_mode
description: computes the mode
param: trials the number of trials
param: prob a probability
return: int: the mean
}
