% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary_functions.R
\name{aux_mode}
\alias{aux_mode}
\title{title: aux_mode
description: determines mode of given probability and given number of trials
param trials representing # of trials (numeric value)
param prob representing probability (numeric value)
return mode (numeric value or numeric vector)}
\usage{
aux_mode(trials, prob)
}
\description{
title: aux_mode
description: determines mode of given probability and given number of trials
param trials representing # of trials (numeric value)
param prob representing probability (numeric value)
return mode (numeric value or numeric vector)
}
