% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{aux_skewness}
\alias{aux_skewness}
\title{name: aux_skewness
description: computes the skewness
param: trials the number of trials
param: prob a probability
return: double: the skewness}
\usage{
aux_skewness(trials, prob)
}
\description{
name: aux_skewness
description: computes the skewness
param: trials the number of trials
param: prob a probability
return: double: the skewness
}
