% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary_functions.R
\name{aux_skewness}
\alias{aux_skewness}
\title{title aux_skewness
description determines skewness of given probability and given number of trials
param trials representing # of trials (numeric value)
param prob representing probability (numeric value)
return skewness (numeric value)}
\usage{
aux_skewness(trials, prob)
}
\description{
title aux_skewness
description determines skewness of given probability and given number of trials
param trials representing # of trials (numeric value)
param prob representing probability (numeric value)
return skewness (numeric value)
}
