% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{aux_variance}
\alias{aux_variance}
\title{name: aux_variance
description: computes the variance
param: trials the number of trials
param: prob a probability
return: double: the variance}
\usage{
aux_variance(trials, prob)
}
\description{
name: aux_variance
description: computes the variance
param: trials the number of trials
param: prob a probability
return: double: the variance
}
