% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary_functions.R
\name{aux_variance}
\alias{aux_variance}
\title{title aux_variance
description determines variance of given probability and given number of trials
param trials representing # of trials (numeric value)
param prob representing probability (numeric value)
return variance (numeric value)}
\usage{
aux_variance(trials, prob)
}
\description{
title aux_variance
description determines variance of given probability and given number of trials
param trials representing # of trials (numeric value)
param prob representing probability (numeric value)
return variance (numeric value)
}
