% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-adele-shen.R
\name{bin_choose}
\alias{bin_choose}
\title{Combination Function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials (numeric)}

\item{k}{number of successful trials (numeric)}
}
\value{
number of combinations in which k successes can occur in n trials (numeric)
}
\description{
Calculates the number of combinations in which k successes can occur in n trials
}
\examples{
# Calculate 5 choose 2
bin_choose(n = 5, k = 2)

# Choose different numbers of successes
bin_choose(5, 1:3)
}
