% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Coefficient}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{Numeric, # of trials in the binomial distribution}

\item{k}{Numeric, # of successes in the binomial distribution}
}
\value{
Returns binomial coefficient
}
\description{
This function calculates the binomial coefficient of a given binomial distribution
}
\examples{
bin_choose(n = 3, k = 1)
bin_choose(6, 3)
bin_choose(5, c(1,2))
}
