% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_function.R, R/workout03-ann-deng.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose()}
\usage{
bin_choose(n, k)

bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}

\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
double of combinations

double of combinations
}
\description{
uses factorial() to compute number of combinations in which k successes can occur in n trials

uses factorial() to compute number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2) ## returns 10
bin_choose(n = 5, k = 2) ## returns 10
}
