% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Coefficient}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{total number of objects, any integer greater than zero}

\item{k}{number of objects to be selected, any integer or vector less than n}
}
\value{
binomial coefficient, or the number of ways to choose k objects from n # of objects
}
\description{
represents the number of ways one can choose k objects from n number of objects
}
\examples{
bin_choose(n = 5, k = 2) 
bin_choose(5, 0)
bin_choose(5, 1:3)
}
