% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials (non-negative integer)}

\item{k}{number of successes (non-negative integer less than trials)}
}
\value{
combinations of trials and success given n and k
}
\description{
alculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2), bin_choose(5, 0), bin_choose(5, 1:3)
}
