% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{Bin Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{an integer, the number of trials}

\item{k}{an integer or vector of numbers, the value of successes}
}
\value{
an integer, the number of combinations for k successes in n trials
}
\description{
A function to calculate the number of combinations in which k successes can occur in n trials.
}
\examples{
bin_choose(10,7)
bin_choose(10,5)
}
