% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\title{Binomial Choose}
\usage{
  bin1 <- bin_choose(5,2)
  bin1 <- bin_choose(5,1:3)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}

\examples{
# standard example
bin1 <- bin_choose(5,2)
 
# Another standard example
bin1 <- bin_choose(5,1:3)
}
