% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Coefficient}
\usage{
bin_choose(trials = 1, success = 1)
}
\arguments{
\item{n}{number of n trials}

\item{k}{number of k sucesses}
}
\value{
returns total number of combinations in which k success can occur in n trials
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
