% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choosing}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{vector of number of trials (numeric)}

\item{k}{vector of number of successes (numeric)}
}
\value{
number of calculated combinations
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{

bin_choose(n = 10, k = 5)

}
