% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Bin Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{is a numeric integer that represents the number of trials}

\item{k}{is a numeric value, or vector of values that represents the number of successes}
}
\value{
this function returns a numeric value that represents the number of combinations in which k successes can occur in n trials.
}
\description{
This function calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
