% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{Choose function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials (numeric)}

\item{k}{number of successes (numeric)}
}
\value{
Number of combinations
}
\description{
Calculates the number of combinations in which k successes can occur in n trials
}
\examples{

#calculates number of combinations 2 successes can occur in 5 trials
 bin_choose(n = 5, k = 2)

#calculates number of combinations 0 successes can occur in 5 trials
 bin_choose(5, 0)

#calculates number of combinations 1,2, or 3 successes can occur in 5 trials
 bin_choose(5, 1:3)

}
