% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials (n >= 0)}

\item{k}{is a variable that repesents the number of success out of n trials (0<=k <= n)}
}
\value{
the number of combinations in which k successes can oocur in n trials.
}
\description{
computes the probabilities about the number of success in a fixed number of random trials performed under identical conditions
}
