% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_choose}
\alias{bin_choose}
\title{Calculate combinations}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{integer the number of trials}

\item{k}{integer(s) the number of successes}
}
\value{
The number of combinatons in which k successes can occur in n trials
}
\description{
Calculates the number of combinatons in which k successes can occur in n trials
}
\examples{
# Calculate 5 choose 3
bin_choose(n = 3, k = 5)

# Calculate 10 choose 0
bin_choose(n = 0, k = 10)
}
