% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n = 5, k = 2)
}
\arguments{
\item{n}{(numeric)}

\item{k}{(numeric)}
}
\value{
the number of combinations
}
\description{
the number of combinations in which k successes can occur in n trials
}
