% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of success (can be vector)}
}
\value{
a numeric value or vector of numeric values
}
\description{
Calculates number of combinations in which k successes can occur in n trials
}
\examples{
# number of combinations of 3 successes that can occur in 5 trials
bin_choose(5, 3)
}
