% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes (must not be greater than n)}
}
\value{
computed number of combinations of success
}
\description{
calculates the number of combinations in which k successes can occur in n trials.
}
\examples{
bin_choose(5,0)
}
