% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_choose}
\alias{bin_choose}
\title{: Binomial Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{:}{n - number of trials; k - number of successes}
}
\value{
: number of combinations
}
\description{
: Calculates the number of combinations that where there are k successes in n trials
}
\examples{
: bin_choose(5, 2) = 10
}
