% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{trials as integer greater than zero}

\item{k}{successes as non-negative integer}
}
\value{
successes as an integer value
}
\description{
Calculates the number of combinations in which k successes occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
#10
bin_choose(5, 0)
bin_choose(5, 1:3)
}
