% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose Function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
Number of combinations
}
\description{
Computes number of successful combinations from given trials and successes.
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
