% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
number of ways k success can occur in n trials
}
\description{
calculates number of ways k successs can occur in n trials
}
\examples{
bin_choose(5, 2)
}
