% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-code.R
\name{bin_choose}
\alias{bin_choose}
\title{Number of Combinations}
\usage{
bin_choose(trials, success)
}
\arguments{
\item{trials}{number of trials}

\item{success}{number of successes}
}
\value{
number of combinations
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
