% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
number of combinations
}
\description{
calculates the number of combinations in which k success can occur in n trials
}
