% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of total trials}

\item{k}{number of desired successes}
}
\value{
n choose k
}
\description{
calculates number of combinations in n choose k
}
\examples{
bin_choose(5,2)
}
