% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{bin_choose}
\alias{bin_choose}
\title{n choose k}
\usage{
bin_choose(n, k)
}
\arguments{
\item{trials(n), }{success(k)}
}
\value{
choose
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2) = 10
}
