% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{Choose Combinations}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of trials}

\item{k}{the number of successes}
}
\value{
combinations an integer representing how many combos
are made with k successes in n trials
}
\description{
computes the number of combinations in which k successes can occur in n trials
}
\examples{
#2 success in 5 trials
bin_choose(n=5, k=2)

#1,2, or 3 successes in 5 trials
bin_choose(n=5, k = 1:3)
}
