% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{bin_choose}
\alias{bin_choose}
\title{Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{vector of success values}
}
\value{
a vector of the number of combinations for each success value
}
\description{
Computes the number of combinations in which k successes can occur in n trials
}
\examples{
choose(8, 5)

choose(10, 2:4)

choose(n = 6, k = 3)

}
