% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_dist_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial choose function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials (numeric)}

\item{k}{number of successes (numeric)}
}
\value{
number of combinations "n choose k"
}
\description{
Outputs number of combinations in which k succeses
can occur in n trials
}
\examples{
bin_choose(n=5,k=2)
bin_choose(5,0)
bin_choose(5,1:3)
}
