% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n, }{number of trials}

\item{k, }{number of succeses}
}
\value{
n choose k
}
\description{
Calculates n choose k
}
