% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindist.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials, integer}

\item{k}{number of successes, integer}
}
\value{
number of combinations in which k successes can occur in n trials
}
\description{
calculate the number of combinations
}
\examples{
bin_choose(5,2)
}
