% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n = 10, k = 3)
}
\arguments{
\item{n}{# of trials}

\item{k}{# of successes}
}
\value{
the # of combinations
}
\description{
Calculates the # of combinations in which k successes occur in n trials
}
\examples{

# default
bin_choose()

# another example
bin_choose(n = 5, k = 2)

}
