% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n = 1, k = 1)
}
\arguments{
\item{n}{positive integer}

\item{k}{positive integer less than or equal to n}
}
\value{
n choose k
}
\description{
computes the choose factor of the binomial probability density
}
\examples{
bin_choose(5,3)
bin_choose(6,1)
bin_choose(6,1:3)
}
