% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n:}{total number of items}

\item{k:}{number of items to be chosen}
}
\value{
n choose k
}
\description{
calculates number of combinations - how many different ways you can choose k items from n items set without repetition and without order
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
