% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of (fixed) trials (n ≥ 0)}

\item{k}{the probability of success on each trial (0 ≤ p ≤ 1)}
}
\value{
the number of combinations in which k successes can occur in n trials
}
\description{
a main function that calculates the number of combinations in which k successes can occur in n trials.
}
