% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_methods.R
\name{bin_choose}
\alias{bin_choose}
\title{number of combinations}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{a vector indicating successes}
}
\value{
\item{comb}{number of combinations}
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
#calculate the number of combinations of 2 successes in 5 trials
bin_choose(n = 5, k = 2)

#calculate the number of combinations of 0 successes in 5 trials
bin_choose(5, 0)
bin_choose(5, 1:3)
}
