% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{bin_choose}
\alias{bin_choose}
\title{Counts number of combinations}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
a vector of combinations corresponding to n choose k
}
\description{
Finds n choose k
}
\examples{
 \dontrun{
 # find combinations for given n and k
 trials <- 5
 success <- c(2, 3)
 choose <- bin_choose(trials, success)
 }
}
