% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{Number of combinations}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of elements}

\item{k}{the size of the subset}
}
\value{
the number of k-length subsets of a n-length set.
}
\description{
Calculates n choose k.
}
\examples{
 \dontrun {
   bin_choose(n = 5, k = 3)
 }
}
