% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(n = 0, k = 0)
}
\arguments{
\item{n}{integer for number of trials (default 10)}

\item{k}{integer for number of successes (default 0.5)}
}
\value{
Count of combinations with k successes over n trials
}
\description{
Calculates number of ways to have k successes in n trials
}
\examples{
# default
bin_choose()

# 2 successes in 5 trials
bin_choose(n = 5, k = 2)

# no successes in 5 trials
bin_choose(5, 0)
}
