% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose()}
\usage{
bin_choose(trials, success)
}
\arguments{
\item{trials}{number of trials}

\item{success}{number of successes}
}
\value{
the number of combinations
}
\description{
calculate the number of combinations in which k successes in n trials
}
