% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes or a vector of number of successes}
}
\value{
number of combinations in which k successes can occur in n trials
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
How many combinations if there are 2 successes in 5 trials?
bin_choose(n = 5, k = 2)
How many combinations if there are 0 successes in 5 trials?
bin_choose(5, 0)
How many combinations if there are 1 to 3 successes in 5 trials?
bin_choose(5, 1:3)
}
