% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{a vector of number of success}
}
\value{
number of combinations of successes in n trials
}
\description{
calculates the number of combinations in which k successes
can occur in n trials
}
\examples{
# the number of combinations in which k = 2 successes can occur in n = 5 trials is
bin_choose(n = 5, k = 2)
}
