% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{number of combination}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of trials}

\item{k}{the number of success}
}
\value{
then number of combination
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)

bin_choose(5, 0)

# k can be a vector
bin_choose(5, 1:3)

}
