% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
the number of combinations in which k successes can occur in n trials
}
\description{
calculates n choose k
}
\examples{
# Calculates the number n choose k for n = 7 and k =0,1,2,3
bin_choose(n = 7, k = 0:3)

}
