% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials, k number of successes}
}
\value{
the number of combinations in which k successes can occur in n trials
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
# Calculates the number of combinations in which 0 through 2 successes can happen in 5 trials
bin_choose(n = 5, k = 0:2)

}
