% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_choose}
\alias{bin_choose}
\title{binominal choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{how many trials}

\item{k}{how many success}
}
\value{
n choose k
}
\description{
choose function
}
\examples{
bin_choose(10,6)
}
