% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose factor}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{:number of trials}

\item{k}{:number of successes}
}
\value{
the choose factor
}
\description{
This computes the number of ways in which k successes could occur in n trials
}
\examples{
#5 choose 2
bin_choose(n = 5, k = 2)

#5 choose 0
bin_choose(5, 0)

#5 choose 1:3
bin_choose(5, 1:3)

}
