% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
the number of combinations of k successes that occur in n trials
}
\description{
calculates the number of combinations of k successes that occur in n trials
}

