% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{trials}

\item{k}{success}
}
\value{
n choose k
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(5, 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
