% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindis.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n, }{number of trials of the distribution}

\item{k, }{size of the combination want to be choose from the distribution}
}
\value{
Number of combinations there are to choose sample of size k from distribution of size n
}
\description{
Calculate the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(5, 3)
bin_choose(10, 4)
}
