% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{Binary Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{Trials}

\item{k}{Successes}
}
\value{
An integer
}
\description{
Returns the combination of k successes from n trials.
}
