% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomal.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose Function}
\usage{
bin_choose(trials, success)
}
\arguments{
\item{trials}{Total number of trials}

\item{success}{Total number of successes}
}
\value{
The number of combinations in which k successes can occur in n trials
}
\description{
Calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
