% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose}
\usage{
bin_choose(n = 1, k = 1)
}
\arguments{
\item{n}{number of objects to choose from}

\item{k}{number to choose}
}
\value{
numeric vector of choose values
}
\description{
calculates the number of ways to choose k objects from n
}
