% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
C(n, k)
}
\description{
computes C(n, k)
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
