% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{The number of trials}

\item{k}{The number of successes}
}
\value{
The number of combinations in which k successes can occur in n trials
}
\description{
Calculates the number of combinations in which k successes can occur in n trials
}
