% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes (can be in vector)}
}
\value{
total number of choose combinations (can be in vector)
}
\description{
The number of combinations in which k successes can occur in n trials.
}
\examples{
# 5 choose 2
bin_choose(n = 5, k = 2)

# 5 choose 0
bin_choose(5, 0)

# 5 choose c(1,2,3) vector form
bin_choose(5, 1:3)

}
