% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-choose.R
\name{bin_choose}
\alias{bin_choose}
\title{calculate the choosing number}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the times of trials}

\item{k}{the times of success}
}
\value{
the number of combinations
}
\description{
calculate the choosing number given the number of success and total trials
}
\examples{
n = 5
k = 3
bin_choose(n,k) #calculating the number of combinations of choosing 2 objects among 5.
}
