% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{is the number of trials}

\item{k}{is the number of successes}
}
\value{
number of combinations of k successes in n trials
}
\description{
calculates number of combinations
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
