% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{: bin_choose}
\usage{
bin_choose(n, k)
}
\value{
: the number of combinations
}
\description{
:calculates the number of combinations in which k successes can occur in n trials
}
