% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workout3.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{an integer with number of trials}

\item{k}{a list with number of success}
}
\value{
an integer represents the number of combinations
}
\description{
the function used to calculate the combinatorial number of
}
\examples{
bin_choose(10, 4)
}
