% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{numeric vector}

\item{k}{numeric vector}
}
\value{
a numeric vctor of all the possible combinations of successes
}
\description{
all possible combinations of k successes
}
\examples{

bin_choose(5, 4)


example <- bin_choose(4, 0:4)


}
