% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials (numeric)}

\item{k}{number of succeses (numeric)}
}
\value{
computed number of combinations
}
\description{
computes the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(5, 2)
bin_choose(5, 1:3)
}
