% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{Number of combinations}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{trials}

\item{k}{successes}
}
\value{
computed number of combinations
}
\description{
calculate number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(5,3)
}
