% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{choose n from k trials}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{numebr of success}
}
\value{
number of combinations
}
\description{
calculates the number of combinaitons in whicc k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
