% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{: bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{:}{n the number of trials}

\item{:}{k the number of successes}
}
\value{
: double: the total number of combinations
}
\description{
: computes the number of combination in which k successes can occur in n trials
}
\examples{
: bin_choose(n = 5,k = 2)
: bin_choose(5,0)
: bin_choose(5,1:3)
}
