% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{integer representing the size of the population}

\item{k}{integer or vector of integers representing the size of the groups to draw from n}
}
\value{
integer or vector of integers, the number of combinations
}
\description{
Calculates n choose k, the number of unique groups of size k that can be drawn from a group of size n
}
\examples{
bin_choose(n = 5, k = 2)

bin_choose(5, 0)

bin_choose(5, 1:3)
}
