% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Combinations}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{trials}

\item{k}{successes}
}
\value{
combo
}
\description{
Calculates number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(5, 2)
}
