% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trial}

\item{k}{number of successes}
}
\value{
number of combinations in which k successes can occur in n trials
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose_5_3 <- bin_choose(5,3)

bin_choose_10_0 <- bin_choose(10,0)

}
