% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
total number of combination.
}
\description{
compute the combination of k successes out of n trials
}
\examples{
bin_choose(5, 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
