% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choosing factor}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{total number of trials (positive integer)}

\item{k}{number of successes in n trials (positive integer)}
}
\value{
return the number(s) of combinations
}
\description{
Calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n=5,k=2)
10
bin_choose(n=5,k=0)
1
bin_choose(n=5,1:3)
5 10 10

}
