% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_choose}
\alias{bin_choose}
\title{Number of Combinations of a Binomial}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of success}
}
\value{
total number of combinations of "successes" which can occur in n trials
}
\description{
Calculates n choose k (choose() in base r)
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(n = 5, k = 0)
bin_choose(n = 5, k = 1:3)
}
