% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{the number of combination}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of trials}

\item{k}{the number of successes}
}
\value{
the calculated number of combination in the situation of k successes in n trials
}
\description{
calculate the number of combination
}
\examples{

bin_choose(n=5,k=2)

}
