% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_bin_func.R
\name{bin_choose}
\alias{bin_choose}
\title{choose binomial}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{some integer}

\item{k}{some integer}
}
\value{
number of combinations of successes
}
\description{
calculates number of combinations of successes in a given number of trials
}
\examples{
bin_choose(n=5, k=2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
