% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Combination}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{numeric, number of trials}

\item{k}{numeric, number of successes}
}
\value{
the number of combinations of possible k successes in n trials
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)


}
