% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(successes, trials)
}
\arguments{
\item{successes}{number of sucesses}

\item{trials}{number of trials}
}
\value{
returns numeric vector of total combinations
}
\description{
uses sapply over bin_find to vectorize possible combinations
}
\examples{
bin_choose(2,5)
bin_choose(successes=5, trials=15)
bin_choose(1:5, 55)
}
