% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{Binary Choose Function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
computed value
}
\description{
Calculates the number of combinations in which k successes can occur in n tirals
}
\examples{
# five trials, two successes
bin_choose(n = 5, k = 2)
10

# five trials, 0 success
bin_choose(5, 0)
1

# five trials, 1-3 successes
bin_choose(5, 1:3)
5 10 10
}
