% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main function.R
\name{bin_choose}
\alias{bin_choose}
\title{: function bin_choose()}
\usage{
bin_choose(trials, success)
}
\arguments{
\item{:}{success,trials}
}
\value{
: the number of combinations in which k successes can occur in n trials
}
\description{
: calculates the number of combinations in which k successes can occur in n trials
}
\examples{
: 
bin_choose(n=5,k=2)
bin_choose(5,0)
bin_choose(5,1:3)
}
