% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial combinations}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of fixed trials}

\item{k}{the number of successes out of n trials}
}
\value{
the number of combination
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
