% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of success}
}
\value{
number of combinations of k successes occuring in n trials
}
\description{
calculates number of combinations in which k successes occur in n trials
}
\examples{
bin_choose(5,2)
}
