% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of trials}

\item{k}{the number of successes}
}
\description{
a ground up choose function using factorial
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
