% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n = 5, k)
}
\arguments{
\item{n}{numeric vaule of number of trials}

\item{k}{numeric value of number of success}
}
\description{
this function calculates the number of combinatons in which k successes can occer in n trials
}
\examples{
Calculate the number of different ways to get 2 successes out of 5 trails.
bin_choose(n = 5, k = 2)
}
