% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{number of possible combinations in which a number of successes occurs}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of trials, a non-negative integer}

\item{k}{the number of successes, a non-negative integer less than n}
}
\value{
computed number of possible combinations in which k successes can occur in n trials
}
\description{
calculates the number of possible combination of successes in a given number of trials
}
\examples{
# the number of possible combinations of 3 successes in 5 trials
bin_choose(5,3)

# the number of 7 successes in 12 trials
bin_choose(12,7)
}
