% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Function of Choosing k from n}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{total number of trials (integer)}

\item{k}{vector of number of successes (integer)}
}
\value{
number of combinations in which k successes occur in n trials
}
\description{
compute the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
