% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_choose}
\alias{bin_choose}
\title{Bin_Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{an integer or vector of numerics - number of trials}

\item{k}{an integer or vector of numerics - number of successes}
}
\value{
an integer - number of combinations in which k successes can occur in n trials
}
\description{
Computes 'n choose k'
}
\examples{
# integers 
combs <- bin_choose(n = 5, k = 2)

# vectorized
combs <- bin_choose(5, 1:3)

}
