% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_choose}
\alias{bin_choose}
\title{The Choose Function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{(integer) the number of trials}

\item{k}{(integer) a vector of the number of successes - -can be a single value or multiple values. Must be less than n.}
}
\value{
function returns the value of n choose k for each value k provided
}
\description{
bin_choose() takes one value for trials and multiple values for sucesses and returns the choose function for trials choose successes
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(10, 4)
bin_choose(n = 20, k = c(5, 10, 15))
}
