% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_chose_bin_probability.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{representing number of elements to combine (numeric value)}

\item{k}{representing number of k-combinations (numeric value)}
}
\value{
a numeric value
}
\description{
calculates the k-combination of n elements
}
\examples{
bin_choose(5,3)
bin_choose(1,1)
}
