% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{bin_choose}
\alias{bin_choose}
\title{choose function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{k:}{number of successes, a non-negative integer}

\item{n:}{number of trials, a non-negative integer}
}
\value{
"n choose k"
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
> bin_choose(5,2)
[1] 10
> bin_choose(5,0)
[1] 1
> bin_choose(5,1:3)
[1] 5 10 10
}
