% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_choose}
\alias{bin_choose}
\title{choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of trials}

\item{k}{the number of success}
}
\value{
the number of combinations in which k successes can occur in n trials.
}
\description{
that calculates the number of combinations in which k successes can occur in n trials.
}
\examples{

bin_choose(n=5,k=2)
#the number of combinations where 2 successes can occur in 5 trials.

}
