% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{desired number of successes}
}
\value{
one numeric value describing number of combinations
}
\description{
calculates the number of combinations in which
k successes can occur in n trials
}
