% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_choose}
\alias{bin_choose}
\title{Function bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{times of success}
}
\value{
the bumber of combinations
}
\description{
calculates the number of combinations in which k successes can occur in n trials.
}
\examples{
# calculates the number of combinations in which 2 successes can occur in 5 trials.
bin_choose(n = 5, k = 2)
}
