% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunction.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomal choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of trials}

\item{k}{the vector of successes}
}
\value{
computer the number of combinations in which k successes can occur in n trials
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
Calcualte the number of combinations in which k successes can occur in n trials
How many combinations when choosing 2 successes in 5 trials?
bin_choose(n = 5, k = 2)
How many combinations when choosing 0 successes in 5 trials?
bin_choose(5, 0)
}
