% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{int}{integer n trials}

\item{int}{integer k successes}
}
\value{
combination value
}
\description{
calculates the number of combinations in which k successes can occur in n trials.
}
\examples{
bin_choose(n = 5, k = 2)

bin_choose(5, 0)

bin_choose(5, 1:3)
}
