% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{: bin_choose()}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of trials}

\item{k}{the number of successes}
}
\value{
: calculated number of combinations when n and k are given
}
\description{
: calculate the number of combinations in which k successes can occur in n trials
}
