% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{Number of combinations}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{vector or single number of successes}
}
\value{
number of combinations
}
\description{
calculates number of combinations of `k` successes in `n` trials
}
\examples{
bin_choose(n = 5, k = 2)

bin_choose(6, 2)
}
