% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{bin_choose}
\alias{bin_choose}
\title{Factorial}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{denotes trials}

\item{k}{denotes succeses}
}
\value{
the factorial of "n chooses k"
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin1 <- bin_choose()

}
