% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{a number, total to choose from}

\item{k}{a number, how many chosen}
}
\value{
n choose k
}
\description{
calculates the choose function
}
\examples{
bin_choose(10,3)
}
