% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose factor}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{numeric value, number of trials}

\item{k}{numeric vector, number of successes}
}
\value{
computed choose factor
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
#5 choose 2
bin_choose(n = 5, k = 2)

#5 choose 0
bin_choose(5, 0)

#5 choose 1:3
bin_choose(5, 1:3)
}
