% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_statmanips.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose Function}
\usage{
bin_choose(n = 1, k = 1)
}
\arguments{
\item{n}{the number of objects to choose from}

\item{k}{the number of objects to choose}
}
\value{
n choose k
}
\description{
Takes n and k and calculates n choose k, given that n is always larger than k
}
\examples{
bin_choose(5,2)
bin_choose(5,0)
bin_choose(5,1:3)
}
