% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_and_methods.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes, less than or equal to n}
}
\value{
the amount of combinations in which k successes can occur in n trials
}
\description{
calculates the amount of combinations in which k successes can occur in n trials.
}
\examples{
example1 <- bin_choose(n = 5, k = 2)
}
