% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose factor}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{integer value for number of trials}

\item{k}{integer value for number of successes}
}
\value{
integer value for choose factor
}
\description{
computes binomial choose factor
}
\examples{

bin_choose(5, 2) returns 10
bin_choose(5, 1:3) returns c(5, 10, 10)
}
