% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_workout3.R
\name{bin_choose}
\alias{bin_choose}
\title{factor in the binomial formula}
\usage{
bin_choose(trials, success)
}
\arguments{
\item{trials}{number of trials}

\item{success}{number of successes}
}
\value{
a certain number of combinations
}
\description{
function that creates the combinations for n and k w/o using choose
}
\examples{
bin_combos = factorial(n)/(factorial(k)*factorial(n-k))
}
