% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{a number}

\item{k}{a number or vector}
}
\value{
a number
}
\description{
calculates prob
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
