% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose Function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{The number of trials}

\item{k}{The number of successes (Could be a vector longer than 1)}
}
\value{
The number of combinations in which k successes occur in n trials.
}
\description{
Function that takes in 2 arguments: n (number of trials) and k (number of successes) and
and calculates the number of combinations in which k successes occurs in n trials.
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(n = 5, k = 0)
bin_choose(n = 5, k = 1:3)
}
