% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials; non-negative integer value}

\item{k}{number of successes; non-negative integer value}
}
\value{
numeric value of combinations that can occur
}
\description{
Calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)

bin_choose(5, 0)

bin_choose(5, 1:3)
}
