% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{represents the trials, and k represents the times of success}
}
\value{
it calculate the result of the choose(n,k)
}
\description{
it is kind of a substitution to the function choose()
}
