% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{Total number of trials}

\item{k}{Number of successes}
}
\value{
Calculates the number of combinations in which k successes can occur in n trials
}
\description{
Calculates the number of combinations in which k successes can occur in n trials
}
