% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{accepts}{n: number of trials and k: size of combinations}
}
\value{
returns the number of combinations of size k
}
\description{
number of ways to choose k in n trials
}
