% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{bin_choose}
\alias{bin_choose}
\title{Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{- number of trials}

\item{k}{- number of successes of n trials}
}
\value{
The number of combinations in which k successes can occur in n trials.
}
\description{
Choose
}
\examples{

#default
bin <- bin_choose()

#simple case
simple <- bin_choose(n = 5, k = 2)

#With vector
vect <- bin_choose(5, 1:3)
}
