% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials (numeric)}

\item{k}{number of successes (numeric vector)}
}
\value{
number of combinations
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
