% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{Number of trials}

\item{k}{Number of successes}
}
\value{
numeric: # of combinations
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
:
bin_choose(7,5)
}
