% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_main.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{Number of trials (numeric)}

\item{k}{Number of successes (numeric)}
}
\value{
Number of combinations (numeric)
}
\description{
Calculates number of combinations in which k successes can occur in n trials.
}
\examples{
bin_choose(n = 5, k = 2)
10

bin_choose(5, 0)
1

bin_choose(5, 1:3)
5 10 10
}
