% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R, R/hello.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(n, k)

bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}

\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
number of combinations

number of combinations
}
\description{
Calculates the number of combinations in which k succeses can occur in n trials

Calculates the number of combinations in which k succeses can occur in n trials
}
\examples{
bin_choose(4, 2) = 6
bin_choose(4, 2) = 6
}
