% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-functions-methods.R
\name{bin_choose}
\alias{bin_choose}
\title{Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
number of combinations in which k successes can occur in n trials
}
\description{
Calculates the number of combinations in which k successes can occur in n trials
}
\examples{
# k = 2 successes in n = 5 trials
choose52 = bin_choose(n = 5, k = 2)

# k = 3 successes in n = 8 trials
choose83 = bin_choose(8, 3)

# k = 1 to 3 successes in n = 5 trials
bin_choose(5, 1:3)
}
