% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials (numeric)}

\item{k}{number of successes (numeric)}
}
\value{
the number of combinations in which k successes can occur in n trials
}
\description{
Calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose (n = 5, k = 2)
bin_choose (n = 5 , k = 0)
bin_choose (n = 5, k = 1:3)
}
