% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of things to choose from}

\item{k}{number of things to choose}
}
\value{
n choose k
}
\description{
Calculates n choose k
}
\examples{
# choosing 3 from 5
bin_choose(5, 3)

# choosing 6 from 12
bin_choose(12, 6)

}
