% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trial}

\item{k}{number of success}
}
\value{
nCk
}
\description{
calculate n choose k
}
\examples{
bin_choose(3, 2)
bin_choose(5, 1:3)
bin_choose(5, 0)
}
