% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bino.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes, can be a vector}
}
\value{
the binomial coefficient for each success
}
\description{
Computes the binomial coefficient
}
\examples{
# 10 success in 20 trials
fac <- bin_choose(20, 10)

}
