% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_function.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose Function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials (numeric)}

\item{k}{number of successes (numeric)}
}
\value{
computed number of combinations in which k successes can occur in n trials (numeric)
}
\description{
Computes the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
