% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of success}
}
\value{
number of combinations if bin_choose output is valid meaning success is less than or equal to trials. If not, return an error message.
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(5, 3)
bin_choose(4, 1)
}
