% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes or a vector (represents several times) of number of successes}
}
\value{
number of combinations
}
\description{
calculates number of combinations that k successes occurs in n trials
}
\examples{
bin_choose(5, 3)
}
