% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials, must be non negative int}

\item{k}{number of successes, must be non negative int less than num trials}
}
\value{
computed number of combinations
}
\description{
Computes the number of combinations possible with given number of successes in
given number of trials
}
\examples{
bin_choose(5, 2)
bin_choose(n = 5, k = 0)
bin_choose(n = 5, k = 1:3)
}
