% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose function}
\usage{
bin_choose(trials, success)
}
\arguments{
\item{trials}{describe the number of trials}

\item{success}{describe the number of success}
}
\value{
number of combinations possible
}
\description{
describe the choose function
}
\examples{
bin_choose(3,2)
bin_choose(5,7)
}
