% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_choose}
\alias{bin_choose}
\title{Choose function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of elements to choose from}

\item{k}{number of elements to choose}
}
\value{
choose(n,k)
}
\description{
computes n choose k
}
\examples{
bin_choose(5,7)
}
