% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Coefficient}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{Number of trials in the binomial distribution (numeric)}

\item{k}{Number of successes in the binomial distribution (numeric)}
}
\value{
Returns the binomial coefficient
}
\description{
Calculates the binomial coefficient of a binomial distribution
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
