% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_prob,distr, cumulative, choose.R
\name{bin_choose}
\alias{bin_choose}
\title{bin choose.}
\usage{
bin_choose(n, k)
}
\arguments{
\item{trials}{the number of trials.}

\item{prob}{the probability of success in any given trial.}
}
\value{
the factorial-derived numerical value.
}
\description{
calculates the number of combinations in which k successes can occur in n trials.
}
\examples{
bin_choose(n = 5, k = 2)

bin_choose(8, 3)
}
