% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial "Choose"}
\usage{
bin_choose(n, k)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
number of combinations in which k successes can occur in n trials.
}
\description{
that calculates number of combinations in which k successes can occur in n trials.
}
\examples{
# default
choose <- bin_choose(3, 2)

}
