% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin choose function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials (intger)}

\item{k}{number of success (intger)}
}
\value{
the calculated number of combinations in which k successes can occur in n trials
}
\description{
calculate the number of combinations in which k successes can occur in n trials
}
