% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binom_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n = 5, k = 2)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
the number of combinations where k successes happen in n trials
}
\description{
gives number of successes in given number of trials
}
\examples{
10
}
