% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials; a numeric vector}

\item{k}{number of successes out of the number of trials; a numeric vector}
}
\value{
the number of combinantions in which k successes can occur in n trials
}
\description{
bin_choose
}
\examples{
bin_choose(n = 5, k = 2)

bin_choose(5, 0)

bin_choose(5, 1:3)
}
