% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{choose from binomial distribution}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n, k}{integer}
}
\value{
the choosen number integer
}
\description{
choose a number given the success and trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
