% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose()}
\usage{
bin_choose(n = 1, k = 1)
}
\arguments{
\item{n}{trials the number of the total trials}

\item{k}{scuuess the probability of a succeess}
}
\value{
bin_chooose returns the number of success canocuur in n trials
}
\description{
bin_chooose() is a function that calculates the number of combinations in which k success canccur in n trials
}
\examples{
bin_choose()
bin_choose(1,1)
}
