% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n = 5, k = 2)
}
\arguments{
\item{n}{value of trials}

\item{k}{value of successes}
}
\description{
calculate the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(5, 2)
10
}
