% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choosing function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of trial (integer)}

\item{k}{the number of success (integer)}
}
\value{
the number of comnination
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
