% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose function}
\usage{
bin_choose(trials, success)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{success}{number of success (numeric)}
}
\value{
the number of combinations in which k successes can occur in n trials
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
