% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.R
\name{bin_choose}
\alias{bin_choose}
\title{the number of combinations}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of trials (numeric)}

\item{k}{a vector of number of success (numeric)}
}
\value{
computed combinations
}
\description{
calculates the number of combinations
}
\examples{
bin_choose(n=5,k=2)
bin_choose(5,4)
bin_choose(4,1:3)
}
