% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{trial numbers}

\item{k}{success numbers}
}
\value{
the combination of k successes can occur in n trials
}
\description{
compute combinations in which k successes can occur in n trials
}
\examples{
bc <- binchoose(3, 2)
}
