% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
a numerical value of the number of combinations if inputs are valid
}
\description{
calculate the number of combinations in which k successes can occur in n trials.
}
