% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose Function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the numeric vector of total trials}

\item{k}{the numeric vector of success}
}
\value{
computed combinations
}
\description{
calculates the number of combinations in which k successes can occur in n trials.
}
\examples{
bin_choose(10, 3)
}
