% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3_Zimeng_Zheng.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of trials}

\item{k}{the list of numbers of success}
}
\value{
the number of combinations
}
\description{
function to calculate the combinatorial number
}
\examples{
bin_choose(n=5,k=2)
bin_choose(5,0)
bin_choose(5,1:3)
}
