% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{Bin Choosing}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{all choices}

\item{k}{number to choose}
}
\value{
n choose k
}
\description{
Uses factorial method to provide n choose k
}
