% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{numeric}

\item{k}{numeric vector}
}
\value{
calculated result
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n=5, k=2)

bin_choose(5,0)

bin_choose(5,1:3)
}
