% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number indicating the number of objects to be chosed from k}

\item{k}{number indicating the total number of objects}
}
\value{
the number of ways to choose k from n
}
\description{
calculate the number of possibilities to choose k from n
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
