% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-adele-shen.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Distribution Function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
data.frame of classes "bincum" and "data.frame" with the probability of the number of successes
occurring in the number of trials and the cumulative probability (of at that number of successes or less)
}
\description{
Calculates probabilities about the number of successes in a fixed number of random trials
performed under identical conditions (assuming a constant probability of success on each trial)
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
