% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of head probability}
}
\value{
a data frame with both the probability distribution and the cumulative probabilities: sucesses in the first column, probability in the second column, and cumulative in the third column
}
\description{
calculates the binomial probabilities and cumulative probabilities of different number of successes occuring in fixed number of trials
}
\examples{

# binomial cumulative distribution
bin_cumulative(trials = 5, prob = 0.5)

# plotting binomial cumulative distribution
dis2 <- bin_cumulative(trials = 5, prob = 0.5)
plot(dis2)

}
