% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{Numeric, # of trials in the given binomial distribution}

\item{prob}{Numeric, probability of success in the given binomial distribution}
}
\value{
Returns a data frame of two classes (bincum and data.frame) containing the columns
success, probability, cumulative.
}
\description{
computes and tabulates the cumulative density for given
binomial distribution
}
\examples{
bin_cumulative(trials = 2, prob = .5)
}
