% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_probability.R,
%   R/workout03-ann-deng.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative()}
\usage{
bin_cumulative(trials, prob)

bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
data frame with probability distribution and cumulative probabilities and class "bincum

data frame with probability distribution and cumulative probabilities and class "bincum
}
\description{
uses bin_distribution to output prob. distribution and cumulative probabilities

uses bin_distribution to output prob. distribution and cumulative probabilities
}
\examples{
bin_cumulative(5, 0.5) ## returns data.frame(c(0, 1, 2, 3, 4, 5), c(0.03125, 0.15625, 0.31250, 0.31250, 0.15625, 0.03125), c(0.03125, 0.18750, 0.50000, 0.81250, 0.96875, 1.00000)) 
bin_cumulative(5, 0.5) ## returns data.frame(c(0, 1, 2, 3, 4, 5), c(0.03125, 0.15625, 0.31250, 0.31250, 0.15625, 0.03125), c(0.03125, 0.18750, 0.50000, 0.81250, 0.96875, 1.00000)) 
}
