% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (non-negative integer)}

\item{prob}{probability of success (real number between 0 and 1)}
}
\value{
a "bincum" class object with "data.frame" secondary class.
data frame has three columns success, probability, cumulative
function also includes and plot function to plot a line plot of success and cumulative
}
\description{
function calculates the cumulative
}
\examples{
dis2 <- bin_cumulative(trials = 5, prob = 0.5)
plot(dis2)
}
