% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3-divya-vijay.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of probability}
}
\value{
dataframe of primary class bincum
}
\description{
creates a dataframe of the probability distribution and cumulative probabilities
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
bin_cumulative(trials = 3, prob = 0.1)
}
