% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Bin Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{an integer, the number of trials}

\item{prob}{an integer of the probability of success for one trial}
}
\value{
a data frame with the probability for all possible number of successes given the number of trials, the probability of success, and the cumulutaive probability of success for all successes values before.
}
\description{
A function to create a data frame displaying the probability for all possible number of successes given the number of trials and the probability of success, along with the culmulative probability.
}
\examples{
bin_cumulative(10,0.5)
}
