% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
data frame with the probability distribution
}
\description{
Computes binomial distribution
}
\examples{
cum <- bin_cumulative(trials = 5, prob = 0.5)
cum

plot(cum)
}
\seealso{
\code{\link{bin_distribution}}
}
