% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\title{Binomial cumulative}
\usage{
  bin1 <- bin_cumulative(5,0.5)
}
\arguments{

\item{trials}{number of trials}

\item{prob}{probability between 0-1}
}
\examples{
# standard example
bin1 <- bin_cumulative(5,0.5)
}
