% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Comulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
returns an object of class \code{c("bincum", "data.frame")}
}
\description{
creates an object of class \code{c("bincum", "data.frame")}
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
