% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{vector of number of trials (numeric)}

\item{prob}{vector of probabilities (numeric)}
}
\value{
data frame containing successes, probability, and cumulative prob
}
\description{
calculates the cumulative binomial distribution with given params
}
\examples{

bin_cumulative(trials = 5, prob = 0.5)

dis2 <- bin_cumulative(trials = 5, prob = 0.5)
plot(dis2)

}
