% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Bin cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{is a numeric integer that represents the number of trials}

\item{prob}{is a numeric value between 0 and 1 that represents the probability of success}
}
\value{
this function returns a data frame with three columns - "success", "probability" and "cumulative". All three contain numeric values.
}
\description{
This function creates a data frame of classes data.frame and bincum with the number of successes in one column, the probability of those successes in the middle, and cumulative probability in the third.
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
bin_cumulative(trials = 100, prob = 0.2)
bin_cumulative(trials = 45, prob = 0.75)
}
