% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Distribution and cumulative probabilties function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
Data frame with sucesses in the first column, probability in the second column, and cumulative in the third column
}
\description{
Generates a data frame with the probability distribution and the cumulative probabilities
}
\examples{

#binomial cumulative distribution
 bin_cumulative(trials = 5, prob = 0.5)

}
