% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial cumulative distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (n >= 0)}

\item{prob}{the probaility to win in oone trials}
}
\value{
A cumulative table of particular probability
}
\description{
To calculate the cumulative distribution of a binomial
}
