% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{prob}{probability of a success}

\item{n}{number of trials}
}
\value{
dataframe
}
\description{
Calculates the cumulative probabilities of each k success in n trials
}
\examples{
# binomial cumulative distribution
bin_cumulative(trials = 5, prob = 0.5)
}
