% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a data frame with the probability distribution (successes in 1st column, probability in 2nd column, & cumulative in the third column)
}
\description{
calculates the probability distribution of successes in given no. of trials, and their cumulative
}
\examples{
bin_distribution(trials = 5, prob = 0.5)

}
