% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
data frame
}
\description{
return a data frame with the probability distribution and cumulative probabilities.
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
