% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{: Binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{:}{trials, prob}
}
\value{
: dataframe with individual and cumulative probabilities for success in trials
}
\description{
: Creates a dataframe with another column for cumulative probabilities
}
\examples{
: bin_cumulative(5, 0.5)
}
