% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a data frame with the cumulative probability of binomial distribution
}
\description{
calculates the cumulative probability of a binomial distribution
}
\examples{
bin_cumulative(trials=5, prob = 0.5)

}
