% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunc.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{Probability that a success occurs}
}
\value{
data frame with : success;probability distribution;cumulative probabilities
}
\description{
calculate binomial probability and cumulative function
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
