% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Distribution Function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{success probability}
}
\value{
Dataframe with binomial cumulative distribution
}
\description{
Creates binomial distribution with cumulative probabilities, using given binomial probability and trials.
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
bin_cumulative(trials = 10, prob = 0.3)
}
