% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
binomial cumulative distribution
}
\description{
to calculate binomial cumulative distribution
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
