% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
data frame of the distribution and cumulative probability
}
\description{
calculates the binomial distribution, along with the cumulative
probability of the distribution
}
\examples{
bin_cumulative(5, 0.5)
}
