% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-code.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative probabilities data frame}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{for number of trials}

\item{prob}{for probability}
}
\value{
a data frame with successes, their probabilities, and cumulative probabilities
}
\description{
creates a data frame showing successes, their probabilities, and cumulative probabilities
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
