% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of total trials}

\item{prob}{probabilty of success in each trial}
}
\value{
data.frame with successes in the first column, probability in the second column, cumulative probability in the third column
}
\description{
calculates probability of successse in trials, as well as the cumulative probability of all results so far
}
\examples{
bin_cumulative(5, 0.5)
}
