% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials, }{prob}
}
\value{
a data.frame with two classes: c("bincum", "data.frame")
}
\description{
to get a dataframe of binomial cumulative distribution for each trial
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
