% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success per trial}
}
\value{
a data.frame with two classes
}
\description{
calculating the  cumulative distribution
}
\examples{
# binomial cumulative distribution
bin_cumulative(trials = 5, prob = 0.5)
}
