% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_dist_objects.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative Binomial Distribution function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
dataframe with number of successes, probability,
and cumulative probability with two classes, bindis and data.frame
}
\description{
Returns dataframe with probability
distribution and cumulative probability
}
\examples{
bin_cumulative(trials=5,prob=0.5)
}
