% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial CDF}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials, }{number of flips}

\item{prob, }{probability of observing heads}
}
\value{
dataframe object detailing the distribution and CDF
}
\description{
Calculates the distribution of a binomial random variable and its CDF
}
