% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Bin Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a dataframe with both prob distribution and cumulative prob
}
\description{
Creates a dataframe of the binom cumulative distribution over n trials
}
\examples{
# bincum distribution of five trials with a half probability
cum <- bin_cumulative(5, 0.5)
}
