% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindist.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials, integer}

\item{prob}{probability of success}
}
\value{
data frame with cumulative probability distribution
}
\description{
computes the cumulative binomial distribution function
}
\examples{
bin_cumulative(5,0.5)
}
