% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{# of trials}

\item{prob}{probability of success}
}
\value{
the probability and cumulative probability distribution
}
\description{
creates a data frame with a distribution of probabilites and cumulative probabilities
}
\examples{

# default
bin_cumulative()

# another example
bin_cumulative(trials = 3, prob = 0.3)

}
