% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability value for the binomial variable}
}
\value{
a dataframe with both the probability distribution and the cumulative probabilities:
successes in the first column, probability in the second column, and cumulative in the third column
}
\description{
calculates the cumulative probability of a binomial variable
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
