% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials:}{total number of trials}

\item{prob:}{probability of success}
}
\value{
data.frame of success, probability, and cumulative probability
}
\description{
adds a column of cumulative probability to the data frame generated by bin_distribution
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
