% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (single value)}

\item{prob}{probability of success}
}
\value{
Data frame with successes from 0 to number of trials,
corresponding probabilities, and cumulative probabilities.
Has classes "bincum" and "data.frame"
}
\description{
Computes distribution with probabilities and cumulative probabilities
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
