% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative distribution function of a binomial random variable}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of I.I.D. Bernoulli trials}

\item{prob}{the shape parameter of each I.I.D. Bernoulli trial}
}
\value{
a data frame representing the Cumulative distribution function of a
  binomial random variable with the specified shape parameters
}
\description{
Cumulative distribution function of a binomial random variable
}
\examples{
cdf = bin_cumulative(trials = 10, prob = 0.5)
support = cdf$success
}
