% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_objects.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative Binomial Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (integer)}

\item{prob}{probability of success in one trial (integer)}
}
\value{
'bincum' data.frame with successes, probability, and cumulative probability
}
\description{
Creates an object of class 'bincum' (cumulative binomial distribution)
}
\examples{
# binomial probability distribution
bin_cumulative(trials = 5, prob = 0.5)
}
