% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative()}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability}
}
\value{
a data frame with successes and probability and cumulative
}
\description{
calcuate the binomial cumulative distribution
}
