% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative()}
\usage{
bin_cumulative(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{prob}
}
\value{
return a data frame with both the probability distribution and the cumulative probabilities
}
\description{
create a data.frame that contains binomial distribution and cumulative probabilities
}
