% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability if success}
}
\value{
a data frame with both the probability distribution and the
cumulative probabilities: sucesses in the first column, probability
in the second column and cumulative in the third column.

an object of class \code{"data.frame"}
}
\description{
compute the probability distribution and
the cumulative distribution
}
\examples{
# binomial cumulative distribution of different successes in five trials
# (assume the probability of success is 0.5)
bin_cumulative(trials = 5, prob = 0.5)
}
