% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{must be an integer, greater than zero}

\item{prob}{a number between 0 and 1}
}
\value{
data frame with # of successes, probability, and cumulative sum
}
\description{
will give a data frame of both the probabilities and the cumulative sum associated with a given number of successes.Can then be plotted directly using plot() function.
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
bin_cumulative(10, 0.2)
}