% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{prob}{probablity of success}

\item{n}{number of trials}
}
\value{
a dataframe with class "bincum"
}
\description{
calculates the binomial CDF with n trials and probability prob
}
\examples{
Calculates the PMF of a binomial distribution of 6 trials and a success probability of .5
bin_cummulative(trials = 6, prob = 0.5)
}
