% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{n}{number of trials, prob probablity of success in each trial}
}
\value{
a dataframe with two classes: "bincum" and "data.frame"
}
\description{
calculates the cumulative probability of 0 to each possible success value with probability prob of success for each trial n
}
\examples{
# Calculates the cumulative distribution of each possible number of successes in a dataframe and object with 5 trials and a probability of .5
bin_cumulative(trials = 5, prob = 0.5)
}
