% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
Dataframe with both the probabilty distribution and the cumulative probabilities
}
\description{
Creates an object of cumulative binomial distribution
}
\examples{
bin_cum <- bin_cumulative(trials = 5, prob = 0.5)
plot(bin_cum)
}
