% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative distribution}
\usage{
bin_cumulative(trials, probs)
}
\arguments{
\item{trials}{: This is the number of trials}

\item{probs}{:This computes the probability of success}
}
\value{
data frame representation of potential successes and related (cumulative) probability
}
\description{
This calculates the cumulative probability of getting the successes
}
\examples{
#build a binomial distribution
dis2 <- bin_cumulative(trials = 5, probs = 0.5)
#You can plot it through the code below
#plot(dis2)
}
