% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}

\arguments{
\item{trials}{number of trials}
\item{prob}{probablity of success}
}

\value{
dataframe with number of success and cumulative probability
}

\description{
calculates the cumulative probability of each success in n trials
}
\examples{
bin_cumulative(5, .5)
}

