% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin cumulative}
\usage{
bin_cumulative(n, p)
}
\arguments{
\item{n}{trials}

\item{p}{prob}
}
\value{
cumulative
}
\description{
calculates the cumulation
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
