% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindis.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials, }{number of trails of the binomial distribution}

\item{prob, }{probability of success of the binomial distribution}
}
\value{
A data frame with both the probability distribution and the cumulative probabilities: successes in the first column, probability in the second column, and cumulative in the thrid column
}
\description{
Given the number of trials and the probability of success, calculate the binomial cumulative distribution
}
\examples{
bin_cumulative(9, 0.45)
bin_cumulative(12, 0.3)
}
