% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binary Cumulation}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{Trials}

\item{prob}{Probability}
}
\value{
A cumulative distribution table
}
\description{
Returns a cumulative table for successes from distribution table.
}
