% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomal.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
A data frame of the computed binomial distribution with classes ("bincum" and "data.frame")
}
\description{
Creates a binomial distribution data frame using the binomial probability function with successes, probabilities, and cumulative probabilities as columns
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
