% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
cumulative distribution of binomial(trials, prob) in a dataframe
}
\description{
computes the cumulative distribution of binomial(trials, prob)
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
