% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on each trial}
}
\value{
a data frame with the probability distribution and cumulative probabilities
}
\description{
create a data frame with the binomial cumulative distribution
}
\examples{
# the binomial cumulative distribution of 5 trials
# assume the probability of success in each trial is 0.5
bin_cumulative(trials =5, prob = 0.5)
}
