% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{A intiger that is the number of trials}

\item{prob}{The probability of getting a success, must be between 0 and 1}
}
\value{
Return a dataframe with the number of successes in one column and the given probability in another column, with another column showing the cumulative odds
}
\description{
Return a table of the probabilities of each possible amounts of success for a given amount of trials, along with a column showing the cumulative odds
}
