% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of event}
}
\value{
dataframe with sucesses in the first column, probability in the second column, and cumulative in the third column of class c(cinbcum, data.frame)
}
\description{
A main function of a cumulative distribution of a binomial function, from 0 to number of trials, can be used with plot().
}
\examples{
# binomial cumulative distribution with 5 trials and prob = 0.5
z = bin_cumulative(trials = 5, prob = 0.5)

# binomial cumulative distribution with 20 trials and prob = 0.3
bin_cumulative(trials = 20, prob = 0.3)

# plot
plot(z)

}
