% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{parameter p to the binomial distribution}
}
\value{
data.frame with two classes: c("bincum","data.frame")
}
\description{
generates a dataframe of number of successes to probability and cumulative probability, up to trials
}
\examples{

# binomial cumulative distribution
bin_cumulative(trials = 5, prob = 0.5)

}
