% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{:bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\value{
: the data frame
}
\description{
:produce the table of the distribution of probability and culmulative probability as getting the given number of trials
}
