% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
returns a dataframe with two classes, bincum and dataframe, with both the probability distribution and the cumulative probabilities (successes in the first column, porbability in the second column, cumulative in the third column)
}
\description{
calculates the binomial cumulative distribution
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
