% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workout3.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{an integer with number of trials}

\item{prob}{the probability of the success occur}
}
\value{
a data.frame containing the success and corresponding probability as well as cumulatve probability
}
\description{
the function used to calculate the cumulative probability and return a dataframe
}
\examples{
bin_cumulative(10, 0.5)
}
