% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunc.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success within one trial}
}
\value{
the probability and cumulative probability distribution specified by the parameters
}
\description{
creates a data frame with a distribution of probabilites for the specified number of trials along with cumulative probabilities
}
\examples{

# default
bin_cumulative()

# another example
bin_cumulative(trials = 3, prob = 0.5)

}
