% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{numeric vector}

\item{prob}{numeric vector}
}
\description{
cumulates the binomial distribution
}
\examples{

bin_cumulative(trials = 5, prob = 0.5)

x <- class(c("binvar", "data.frame"))

x <- bin_cumulative(trials = 6, prob = 0.3)


}
