% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probabililty of success (numeric)}
}
\value{
data frame with number of successes, respective probabilities, and cumulative probabilities (class bincum)
}
\description{
computes the binomial cumulative distribution for the given number of trials and probability
}
\examples{
bin_cumulative(5, 0.5)
}
