% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{input number of trials}

\item{prob}{input probability}
}
\value{
computed cumulative probability in which k successes can occur in n trials
}
\description{
calculate the cumulative probability from 1 to k successes which can occur in n trials
}
\examples{
bin_cumulative(5,0.5)
}
