% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
dataframe with two classes, bincum and dataframe
}
\description{
computes the cumulative distribution of binomial(trials, prob)
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
