% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{integer representing n, the total number of trials}

\item{prob}{float representing p, the probability of success on a given trial}
}
\value{
Data frame with primary class bincum and with columns trials, probability, and cumulative
}
\description{
Calculates the probability X = k a binomial random variable X = Binomial(n,p) for all k, as well as the cumulative distribution, probability X <= k for all k
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)

bin_cumulative(100,.75)
}
