% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{numeric value for number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
computed cumulative probability
}
\description{
computes the cumulative binomial distribution given number of trials and probability of success
}
\examples{
# binomial cumulative distribution
bin_cumulative(trials = 5, prob = 0.5)

# plotting binomial cumulative distribution
dis2 <- bin_cumulative(trials = 5, prob = 0.5)
plot(dis2)
}
