% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{trials}

\item{prob}{probability}
}
\value{
bincum
}
\description{
Calculates a distribution given a probability of success and given number of trials
}
\examples{
bin_cumulative(5, 0.5)
}
