% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\value{
probability distribution and cumulative probabilities
}
\description{
return data frame with probability distribution and respective cumulative probabilities
}
\examples{
a <- bin_cumulative(5,0.5)
plot(a)
}
