% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
dataframe depicting cumulative binomial distribution
}
\description{
constructs dataframe depicting cumulative binomial distribution
}
\examples{
bin_cum_5 <- bin_cumulative(5,0.5)

bin_cum_10 <- bin_cumulative(10,0.1)

}
