% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{(integer)}

\item{prob}{(real)}
}
\value{
data frame with cumulative probabilities
}
\description{
computes cumulative probabilities and binomial distribution
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
bin_cumulative(7, 0.4)
}
