% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative probabilities}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{the probability of success of each trial}
}
\value{
return a dataframe listing all cumulative probabilities of the binomial distribution
}
\description{
Calculating cumulative probabilities for a binomial distribution
}
\examples{
 success probability cumulative
1       0     0.03125    0.03125
2       1     0.15625    0.18750
3       2     0.31250    0.50000
4       3     0.31250    0.81250
5       4     0.15625    0.96875
6       5     0.03125    1.00000
}
