% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulation of Probability for a Binomial Random Variable}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{success}{probability of success for each trial}
}
\value{
data frame with three columns, one for successes , one for its corresponding probability, and one for the cumulative probability
}
\description{
calculates the probability and cumulative probability of different number of successes for a given number of trials
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
