% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial culmulative distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of a single success}
}
\value{
a data frame displaying the binomial culmulative distribution
}
\description{
display the binomial culmulative distribution
}
\examples{

bin_cumulative(trials = 5, prob = 0.5)

}
