% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_bin_func.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{cumulative binomial}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{some integer}

\item{prob}{some real number}
}
\value{
data frame with cumulative probabilities
}
\description{
computes cumulative probabilities of binomial distribution
}
\examples{
bin_cumulative(trials=5, prob=0.5)
bin_cumulative(7, 0.4)
}
