% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Distribution}
\usage{
bin_cumulative(trials, prob, class = c("bincum", "data.frame"))
}
\arguments{
\item{trials}{numeric, number of trials}

\item{prob}{numeric, probability of success}
}
\value{
a data frame with both the probability distribution and the cumulative probabilities, of two classes-bincum and data.frame
}
\description{
calculates the probability of success for each of the possible number of successes given the number of trials,
as well as the cumulative probability, which adds the current probability to the past ones
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)


}
