% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binary Cumulative Function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
data frame with both the probability distribution and the cumulative probabilities
}
\description{
Calculate the binomial cumulative distribution
}
\examples{
# binolmial cumulative distribution of 5 trials and probability of 0.5
bin_cumulative(trials = 5, prob = 0.5)
 success probability cumulative
1      0     0.03125    0.03125
2      1     0.15625    0.18750
3      2     0.31250    0.50000
4      3     0.31250    0.81250
5      4     0.15625    0.96875
6      5     0.03125    1.00000
}
