% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulation method}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
binomial cumulative distribution of given trials with given probability of success
}
\description{
calculates binomial cumulative distribution
}
\examples{

# binomial cumulative distribution of heads out of five fair coin toss
bin_cumulative(5, 0.5)

# binomial cumulative distribution of heads out of ten loaded coin toss with 25\% chance of heads
bin_cumulative(10, 0.25)

}
