% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main function.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{: function bin_cumulative()}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{:}{trials and prob}
}
\value{
: an object of class \code{"bindis"}
}
\description{
: create binomial cumulative distribution
}
\examples{
:
bin_cumulative(trials = 5, prob = 0.5)
}
