% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
dataframe of cumulative probability of success
}
\description{
create a dataframe of cumulative probability of success
}
\examples{
bin_cumulative(2,.5)
  \dontrun{
 # plot the cumulative probability
 cum = bin_cumulative(2,.5)

 plot(cum)
 }
}
