% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial cumulative density function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on each trial}
}
\description{
the probability of all possible outcomes of a binomial task summed with all less than that value
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
