% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial_main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
return a data frame with both the probability distribution and the cumulative probabilities
}
\description{
list the cumulative probabality of each successe with probability p in n trials
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
