% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{numeric value of number of trials}

\item{prob}{numeric value of probability of success of each trial}
}
\description{
this function calculates the probabilty of getting successes in n trails with prob p of success
and the cumulative probability
}
\examples{
Generate a dataframe of differnet number of successes and the cumulative probability within 5 trials of 0.5
probability of success.
  bin_distribution(trials = 5, prob = 0.5)
}
