% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Probability Distribution and Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of trials, a non-negative integer}

\item{prob}{the probability of success, a number between 0 and 1 both included}
}
\value{
a data frame of primary class bincum and secondary class data.frame with the following columns: success, probability and cumulative probability
}
\description{
calculates and stores in a data frame the probabilities of getting different number of successes in a given number of trials and the cumulative probabilities
}
\examples{
# the distribution of a random variable in 4 trials and probability of success 0.2
bin_cumulative(4,0.2)

# the distribution of a random variable in 60 trials and probability of success 0.6
bin_cumulative(60,0.6)
}
