% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{an integer}

\item{prob}{a double value}
}
\value{
created object
}
\description{
Create an object of class *bincum*
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
