% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{trial numbers}

\item{prob}{success probability}
}
\value{
a data frame containing cumulative distribution of a n trials
}
\description{
get cumulative distribution of a n trials experiment
}
\examples{
dis2 <- bin_cumulative(trials = 5, prob = 0.5)
plot(dis2)
}
