% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Distribution Function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{total number of trials (integer)}

\item{prob}{probability of making success in one trial (decimal)}
}
\value{
a data frame of class "bincum" which contains the different numbers of successes and the corresponding probabilities
and cumulative probabilities
}
\description{
compute the cumulative distribution of making successes in several trials with
corressponding binomial probablity
}
\examples{
# binomial cumulative distribution
bin_cumulative(trials = 5, prob = 0.5)
}
