% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Bin_Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{an integer or vector of numerics - number of trials}

\item{prob}{a numeric value between 0 and 1}
}
\value{
a data frame with the distribution of successes (both individual and cumulative)
}
\description{
takes in trials and prob and outputs a data frame for all success possibilities.
Includes a column for that success alone, as well as a column for that success or fewer
}
\examples{
# default
cum <- bin_cumulative(trials = 5, prob = 0.5)

}
