% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{function used to compute cumulative bins}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{is the probability of success}
}
\value{
returns the daata talbe of success, probability and cumulative
}
\description{
generates cumulative table for trials and probabilities of success
}
\examples{

#cumulative toss graph
bin_cumulative(trials=5, prob=0.5)
dis2 <- bin_cumulative(trials=5, prob=0.5)
plot(dis2)

}
