% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cum_and_related.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{representing # of trials (numeric value)}

\item{prob}{representing probability (numeric value)}
}
\value{
a list of classes \code{"bincum"} and \code{"data.frame"}
}
\description{
determines cumulative distribution of successes a given # of trials and probability
}
\examples{
bin_cumulative(5,0.5)
bin_cumulative(1,1)
}
