% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{create a binomial cumulative distribution table}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success for each trials}
}
\value{
a data frame with the cumulative probability distribution: sucesses in the first column, probability in the second column, cumulative distribution probability in the third column.
}
\description{
the primary class is "bincum" indicating that this is an object of class binomial cumulative distribution. Additionally, to keep this object as a data frame, we still need to include a class "data.frame".
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
