% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{cumulative distribution of binomial}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials:}{number of trials, a non-negative integer}

\item{prob:}{probability of success, a number between 0 and 1, inclusive}
}
\value{
probability and cumulative distributions
}
\description{
calculates the probability distribution and the cumulative distribution
}
\examples{
> bin_cumulative(trials = 5, prob = 0.5)
}
