% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success in 1 trial}
}
\value{
an object of class code "bincum" and "data.frame" with the following elements:
}
\description{
that returns a data frame with the probability distribution and the cumulative distribution
}
\examples{

# binomial cumulative distribution
bin_cumulative(trials = 5, prob = 0.5)
}
