% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
data frame of class "bincum" and "data.frame" with columns success,
probability, and cumulative
}
\description{
computes a data frame with columns listing values for success,
probability, and cumulative distribution
}
