% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Function bin_cumulative()}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability of success}
}
\value{
the probability and cumulative probability distribution data frame
}
\description{
dispaly the probability and cumulative probability of each success times
}
\examples{
# binomial cumulative distribution
 bin_cumulative(trials = 5, prob = 0.5)
}
