% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunction.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomal cumulative}
\usage{
bin_cumulative(n, p)
}
\arguments{
\item{n}{the number of trials}

\item{p}{the probability of successes}
}
\value{
create a data frame of cumilative  probability of a binomial probability distribution

create a line graph of cumilative  probability of a binomial probability distribution
}
\description{
list the  cumilative probabality of each successe with probability p in n trials
}
\examples{
# list a binomial cumilative probability distribution
dis2 <- bin_cumulative(n = 5, p = 0.5)
# show a line graph of a binomial cumilative  probability distribution
plot(dis2)
}
