% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{int}{integer trials}

\item{int}{integer prob}
}
\value{
dataframe with bincum class
}
\description{
returns a dataframe with the probability distribution and the cumulative probabilities
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
