% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{: bin_cumulative()}
\usage{
bin_cumulative(trials, prob)
}
\value{
: data frame with both the probability distribution and the cummulative probabilities
}
\description{
: get data frame with both the probability distribution and the cummulative probabilities
}
