% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative binomial distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
data frame with cumulative binomial probability distribution
}
\description{
displays binomial distrbution with cumulative probabilities
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
