% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{denotes trials}

\item{prob}{denotes prob of  succeses}
}
\value{
the factorial of "n chooses k"
}
\description{
create a cumulative probability distribution
}
\examples{
bin6 <- bin_cumulative()


}
