% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Bin Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
data frame with bincum and data.frame classes
}
\description{
Creates a data frame with probability distribution and cumulative probabilities
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
