% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{numb of trials}

\item{prob}{prob of success}
}
\value{
data frame containing cumulative probability of success in each trial
}
\description{
calculates cumulative chances of all success
}
\examples{
# chances of successes in 10 trials with p=.5
bin_cumulative(10, .5)

# chances of successes in 8 trials with p=.7
bin_cumulative(8, .7)

}
