% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of n attempts (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
a dataframe
}
\description{
creates a dataframe with success, probability, and cumulative as columns
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
