% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{chance of success}
}
\value{
a dataframe containing the distribution and cumulative
}
\description{
gets the discrete distribution and the CDF
}
\examples{
bin_cumulative(5,0.5)
}
