% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{numeric value, number of trials}

\item{prob}{numeric value between 0 and 1, probability of success}
}
\value{
a data frame of possible successes and related (cumulative) probability
}
\description{
computes the (cumulative) probability of getting the possible successes
}
\examples{
#build a binomial distribution
dis2 <- bin_cumulative(trials = 5, prob = 0.5)
#plot it
plot(dis2)
}
