% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_statmanips.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Function}
\usage{
bin_cumulative(trials = 1, prob = 1)
}
\arguments{
\item{trials}{the number of trials for the Binomial Distribution}

\item{prob}{the probability for each independent Bernoulli trial}
}
\value{
an object bincum which contains a dataframe named datafr which contains a column of all potential successes values
and then their corresponding probabilities, and a third column with the cumulative probability.
}
\description{
For a given number of trials and probability of success, calculates the given distribution of successes and the 
cumulative distribution of successes.
}
\examples{
bin_cumulative(5,0.5)
}
