% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_and_methods.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials, }{number of trials}

\item{prob, }{probablity of successes}
}
\value{
return a data frame with both the probability distribution and the cumulative probabilities: sucesses in the first column, probability in the second column, and cumulative in the third column.
}
\description{
creates a data frame of probability, success and cumulative
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
