% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{integer value for number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
dataframe with two classes: bincum and data.frame
#' @examples

bin_cumulative(5, 0.5)
bin_cumulative(5, 0.2)
}
\description{
computes the cumulative probability for success for all values of success
}
