% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_workout3.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Finds the cumulative probability of a distribution (CDF)}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{how many times was the experiment repeated}

\item{prob}{the probability of a success ocurring}
}
\value{
a dataframe like functions above but with additional cumulative function
}
\description{
Data frame with an additional cdf column
}
\examples{
cumulative = cumsum(probability)
}
