% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cum}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{a number}

\item{prob}{a number}
}
\value{
a number
}
\description{
calculates cummulative prob
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
