% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative Function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{trials number of fixed trials}

\item{prob}{probability of success in each trial}
}
\value{
data frame of two classes (bincum and data.frame) that has 3 columns containing the probability
distribution and cumulative probabilities. The first column is successes, second column is probabilities,
and third column is cumulative.
}
\description{
The function takes in two arguments and returns a data frame with two classes: the primary 
class is bincum which is an object with class binomial cumulative distribution, and data frame to keep
the object as a data frame. The data frame returned will contain 3 columns. The first column are the 
successes of the probability distribution, the second column is the respective probabilities, and the 
third column is the cumulative of all the probabilities. The function will also check whether the 
inputs are valid and will return an error if they are not.
}
\examples{
bin_cumulative(trials = 5, prob = 0.2)
}
