% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials; non-negative integer value}

\item{prob}{probability of success; must be a numeric value between 0 and 1 (inclusive)}
}
\value{
data frame with two classes: "bincum" (binomial cumulative distribution) and "data.frame"
}
\description{
Calculates both the probabiltiy distribution and the cumulative probabiltiies of a given number of trials and probability value
}
\examples{
dis2 <- bin_cumulative(5, 0.5)
dis2

# plot cumulative distribution

plot(dis2)
}
