% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{for the time of success}

\item{prob}{for the probability of success}
}
\value{
a data frame with two clss
}
\description{
it forms a data frame that both the probability distribution and the cumulative probability
}
