% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative Binomial Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
An object representing the cumulative binomial distribution
}
\description{
Creates an object of cumulative binomial distribution
}
\examples{
# Create a binomial cumulative distribution object
dis2 <- bin_cumulative(trials = 5, prob = 0.5)
# Create a line graph of the cumulative distribution
plot(dis2)

}
