% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial umulative probability}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{accepts}{trials, probability}
}
\value{
returns array of probabilities for each number of successes increasing cumulatively
}
\description{
cumulative probability of getting any number of successes in n trials
}
