% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{- number of trials}

\item{prob}{- probability of success}
}
\value{
A data frame describing the discrete and probabilities for each possible binomial outcome
from 'trials' trials
}
\description{
Cumulative distribution
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)

}
