% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success on each trial (numeric)}
}
\value{
data frame with probability distribution & cumulative probabilities
}
\description{
returns a data frame with the probability distribution & cumulative probabilities, of class "bincum"
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
