% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{probability of success in a single trial}
}
\value{
bincum : Returns dataframe with primary class 'bincum' and second class 'data.frame'
}
\description{
Displays probability density and cumulative probability of binomial distribution
given # of trials and probability of a single success
}
\examples{
:
bin_cumulative(trials = 5, prob = 0.3)
}
