% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{Number of trials (numeric)}

\item{prob}{Probability of successs (real)}
}
\value{
Returns a data.frame of the number of successes and their probabilities as well as the cumulative probability up to a certain number of successes
}
\description{
Computes the binomial distribution of a certain number of success over a given number of trials as well as the cumulative probability up to that number of successes
}
\examples{
dis2 <- bin_cumulative(trials = 5, prob = 0.5)
dis2
1       0     0.03125    0.03125
2       1     0.15625    0.18750
3       2     0.31250    0.50000
4       3     0.31250    0.81250
5       4     0.15625    0.96875
6       5     0.03125    1.00000

plot(dis2)
Returns a plot of the cumulative probability versus number of successess
}
