% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R, R/hello.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulation}
\usage{
bin_cumulative(trials, prob)

bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability that a trial will be a success}

\item{trials}{number of trials}

\item{prob}{probability that a trial will be a success}
}
\value{
data frame with two classes, succes in the first column, probability in the second column, and cumulative in the third column

data frame with two classes, succes in the first column, probability in the second column, and cumulative in the third column
}
\description{
Returns a data frame with probability distribution and cumulative probabilities

Returns a data frame with probability distribution and cumulative probabilities
}
\examples{
bin_cumulative(5, .5)
bin_cumulative(5, .5)
}
