% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-functions-methods.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
table with probability distribution and cumulative distribution
}
\description{
Returns a table of the binomial probability distribution and the cumulative probability distribution.
}
\examples{
# binomial cumulative distribution
bin_cumulative(trials = 5, prob = 0.5)
}
