% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the specified number of trials for this binomial distribution (numeric)}

\item{prob}{the specificed probability of success for this binomial distribution (numeric)}
}
\value{
a data frame that displays the binomial distribution with number of successes in the left column, the probability in the middle column, and the cumulative probability in the right column, with classes "bindcum" and "data.frame"
}
\description{
Calculates the binomial probability distribution and cumulative binomial probability distribution of a binomial distribution with specified probability and number of trials and displays the result in data frame
}
\examples{
bin_cumulative (5, 0.5)
}
