% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Bin Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
data frame of number of successes, probabilities, and cumulative probabilities
}
\description{
computes a data frame with probability distribution and cumulative probabilities for each success value
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
