% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (n)}

\item{prob}{probability (p)}
}
\value{
dataframe of distributions
}
\description{
calculate both binomial distribution and binomial cumulative distribution
}
\examples{
bin_cumulative(5, 0.5)
}
