% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bino.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
an object of class \code{"bincum"} or \code{"data.frame"}
}
\description{
Gives a table with cumuative density of the binomial distribution
}
\examples{
# 20 trials with probability of success 0.5
cumulative <- bin_cumulative(20, 0.5)

}
