% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_function.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of getting a success in a trial (numeric)}
}
\value{
computed binomial distribution data frame with classes ("bincum" and "data.frame")
}
\description{
Creates a binomial distribution data frame using the binomial probability function with successes, probabilities, and cumulative probabilities as columns
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
