% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
data frame of a binomial cumulative distribution
}
\description{
generate a binomial cumulative distribution
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)

}
