% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
data.frame with probability distribution that has successes in the first column and probability in the second column and cumulative in the third column
}
\description{
calculates probability distribution and cumulative probabilities gets an output that has two classes consisting of bincum and data.frame
}
\examples{
bin_cumulative(5, 0.5)
bin_cumulative(4, 0.6)
}
