% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a data frame containing cumulative probability of binomial distribution
}
\description{
creates a dataframe for cumulative probability of certain success with certain probability in certain number of trials
}
\examples{
bin_cumulative(5, 0.5)
}
