% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Density Function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials, must be non negative int}

\item{prob}{probability of success, must be numeric between 0 and 1}
}
\value{
dataframe with successes in the first column, probabality in the
second column, and cumulative probabilty in the third
}
\description{
creates dataframe with both the probability distribution and
the cumulative probability distribution for binomial with given number of
trials and prob of success
}
\examples{
bin_cumulative(19, .25)
bin_cumulative(trials = 100, prob = .5)
}
