% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{return the distribution of success with cumulative probability}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trails}

\item{prob}{the probability}
}
\value{
a dataframe of distribution with cumulative probability
}
\description{
return the distribution of success with cumulative probability
}
\examples{
bin_cumulative(5, 0.5)
}
