% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{cumulative binomial}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{describe the number of files}

\item{prob}{describe probability}
}
\value{
probability and cumulative probability for each success
}
\description{
gives table for cumulative binomial distribution given parameter
}
\examples{
bin_cumulative(3,0.3)
}
