% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative binomial distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials to attempt}

\item{prob}{probability of success}
}
\value{
returns a bincum variable which is a data.frame containing the number of successes, the probability of that success, and the cumulative probability up to that point.
}
\description{
Computes the binomial distribution while keeping track of the cumulative probability
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
