% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Table}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{Number of trials in the binomial distribution (numeric)}

\item{prob}{Probability of success in the binomial distribution (numeric)}
}
\value{
Returns a dataframe of the binomial distribution with cumulative distribution
}
\description{
Calculates a binomial distribution table with cumulative distributions
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
