% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_prob,distr, cumulative, choose.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative.}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of trials.}

\item{prob}{the probability of success in any given trial.}
}
\value{
a data frame with sucesses in the first column, probability in the second column, and cumulative in the third column.
}
\description{
computes the probability distribution and the cumulative probabilities.
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)

bin_cumulative(8, 0.1)
}
