% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Bin Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
an object of class \code{c("bincum", "data.frame")}
}
\description{
Creates an object of class \code{c("bincum", "data.frame")}
}
\examples{
# default
bincum <- bin_cumulative(5, 0.5)

}
