% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin cumulative function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (intger)}

\item{prob}{number of success (intger)}
}
\value{
a dataframe with both the probability distribution and the cumulative distribution of each successes certain with probability and trials
}
\description{
calculate the binomial probability distribution and the cumulative distribution
}
