% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binom_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(n = 5, p = 0.5)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
dataframe with successes and cumulative successes
}
\description{
returns dataframe with both the probabilities and the successes in the first column
}
