% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Max Sima .R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Function bin_cmulative()}
\usage{
bin_cumulative(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
data frame with the probability distribution
}
\description{
create a dataframe with 2 classes of the binomial distribution
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
