% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic Functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{n}{number of trials}

\item{p}{success rate of each trials}
}
\value{
A data frame with number of success, probability and cumulative probability
}
\description{
a dataframe with the probability and cumulative probability
}
\examples{
plot_bincum(bin_cumulative(trials=5, prob=0.5))
}
