% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Bin cumulative function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{Number of (fixed) trials in the binomial distribution.}

\item{prob}{the probability of success on each trial.}
}
\value{
a data frame with both the probability distribution and the cumulative probabilities
}
\description{
a data frame with both the probability distribution and the cumulative probabilities
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
