% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of (fixed) trials}

\item{prob}{the probability of success on each trial}
}
\value{
a data frame with the probability distribution and cumulative probability
}
\description{
return a data frame with both the probability distribution and the cumulative probabilities: success, probability, and cumulative
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
