% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{a dataframe of cumulative binomial distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{integer prob double}
}
\value{
the dataframe of cumulative binomial distribution
}
\description{
base on the binomial probability, create a dataframe ofcumulative binomial distribution
}
\examples{
bin_cumulative(trials=5,prob=0.5)
}
