% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function-bincumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulative Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{numeric values. number of trials}

\item{prob}{numeric values. probability of success}
}
\value{
Cumulative Distribution data frame
}
\description{
A table of the cumulative probability of all possible successes
}
\examples{
  bin_cumulative(trials = 5, prob = 0.5)
}
