% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative()}
\usage{
bin_cumulative(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{the number of the total trials}

\item{prob}{the probability of a succeess}
}
\value{
dataframe of a binomial distribution it returns a da.frame with two classes
bindis and data.frame under the binomial distribution of given trials and probability
}
\description{
bin_cumulative() is a function that compute the total probability under given numbers of trials and success
}
\examples{
# binomial cumulative density distribution,
 bin_cumulative()
 bin_cumulative(5, 0.5)
}
