% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{value of trials}

\item{prob}{probability of one independent success}
}
\description{
obtain a data frame with both the probability distribution and the cumulative probabilities
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)

}
