% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial distribution function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of totoal trials (integer)}

\item{prob}{the possibility of each success (numeric)}
}
\value{
a dataframe of probabilities distribution
}
\description{
show the distribution of probabilities of each success in n trials
and the cumulative probabilities.
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
