% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability (numeric)}
}
\value{
computed binomial cumulative distribution
}
\description{
calculates the binomial cumulative distribution
}
\examples{
bin_cumulative(trials=5,prob=0.5)
}
