% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of the trials}

\item{prob}{the probability of success on each trial}
}
\value{
an object of class \code{"bindcum"}
}
\description{
binomial cumulative distribution
}
\examples{
bin_cumulative(trials = 10, prob = 0.3)
}
