% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{number of success}
}
\value{
the probability distribution and cumulative probabilities
}
\description{
Calculates the probability distribution and cumulative probabilities
}
\examples{
 \dontrun{
bin_cumulative(trials=5L,prob=0.5)
 }
}
