% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability of success in each trial}
}
\value{
an object of both 'bincum' and 'data.frame' that has columns of success, probability, and cumulative probability
}
\description{
constructs a table of successes, probability, and cumulative probability.
}
