% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions and methods.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{success}{success rate of each trials}
}
\value{
A data frame with number of success, probability and cumulative probability
}
\description{
calculates the probability and cumulative probability of different success times
}
\examples{
bin_cumulative(5,0.5)
}
