% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the numeric vector of total trials}

\item{prob}{the numeric vector of probability for success of one single trial}
}
\value{
the returned output: a data.frame with two classes: c("bincum", "data.frame")
}
\description{
output a data frame with both the probability distribution and the cumulative probabilities: sucesses in the first column, probability in the second column, and cumulative in the third column
}
\examples{
bin_cumulative(10, 0.3)
}
