% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial cumulative distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability if success}
}
\value{
an object of class \code{"bincum"}

an object of class \code{"data.frame"}
}
\description{
compute the probability distribution and
the cumulative distribution
}
\examples{
# binomial cumulative distribution of different success in five trials
bin_cumulative(trials = 5, prob = 0.5)
}
