% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Bin Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability}
}
\value{
data.frame -> df
}
\description{
forms a table with which each success is given a probability and cumulative is added
}
