% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Cumulated Distribution}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
Data frame with the Probability Value and Cumulated Value
}
\description{
Computes Cumulated distribution
}
\examples{
dis2 <- bin_cumulative(trials = 5, prob = 0.5)
plot(dis2)
}
