% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All Functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability value (numeric)}
}
\value{
data frame with probability distribution and cumulative probabilities
}
\description{
create a data frame with probability distribution and cumulative probabilities
}
\examples{
bin_cumulative(5, 0.5)
}
