% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number indicating the number of trials}

\item{prob}{number indicating teh probability of getting success}
}
\value{
the data frame with the probability distribution and the cumulative of probability
}
\description{
get a data frame that the first column is the number of success, the second column is the probability, and the third column is the cumulative probability
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
