% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-adele-shen.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Probability Distribution Function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
data.frame of classes "bindis" and "data.frame" with the probability of the number of successes
occurring in the number of trials
}
\description{
Calculates probabilities about the number of successes in a fixed number of random trials
performed under identical conditions (assuming a constant probability of success on each trial)
}
\examples{
# binomial probability distribution
bin_distribution(trials = 5, prob = 0.5)
}
