% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of head probability}
}
\value{
a data frame with the probability distribution: sucesses in the first column, probability in the second column
}
\description{
calculates the probabilities of different number of successes occuring in fixed number of trials
}
\examples{

# binomial probability distribution
bin_distribution(trials = 5, prob = 0.5)

}
