% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_probability.R, R/workout03-ann-deng.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution()}
\usage{
bin_distribution(trials, prob)

bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
data frame with probability distribution with class "bindis"

data frame with probability distribution with class "bindis"
}
\description{
uses bin_probability to output probability distribution

uses bin_probability to output probability distribution
}
\examples{
bin_distribution(5, 0.5) ## returns data.frame(c(0, 1, 2, 3, 4, 5), c(0.03125, 0.15625, 0.31250, 0.31250, 0.15625, 0.03125))
bin_distribution(5, 0.5) ## returns data.frame(c(0, 1, 2, 3, 4, 5), c(0.03125, 0.15625, 0.31250, 0.31250, 0.15625, 0.03125)) 
}
