% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomal distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials (non-negative integer)}

\item{prob}{probability of success (real number between 0 and 1)}
}
\value{
a "bindis" class object with "data.frame" secondary class.
data frame has two columns success and probability
function includes plot function to plot histogram of success and probability
}
\description{
calculates the binomial distribution given number of trials and probability of success
}
