% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3-divya-vijay.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a dataframe of class binomial distribution
}
\description{
creates a dataframe of a probability distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
