% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Bin Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{an integer, the number of trials}

\item{prob}{an integer of the probability of success for one trial}
}
\value{
a data frame with the probability for all possible number of successes given the number of trials and the probability of success.
}
\description{
A function to create a data frame displaying the probability for all possible number of successes given the number of trials and the probability of success.
}
\examples{
bin_distribution(10,0.5)
}
