% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
data frame with the probability distribution
}
\description{
Computes binomial distribution
}
\examples{
dis <- bin_distribution(trials = 5, prob = 0.5)
dis

plot(dis)
}
\seealso{
\code{\link{bin_cumulative}}
}
