% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability value}
}
\value{
an object of class \code{c("bindis", "data.frame")}
}
\description{
creates an object of class \code{c("bindis", "data.frame")}
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
