% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{vector of number of trials (numeric)}

\item{prob}{vector of probabilities (numeric)}
}
\value{
data frame containing successes in one column and probabilities in another
}
\description{
calculates the binomial distribution with given params
}
\examples{

bin_distribution(trials = 5, prob = 0.5)

dis11 <- bin_distribution(trials = 5, prob = 0.5)
plot(dis1)

}
