% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials (n >= 0)}

\item{prob}{the probaility to win in oone trials}
}
\value{
A distribution table of particular probability
}
\description{
To calculate the distribution of particular probaility
}
