% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Calculate probability distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{integer the number of trials}

\item{prob}{the probability to success in each trials}
}
\value{
The probability of getting every possible number of success(es) in k trials
}
\description{
Calculates the distribution of the probability of getting possible number of success(es) in k trials
}
\examples{
# Calculate the probability of getting 0, 1, 2, 3 success out of 3 trials, each trial with
# independent probability 0.5
bin_distribution(trials = 3, prob = 0.5)
}
