% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a data frame with the probability distribution (successes in 1st column, & probability in 2nd column)
}
\description{
calculates the probability distribution of successes in given no. of trials
}
\examples{
bin_distribution(trials = 5, prob = 0.5)

}
