% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
computed probability
}
\description{
calculates the probability of getting x number of successes in y trials.
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
