% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{: Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{:}{rials, prob}
}
\value{
: dataframe of probabilities of different success in trials
}
\description{
: Calculates the different probabilities of successes in n trials
}
\examples{
: bin_distribution(5, 0.5)
}
