% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Bin Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a data frame with the binomial distribution
}
\description{
calculates the probability of binomial distribution
}
\examples{
bin_distribution(trials=5, prob = 0.5)

}
