% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunc.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{Probability that a success occurs}
}
\value{
data frame with: success; probability distribution
}
\description{
calculate Binomial distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
