% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution Function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{success probability}
}
\value{
Dataframe with binomial distribution
}
\description{
Creates binomial distribution, using given binomial probability and trials.
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
bin_distribution(trials = 10, prob = 0.3)
}
