% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
binomial probability distribution
}
\description{
to calculate binomial probability distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
