% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
data frame of the binomial distribution
}
\description{
calculates the binomial distribution given the number of trials
and probability of success
}
\examples{
bin_distribution(5, 0.5)
}
