% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-code.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution Data Frame}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of getting a number of successes}
}
\value{
a probability of success data frame
}
\description{
creates data frame with the probabilities of successes within a number of trials
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
