% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of total trials}

\item{prob}{probabilty of success in each trial}
}
\value{
data.frame with successes in the first column, probaility in the second column
}
\description{
calculates the probability of all possible numbers of successes occurring
}
\examples{
bin_distirbution(5, 2)
}
