% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials, }{prob}
}
\value{
a data.frame with two classes: c("bindis", "data.frame")
}
\description{
to get a dataframe of binomial probability distribution for each trial
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
