% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success per trial}
}
\value{
a data.frame with two classes
}
\description{
calculating the probability distribution
}
\examples{
# binomial probability distribution
bin_distribution(trials = 5, prob = 0.5)
}
