% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{number that is the probability of success}
}
\value{
bindis a dataframe with the probability distribution: successes in first column, probability in second
}
\description{
computes probability distribution for set amount of trials
}
\examples{
#2 success in 5 trials
bin_distribution(trials=5, success=2, prob = 0.5)

bin_distribution(trials=3, succcess=1, prob =0.3)
}
