% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials, }{number of flips}

\item{prob, }{probability of observing heads}
}
\value{
dataframe object detailing the distribution
}
\description{
Calculates the distribution of a binomial random variable
}
