% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindist.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials, integer}

\item{prob}{probability of success}
}
\value{
data frame with probability distribution
}
\description{
computes the binomial distribution
}
\examples{
bin_distribution(5,0,5)
}
