% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability value}
}
\value{
data.frame with two classes: c("bindis", "data.frame")
}
\description{
creates a data frame of a binomial probability distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
