% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of achieving success}

\item{success}{number of success}
}
\value{
a data frame with the probability distribution: sucesses in the first column, probability in the second column
}
\description{
create a data frame with the probability distribution: sucesses in the first column, probability in the second column
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
