% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability value of the binomial variable}
}
\value{
a dataframe with probability distribution: success in first column, probability in the second column
}
\description{
calculates the distribution for a binomial variable
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
