% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials:}{total number of trials}

\item{prob:}{probability of success}
}
\value{
data.frame of success and probability
}
\description{
calculates probability distribution in a table
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
