% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials (single value)}

\item{prob}{probability of success}
}
\value{
Data frame with successes from 0 to number of trials, and
corresponding probabilities. Has both classes "bindis" and "data.frame"
}
\description{
Computes probability distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
