% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the probability of success on each trial (0 ≤ p ≤ 1)}

\item{prob}{the probability of success on each trial (0 ≤ p ≤ 1)}
}
\value{
a data frame with the probability distribution: sucesses in the first column, probability in the second column
}
\description{
a main function that calculates probability
}
