% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_methods.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution Function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success of individual trial}
}
\value{
\item{data}{a data.frame with two classes: c("bindis","data.frame")}
}
\description{
creates a data frame of binomial distribution
}
\examples{
# binomial probability distribution
bin_distribution(trials = 5, prob = 0.5)
}
