% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Finds binomial distribution}
\usage{
bin_distribution(prob, trials)
}
\arguments{
\item{prob}{probability of random variable}

\item{trials}{number of trials}
}
\value{
an object of class \code{"c("bindis", "data.frame")"}
}
\description{
Finds the probability for each possible number of successes
}
\examples{
 \dontrun{
 # find probability distribution of a binomial random variable
 prob <- 0.5
 trials <- 5
 pdf <- bin_distribution(prob, trials)
 plot(pdf)
 }
}
