% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Probability mass function of a binomial random variable}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of I.I.D. Bernoulli trials}

\item{prob}{the shape parameter of each I.I.D. Bernoulli trial.}
}
\value{
a data frame representing the probability mass function of a binomial
  random variable with the specified shape parameters.
}
\description{
Probability mass function of a binomial random variable
}
\examples{
  \dontrun {
  dis = bin_distribution(trials = 10, prob = 0.5)
  support = dis$success
  }
}
