% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution()}
\usage{
bin_distribution(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{prob}
}
\value{
two data.frame with two classes: c("binds", "data.frame")
}
\description{
create a data.frame that contains binomial distribution
}
