% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of successes}
}
\value{
a data frame indicating the probabality of each number of successes
}
\description{
creates a dataframe for the probability of each number of successes with specified probability in a specified number of trials
}
\examples{
What is the binomial distribution with prob of success = 0.5 in 5 trials?
bin_distribution(trials = 5, prob = 0.5)
}
