% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability if success}
}
\value{
a data.frame with two classes: c("bindis", "data.frame")
}
\description{
compute the probability distribution
}
\examples{
# binomial probability distribution of different successes in five trials
# (assume the probability of success is 0.5)
bin_distribution(trials = 5, prob = 0.5)
}
