    
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{must be an integer, greater than zero}

\item{prob}{a number between 0 and 1}
}
\value{
a dataframe with the # of successes, and the associated probability distribution
}
\description{
probability distribution for a given number of trials, output can also be directly plotted using plot()
}
\examples{
bin_distribution(trials = 10, prob = 0.5)
bin_distribution(8, 0.4)
}