% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a data frame with the probability distribution
}
\description{
make a dataframe that shows probability for each number of success
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
