% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial probability distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on rach trial}
}
\value{
binomial probability distribution
}
\description{
calculates binomial probability distribution
}
\examples{
# binomial probability distribution
bin_distribution(trials = 5, prob = 0.5)

}
