% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{n}{number of trials, prob probablity of success in each trial}
}
\value{
a dataframe with two classes: "bindis" and "data.frame"
}
\description{
calculates the binomial distribution with n trials and probability prob
}
\examples{
# Calculates the distribution of probabilities of each possible number of successes in a dataframe with 5 trials and a probability of .5
bin_distribution(trials = 5, prob = 0.5)

}
