% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{prob}{probablity of success}

\item{n}{number of trials}
}
\value{
a dataframe with class "bindis"
}
\description{
calculates the binomial PMF with n trials and probability prob
}
\examples{
# Calculates the PMF of a binomial distribution of 6 trials and a success probability of .5
bin_distribution(trials = 6, prob = 0.5)

}
