% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probablity of success}
}
\value{
dataframe with probability distribition: successes in first column and probablity in second
}
\description{
Binomial Distribution as a dataframe with successes and probabliyu
}
\examples{
bin_dis <- bin_distribution(trials = 5,prob=0.5)
plot(bin_dis)
}
