% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binominal distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{how many trials}

\item{prob}{probability of a success in any given trial}
}
\value{
data frame with the probability distribution: sucesses in the first column, probability in the second
}
\description{
gives data frame with the corresponding probability distribution
}
\examples{
bin_distribution(5, 0.5)
}
