% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial probability distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{(numeric) the number of trials}

\item{prob}{(numeric) the probability of success on each trial}
}
\value{
A dataframe describing the the full binomial probability distribution given by the trial and probability parameters.
}
\description{
Computes the full binomial probability distribution given by the trial and probability parameters.
}
\examples{
# binomial probability distribution with 20 trials and 45\% chance of success on each trial
bin_distribution(20, 0.45)
}
