% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, probs)
}
\arguments{
\item{trials}{: The number of trials for the experiment}

\item{prob}{: This is a probability value}
}
\value{
a data frame of possible successes and associated probability values
}
\description{
This calculates the probability of successes
}
\examples{
#Builds a binomial distribution
dis1 <- bin_distribution(trials = 5, probs = 0.5)
#plot it
plot(dis1)
}
