% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{prob}{probablity of success for each trial}
\item{n}{number of trials}
}
\value{
a dataframe that contains the number of success and probability
}
\description{
calculates the probability of different number of success trials
}
\examples{
bin_distribution(5, 0.5)
}
