% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin distribution}
\usage{
bin_distribution(n, p)
}
\arguments{
\item{n}{trials}

\item{p}{prob}
}
\value{
probability
}
\description{
calculates the distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
