% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials.}

\item{prob}{probability of success.}
}
\value{
binomial distribution data frame
}
\description{
create a data frame with the binomial probability distribution
}
\examples{

bin_distribution(trials = 5, prob = 0.5)
bin_distribution(trials = 100, prob = 0.45)

}
