% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomal.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial distribution Function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
A dataframe representing the binomial distribution with two classes ("bindis" and "data.frame)
}
\description{
Creates a binomial distribution data frame using the binomial probability function with successes and probabilities as columns
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
