% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
probability distribution of binomial(trials, prob) in a dataframe
}
\description{
computes the probability distribution of binomial(trials, prob)
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
