% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on each trial}
}
\value{
a data frame with the probability distribution
}
\description{
create a data frame with the probability distribution
}
\examples{
# the probability distribution of 5 trials
# assume the probability of success in each trial is 0.5
bin_distribution(trials =5, prob = 0.5)
}
