% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of event}
}
\value{
dataframe of class c(bindis, data.frame)
}
\description{
A main function including a dataframe of successes and probabilitis with two classes c(bindis, data.frame), can be used with plot().
}
\examples{
# binomial probability distribution with 5 trials and prob = 0.5
bin_distribution(trials = 5, prob = 0.5)

# plot of binomial distribution with 5 trials and prob = 0.5
plot(bin_distribution(trials = 5, prob = 0.5))

}
