% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(n, p)
}
\arguments{
\item{n}{the total times of trials}

\item{p}{the probability of success of single trial}
}
\value{
a data frame recording the times of success and corresponding probabilities
}
\description{
form the distribution table
}
