% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{parameter p to the binomial distribution}
}
\value{
data.frame with two classes: c("bindis","data.frame")
}
\description{
generates a dataframe of number of successes to probability, up to trials
}
\examples{

# binomial probability distribution
bin_distribution(trials = 5, prob = 0.5)

}
