% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{is the number of trials}

\item{prob}{is the probability of success}
}
\value{
a data frame with the probability distribution
}
\description{
calculates a data frame with the probability distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
