% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{: bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\value{
: the data frame
}
\description{
:produce the table of the distribution of probability as getting the given number of trials
}
