% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
data.frame with two classes, "bindis" and "data.frame" with the probability distribution (successes in the first column, probability in the second column)
}
\description{
calculates the probabilities of all possible number of successes for a binomial distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
