% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workout3.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{an integer with number of trials}

\item{prob}{the probability of the success occur}
}
\value{
a data.frame containing the success and corresponding probability
}
\description{
the function used to calculate the probability and return a dataframe
}
\examples{
bin_distribution(10, 0.5)
}
