% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials (numeric)}

\item{prob}{the probability of one success (numeric)}
}
\value{
an object containing the probability distribution of the given parameters (data.frame, bindis)
}
\description{
creates a bindis and data.frame object
}
\examples{
dis1 <- bin_distribution(trials = 5, prob = 0.5)
plot(dis1)
}
