% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probabililty of success (numeric)}
}
\value{
data frame with number of successes and respective probabilities (class bindis)
}
\description{
computes the binomial distribution for the given number of trials and probability
}
\examples{
bin_distribution(5, 0.5)
}
