% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial probability}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{input number of trials}

\item{prob}{input probability}
}
\value{
computed probability in which k successes can occur in n trials
}
\description{
calculate the probability from 1 to k successes which can occur in n trials
}
\examples{
bin_distribution(5,0.5)
}
