% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success in each trial}
}
\value{
binomial distribution
}
\description{
displays the distribution in table format
}
\examples{
# binomial probability distribution
bin_distribution(trials = 5, prob = 0.5)
}
