% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
dataframe with two classes, bindis and dataframe
}
\description{
computes the probability distribution of binomial(trials, prob)
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
