% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{integer representing n, the total number of trials}

\item{prob}{float representing p, the probability of success on a given trial}
}
\value{
Data frame with primary class bindis and with columns trials and probability
}
\description{
Calculates the probability X = k a binomial random variable X = Binomial(n,p) for all k. Returns a data frame with primary class bindis.
}
\examples{
bin_distribution(trials = 5, prob = 0.5)

bin_distribution(100,.75)
}
