% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{numeric value for number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
computed binomial probability
}
\description{
computes the binomial distribution given number of trials and probability of success
}
\examples{
# binomial probability distribution
bin_distribution(trials = 5, prob = 0.5)

# plotting binomial probability distribution
dis1 <- bin_distribution(trials = 5, prob = 0.5)
plot(dis1)
}
