% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{trials}

\item{prob}{probability}
}
\value{
bindis
}
\description{
Calculates a distribution given a probability of success and given number of trials
}
\examples{
bin_distribution(5, 0.5)
}
