% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\value{
probability distribution
}
\description{
return a data frame with the probability distribution
}
\examples{
#Build a binomial distribution with 5 trials, prob = 0.5
a <- bin_distribution(trials=5,prob=0.5)
plot(a)
}
