% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials:}{the number of trials.}

\item{prob:}{probability of success}
}
\value{
data frame of numeric output of successes and probability
}
\description{
dives a data frame with the probability distribution, successes in first column and probabilites in the second column.
}
