% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
dataframe depicting binomial distribution
}
\description{
constructs dataframe depicting binomial distribution
}
\examples{
bin_dis_5 <- bin_distribution(5,0.5)

bin_dis_10 <- bin_distributions(10,0.1)
}
