% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
With valid input of parameters, return the a data frame of the probability distribution
}
\description{
compute the binomial distribution with given trial and success input.
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
