% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Distribution of a Binomial Random Variable}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success for each trial}
}
\value{
data frame with two columns, one for successes and one for its corresponding probability
}
\description{
calculates the probability of different number of successes for a given number of trials
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
