% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial proability distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the proability of a single success}
}
\value{
a data frame displaying the binomial proability distribution
}
\description{
display the data frame of a binomial proability distribution
}
\examples{

 bin_distribution(trials = 5, prob = 0.5)

}
