% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_bin_func.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{some integer}

\item{prob}{some real number}
}
\value{
data frame with binomial probability distribution
}
\description{
computes binomial probability distribution
}
\examples{
bin_distribution(trials=5, prob=0.5)
bin_distribution(7, 0.4)
}
