% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob, class = c("bindis", "data.frame"))
}
\arguments{
\item{trials}{numeric, number of trials}

\item{prob}{numeric, probability of success}
}
\value{
a data frame with the probability distribution of two classes-bindis and data.frame
}
\description{
calculates the probability of success for each of the possible number of successes given the number of trials
}
\examples{
bin_distribution(trials = 5, prob = 0.5)


}
