% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binary Distribution Function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
data frame with the probability distribution: success in the first column, probability in the second column
}
\description{
Calculate binary distribution using bin_probability()
}
\examples{
# binomial probability distribution with trials = 5, prob = 0.5
bin_distribution(trials = 5, prob = 0.5)
 success probability
1      0     0.03125
2      1     0.15625
3      2     0.31250
4      3     0.31250
5      4     0.15625
6      5     0.03125
}
