% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution method}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
binomial probability distribution of given trials with given probability of success
}
\description{
calculates binomial probability distribution
}
\examples{

# binomial probability distribution of heads out of five fair coin toss
bin_distribution(5, 0.5)

# binomial probability distribution of heads out of ten loaded coin toss with 25\% chance of heads
bin_distribution(10, 0.25)

}
