% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-methods.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Probability Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
a data frame with the probability distribution: sucesses in the first column, probability in the second column
}
\description{
create a data frame with the binomial probability distribution
}
\examples{
# Binomial Probability Distribution
bin_distribution(trials = 5, prob = 0.5)

}
