% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{nubmer of trials}

\item{prob}{probsbility of success}
}
\value{
data.frame dataframe of binomial distribution
}
\description{
create dataframe of binomial distribution
}
\examples{
bin_distribution(2,.5)
  \dontrun{
 # plot the distribution
 dis = bin_distribution(2,.5)

 plot(dis)
 }
}
