% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial probability distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on each trial}
}
\description{
the probability of all possible outcomes of a binomial task
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
