% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial_main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
return a data frame with the probability distribution
}
\description{
list the probabality of each successe with probability p in n trials
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
