% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials, a non-negative integer}

\item{prob}{the probability of success, a number between 0 and 1 both included}
}
\value{
a data frame of primary class bindis and secondary class data.frame with two columns and that contains the probabilities of getting different numbers of successes under a given number of trials
}
\description{
calculates and stores in a data frame the probabilities of getting different number of successes in a given number of trials
}
\examples{
# the binomial distribution of a random variable in 11 trials and probability of success of 0.42 per trial
bin_distribution(11,0.42)

 # the binomial distribution of a random variable in 3 trials and probability of success of 0.8 per trial
bin_distribution(3,0.8)
}
