% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{an integer}

\item{prob}{a double value}
}
\value{
created object
}
\description{
Create an object of class *binomial distribution*
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
