% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{trial numbers}

\item{prob}{success probability}
}
\value{
a data frame containing distribution of a n trials experiment
}
\description{
get distribution of a n trials using the probability
}
\examples{
dis1 <- bin_distribution(trials = 5, prob = 0.5)
plot(dis1)
}
