% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Probability Distribution Function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{total number of trials (integer)}

\item{prob}{probability of making success in one trial (decimal)}
}
\value{
a data frame of class "bindis" which contains the different numbers of successes and the corresponding probabilities
}
\description{
compute the possibility distribution of making successes in several trials with
corressponding binomial probablity
}
\examples{
# binomial probability distribution
bin_distribution(trials = 5, prob = 0.5)
}
