% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Bin_Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{an integer or vector of numerics - number of trials}

\item{prob}{a numeric value between 0 and 1}
}
\value{
a data frame with the distribution of successes
}
\description{
takes in trials and prob and outputs a data frame for all success possibilities.
}
\examples{
# default
distr <- bin_distribution(trials = 5, prob = 0.5)

}
