% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Probability Distribution}
\usage{
bin_distribution(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a data frame with the probability distribution
}
\description{
to create a dataframe of the binomial distribution
}
\examples{
bin_distribution(5, 0.5)
}
