% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{(integer) the number of trials for the binomial distribution you want to calculate the distribution of}

\item{prob}{(double) a number giving the probability of success on each trial for the binomial distribution}
}
\value{
a data.frame with class "bindis" and "data.frame" that contains a column for the number of successes from 0 to trials and a column for the probability of observing that number of successes in the given number of trials
}
\description{
given a number of trials and a probability of success produces a data.frame with class bindis that contains the probability of each number of successes
}
\examples{
bin_distribution(15, 0.5)
bin_distribution(trials = 5, prob = 0.75)
}
