% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution()}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of one success}
}
\value{
the binomial distribution
}
\description{
displays the binomial distribution as data frame
}
