% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
data frame
}
\description{
creates a data frame with the probability distribution: successes in the first column, probability in the second column
}
