% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_dis_and_related.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{representing # of trials (numeric value)}

\item{prob}{representing probability (numeric value)}
}
\value{
a list of classes \code{"bindis"} and \code{"data.frame"}
}
\description{
determines probability distribution of successes a given # of trials and probability
}
\examples{
bin_distribution(5,0.5)
bin_distribution(1,1)
}
