% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{create a binomial distribution table}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success for each trials}
}
\value{
return a data frame with the probability distribution: sucesses in the first column, probability in the second column.
}
\description{
the primary class is "bindis" indicating that this is an object of class binomial distribution.This function should return a data frame with the probability distribution: sucesses in the first column, probability in the second column.
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
