% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{bin_distribution}
\alias{bin_distribution}
\title{distribution of binomial}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials:}{number of trials, a non-negative integer}

\item{prob:}{probability of success, a number between 0 and 1, inclusive}
}
\value{
probability distribution of each number of successes
}
\description{
calculates the probability distribution given trials and success prob
}
\examples{
> bin_distribution(trials = 5, prob = 0.5)
}
