% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_distribution}
\alias{bin_distribution}
\title{distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success in 1 trial}
}
\value{
an object of class "bindis" and "data.frame" with the following elements:
}
\description{
that returns a data frame with the probability distribution
}
\examples{

# binomial probability distribution
bin_distribution(trials = 5, prob = 0.5)
}
