% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
data frame of class "bindis" and "data.frame" with first
column successes and second column probability
}
\description{
creates a data frame with the probability distribution
of successes and respective probability
}
