% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Function bin_distribution()}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability of success}
}
\value{
the probability distribution data frame
}
\description{
dispaly the probability of different success times.
}
\examples{
# binomial probability distribution
bin_distribution(trials = 5, prob = 0.5)
}
