% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunction.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomal distribution}
\usage{
bin_distribution(n, p)
}
\arguments{
\item{n}{the number of trials}

\item{p}{the probability of successes}
}
\value{
create a data frame of a binomal probability distribution

create a barplot of a binomal probability distribution
}
\description{
list the probabality of each successe with probability p in n trials
}
\examples{
# list a binomial probability distribution
dis1 <- bin_distribution(n= 5, p = 0.5)
# show a barplot of a binomial probability distribution
plot(dis1)
}
