% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{int}{integer trials}

\item{int}{integer prob}
}
\value{
dataframe with bindis class
}
\description{
returns a data frame with the probability distribution, columns are successes and probability
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
