% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{: bin_distribution()}
\usage{
bin_distribution(trials, prob)
}
\value{
: data frame with the binomial probability distribution
}
\description{
: get data frame with the binomial probability distribution
}
