% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Probability Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{denotes trials}

\item{prob}{denotes probability of success}
}
\value{
a table of each probability given each value of success
}
\description{
finds a probability distribution given values of trials, success, and probability
}
\examples{

 bin5 <- bin_distribution()

}
