% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a data frame with two classes: trials and the probability of successes in each trial
}
\description{
calculates probability of all successes
}
\examples{
# chances of successes in 10 trials with p=.5
bin_distributoin(10, .5)

# chances of successes in 8 trials with p=.7
bin_distribution(8, .7)

}
