% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials, }{total number of trials}

\item{prob, }{chance of success}
}
\value{
data frame contains the distribution
}
\description{
gets the whole discrete distribution
}
\examples{
bin_distribution(5,0.5)
}
