% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{numeric value, number of trials}

\item{prob}{numeric value between 0 and 1, probability of success}
}
\value{
a data frame of possible successes and related probability
}
\description{
computes the probability of getting the possible successes
}
\examples{
#Build a binomial distribution
dis1 <- bin_distribution(trials = 5, prob = 0.5)
#plot it
plot(dis1)
}
