% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_statmanips.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution Function}
\usage{
bin_distribution(trials = 1, prob = 1)
}
\arguments{
\item{trials}{the number of trials for the Binomial Distribution}

\item{prob}{the probability for each independent Bernoulli trial}
}
\value{
an object bindis which contains a dataframe named datafr which contains a column of all potential successes values
and then their corresponding probabilities.
}
\description{
For a given number of trials and probability of success, calculates the given distribution of successes.
}
\examples{
bin_distribution(5,0.5)
}
