% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
data.frame with two classes: c("bindis", "data.frame")
}
\description{
calculate the probabilities about the number of successes in a fixed number of random trials performed under identical conditions
}
\examples{

bin_distribution(trials = 5, prob = 0.5)
}
