% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_and_methods.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials, }{number of trials}

\item{prob, }{probablity of successes}
}
\value{
a data frame with the probability distribution: sucesses in the first column, probability in the second column
}
\description{
creates a data frame of probability distrubtion and successes
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
