% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{integer value for number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
vector of binomial probabilities
#' @examples

bin_distribution(5, 0.5)
bin_distribution(5, 0.2)
}
\description{
computes the probability for success for all values of success
}
