% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_workout3.R
\name{bin_distribution}
\alias{bin_distribution}
\title{The entire Binomial Formula with n independent trials and binary options}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{success}{how many successes can be measured in n trials}

\item{trials}{how many times was the experiment repeated}

\item{prob}{the probability of a success occurring}
}
\value{
a certain probability
}
\description{
An output of a dataframe and its plot over a course of n trials
}
\examples{
bin_dataframe <- data.frame(bin_probability(success, trials, prob))
}
