% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{a number}

\item{prob}{a number}
}
\value{
a dataframe
}
\description{
calculates distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
