% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution Function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{probability of success in each trial}
}
\value{
data frame with two classes with the probability distribution and the first column will be
successes and the second column will be the probability.
}
\description{
The bin_distribution Function takes in two arguments and returns a data frame with two classes:
The primary class is "bindis" which indicates the object is of the binomial distribution class. The other class
is the data.frame class. This keeps the object as a data frame. The first column of the data frame will be
the successes of the probability distribution and the second column will contain its respective probability.
The function will also return an error if either arguments entered are invalid.
}
\examples{
class(bin_distribution(trials = 10, prob = 0.5))
}
