% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials; non-negative integer value}

\item{prob}{probability of success; must be a numeric value between 0 and 1 (inclusive)}
}
\value{
data frame with two classes: "bindis" (binomial distribution) and "data.frame"
}
\description{
Calculates values of a probability distribution
}
\examples{
dis1 <- bin_distribution(5, 0.5)
dis1

# plot probability histogram of binomial distribution

plot(dis1)
}
