% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{The probability of getting a success}
}
\value{
Return a dataframe (table) of the binomial distribution
}
\description{
form a table of the probabilities under different time of success
}
