% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
An object representing the binomial distribution
}
\description{
Creates an object of binomial distribution
}
\examples{
# Create a binomial distribution object
dis1 <- bin_distribution(trials = 5, prob = 0.5)
# Create a bar plot of the binomial distribution
plot(dis1)

}
