% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{accepts}{trials, probability}
}
\value{
returns array of probabilities for each number of successes
}
\description{
probability of getting any number of successes in n trials
}
