% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{- number of trials}

\item{prob}{- probability of success}
}
\value{
A data frame describing the discrete probability for each possible binomial outcome
from 'trials' trials
}
\description{
Binomial distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)

}
