% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{probability of success in a single trial}
}
\value{
bindis: primary class 'bindis' and second class 'data.frame'
}
\description{
Displays probability density of binomial distribution
given # of trials and probability of a single success
}
\examples{
:
bin_distribution(trials = 5, prob = 0.5)
}
