% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{Number of trials (numeric)}

\item{prob}{Probability of successs (real)}
}
\value{
Returns a data.frame of the number of successes and their probabilities
}
\description{
Computes the binomial distribution of a certain number of success over a given number of trials
}
\examples{
x <- bin_distribution(trials = 5, prob = 0.5)
x
success probability
1 0 0.03125
2 1 0.15625
3 2 0.31250
4 3 0.31250
5 4 0.15625
6 5 0.03125

plot(x)
Returns a barplot of the distribution

}
