% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R, R/hello.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)

bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability that a trial will be a success}

\item{trials}{number of trials}

\item{prob}{probability that a trial will be a success}
}
\value{
probability of getting number of successes from trials for 0 through trials successes

probability of getting number of successes from trials for 0 through trials successes
}
\description{
Calculates the probabilities  of getting 0 through trials success s

Calculates the probabilities  of getting 0 through trials success s
}
\examples{
bin_probability(5, .5)
bin_probability(5, .5)
}
