% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-functions-methods.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
probability of success for all successes given number of trials
}
\description{
Returns a table of the binomial distribution over all possible successes for a certain number of trials
}
\examples{
# binomial probability distribution
bin_distribution(trials = 5, prob = 0.5)
}
