% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Bin Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
data frame of number of successes and probabilities
}
\description{
computes a table of number of successes and their respective probabilities
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
