% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials (n)}

\item{prob}{probability (p)}
}
\value{
dataframe of distribution
}
\description{
calculate binomial distribution
}
\examples{
bin_distribution(5, 0.5)
}
