% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution()}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a dataframe with the probability distribution
}
\description{
calculate the binomial distribution
}
