% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bino.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
An object of class \code{"bindis"} or \code{"data.frame"}
}
\description{
Gives a table of binomial distribution with probability
}
\examples{
# 20 trials with probability of success 0.5
distribution <- bin_distribution(20, 0.5)

}
