% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_function.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution Function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of getting a success in a trial (numeric)}
}
\value{
computed binomial distribution data frame with classes ("bindis" and "data.frame")
}
\description{
Creates a binomial distribution data frame using the binomial probability function with successes and probabilities as columns
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
