% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
data.frame with probability distribution that has successes in the first column and probability in the second column
}
\description{
calculates probability distribution and gets an output that has two classes consisting of bindis and data.frame
}
\examples{
bin_distribution(5, 0.5)
bin_distribution(4, 0.6)
}
