% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a data frame containing all possible probabilties under specific number of trials
}
\description{
creates a dataframe for the possible probabilities under specific number of trials
}
\examples{
bin_distribution(5, 0.5)
}
