% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials, must be non negative int}

\item{prob}{probability of success, must be numeric between 0 and 1}
}
\value{
dataframe with successes in the first column and associated probabality
in the second column
}
\description{
creates dataframe with probability distribution for binomial with given
number of trials and prob of success
}
\examples{
bin_distribution(19, .25)
bin_distribution(trials = 100, prob = .5)
}
