% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{return the distribution of success}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trails}

\item{prob}{the probability}
}
\value{
a dataframe of distribution
}
\description{
return the distribution of success
}
\examples{
bin_distribution(5, 0.5)
}
