% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{describe the number of trials}

\item{prob}{describe probability}
}
\value{
probability for each success
}
\description{
returns the probability for each possible success in the trials
}
\examples{
bin_distribution(3,0.3)
}
