% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials to attempt}

\item{prob}{probability of success}
}
\value{
a data.frame with all probabilities of any success over n trials of type bindist
}
\description{
gives the binomial distribution of all number of successes over a certain number of trials
}
\examples{
bin_distribution(5,0.5)
}
