% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_prob,distr, cumulative, choose.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin distribution.}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials.}

\item{prob}{the probability of success in any given trial.}
}
\value{
a data frame with the probability distribution.
}
\description{
displays table.
}
\examples{
bin_distribution(trials = 5, prob = 0.5)

bin_distribution(8, 0.1)
}
