% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Bin Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
an object of class \code{c("bindis", "data.frame")}
}
\description{
Creates an object of class \code{c("bindis", "data.frame")}
}
\examples{
# default
bindis <- bin_distribution(5, .5)

}
