% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin distribution function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials (intger)}

\item{prob}{number of success (intger)}
}
\value{
a dataframe with the probability distribution of each successes with certain probability and trials
}
\description{
calculate the binomial distribution
}
