% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Max Sima .R
\name{bin_distribution}
\alias{bin_distribution}
\title{Function bin_distribution()}
\usage{
bin_distribution(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
data frame with the probability distribution
}
\description{
create a dataframe with 2 classes of the binomial distribution
}
\examples{
bin_distribution(5, 0.5)
}
