% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials; numeric vector}

\item{prob}{probability of success; numeric vector}
}
\value{
return a data frame with the probability distribution (sucesses in the first column, probability in the second column)
}
\description{
caculates the probability distribution of success out of a given number trials with a given probability
}
\examples{
bin_distribution(trials = 5, prob = 0.5)

X <- bin_distribution(5, 0.5)

plot(X)
}
