% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of (fixed) trials}

\item{prob}{the probability of success on each trial}
}
\value{
a data frame with the probability distribution
}
\description{
return a data frame with the probability distribution: successes in the first column, probability in the second column
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
