% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{a dataframe of binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{integer prob double}
}
\value{
dataframe of binomial distribution
}
\description{
base on the binomial probability, create a dataframe of binomial distribution
}
\examples{
bin_distribution(trials=5,prob=0.5)
}
