% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{value of trials}

\item{prob}{probability of one independent success}
}
\description{
obtain a data frame with the probability distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
success probability
}
