% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{Number of trials (numeric)}

\item{prob}{Probability of success (numeric)}
}
\value{
A dataframe of binomial distribution
}
\description{
Calculates a binomial distribution and form a table
}
\examples{
plot.bindis(bin_distribution(trials = 5, prob = 0.5))
}
