% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability (numeric)}
}
\value{
computed binomial distribution
}
\description{
calculates the binomial distribution
}
\examples{
bin_distribution(trials=5,prob=0.5)
}
