% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{number of success}
}
\value{
the probability distribution
}
\description{
Calculates the probability distribution
}
\examples{
 \dontrun{
bin_distribution(trials=5L,prob=0.5)
 }
}
