% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{trial numbers}

\item{prob}{success probability}
}
\value{
a data frame containing distribution of a n trials experiment
}
\description{
compute distribution of a n trials experiment
}
\examples{
bd <- bin_distribution(3, 0.5)
}
