% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability of success in each trial}
}
\value{
an object of both 'bindis' and 'data.frame' that displays the success and its corresponding probability
}
\description{
constructs a table of successes and probability.
}
