% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout03-Zihan-Liu.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials (integer)}

\item{prob}{probability of success (numeric)}
}
\value{
a data.frame with two classes: c("bindis", "data.frame")
}
\description{
output a data.frame with two classes: c("bindis", "data.frame")
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
