% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions and methods.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{success rate of each trials}
}
\value{
A data frame with number of success and its probability
}
\description{
calculates the probability of different number of success trials
}
