% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution Function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the numeric vector of total trials}

\item{prob}{the numeric vector of probability for success of one single trial}
}
\value{
the returned output: a data.frame with two classes: c("bindis", "data.frame")
}
\description{
output a data frame with the probability distribution: sucesses in the first column, probability in the second column
}
\examples{
bin_distribution(10, 0.5)
}
