% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability if success}
}
\value{
an object of class \code{"bindis"}

an object of class \code{"data.frame"}
}
\description{
compute the probability distribution
}
\examples{
# binomial probability distribution of different success in five trials
bin_distribution(trials = 5, prob = 0.5)
}
