% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3_Zimeng_Zheng.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability that success occurs}
}
\value{
a data frame with the probability distribution
}
\description{
function to calculate the distribution of binomial random variable
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
