% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
Data frame with the Probability Value
}
\description{
Computes binomial distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
dis1 <- bin_distribution(5, 0.5)
plot(dis1)
}
