% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All Functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability value (numeric)}
}
\value{
data frame with the probability distribution
}
\description{
create a data frame with the probability distribution
}
\examples{
bin_distribution(5, 0.5)
}
