% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number indicating the number of trials}

\item{prob}{number indicating teh probability of getting success}
}
\value{
the data frame with the probability distribution
}
\description{
get a data frame that the first column is the number of success, and the second column is the probability
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
