% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-adele-shen.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
kurtosis (numeric)
}
\description{
Calculates the kurtosis of a binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)

}
