% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func_measures.R, R/workout03-ann-deng.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis()}
\usage{
bin_kurtosis(trials, prob)

bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
double of kurtosis

double of kurtosis
}
\description{
returns kurtosis of binvar with trials and prob

returns kurtosis of binvar with trials and prob
}
\examples{
bin_kurtosis(10, 0.3) ## returns -0.1238095
bin_kurtosis(10, 0.3) ## returns -0.1238095
}
