% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3-divya-vijay.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of probability}
}
\value{
kurtosis summary measure
}
\description{
calculates kurtosis for binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
