% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
kurtosis value
}
\description{
Kurtosis of a Binommial Distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
