% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of random variable}
}
\value{
returns kurtosis given by the number of trials and its probability
}
\description{
gives the kurtosis of a binomial random variable
}
\examples{
bin_kurtosis(10,0.3)
}
