% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Bin kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{is a numeric integer that represents the number of trials}

\item{prob}{is a numeric value, between 0 and 1, that represents the probability or success}
}
\value{
this function returns a numeric value that represents the kurtosis of the probability distribution
}
\description{
This function gives the kurtosis of a binomial distribution
}
\examples{
bin_kurtosis(trials = 5, prob = 0.5)
bin_kurtosis(trials = 100, prob = 0.2)
bin_kurtosis(trials = 45, prob = 0.75)
}
