% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Kurtosis function}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
Kurtosis
}
\description{
Calculates the kurtosis
}
\examples{

#computes kurtosis with corresponding arguments
 bin_kurtosis(10, 0.3)

}
