% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Calculate kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{integer the number of trials}

\item{prob}{the probability to success in each trials}
}
\value{
The kurtosis of a binomial distribution
}
\description{
Computes the kurtosis of a binomial distribution
}
\examples{
# Calculate the kurtosis for 10 trials, each trial with
# independent probability 0.5
bin_kurtosis(trials = 10, prob = 0.5)
}
