% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{(numeric)}

\item{prob}{(numeric)}
}
\value{
the kurtosis value
}
\description{
the kurtosis of the binomial distribution
}
