% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a numeric value
}
\description{
Calculates the kurtosis
}
\examples{
kurtosis of binomial random distribution with 10 trials and prob = 0.5
# bin_kurtosis(10, 0.5)
}
