% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
kurtosis value
}
\description{
calculates kurtosis
}
\examples{
bin_kurtosis(10, 0.3)
}
