% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{: Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{:}{trials, prob}
}
\value{
: kurtosis
}
\description{
: Calculates kurtosis given a binomial distribution
}
\examples{
: bin_kurtosis(10, 0.3) = -0.1238095
}
