% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{as number of trials}

\item{prob}{as probability number between 0 and 1}
}
\value{
kurtosis as numeric value
}
\description{
calculate measure of "tailedness" of probability dist. of a random variable
}
\examples{
bin_kurtosis(10, 0.3)
}
