% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunc.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{Probability that a success occurs}
}
\value{
Kurtosis of binomial variable
}
\description{
calculate the kurtosis of binomial variable
}
\examples{
bin_kurtosis(10, 0.3)
}
