% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis Function}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{success probability}
}
\value{
kurtosis, as a numeric value. 2 modes if value is an exact integer.
}
\description{
Computes kurtosis for binomial variable with given trials and prob. Uses auxiliary function.
}
\examples{
bin_kurtosis(10, 0.3)
}
