% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
kurtosis value of binomial distribution described by parameters
}
\description{
calulates the kurtosis value of a binomial distribution
}
\examples{
bin_kurtosis(5, 0.5)
}
