% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-code.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{tailedness of binomial distribution}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
tailedness of the binomial distribution
}
\description{
calculates tailedness of a binomial distribution; kurtosis formula
}
