% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_kurtosis.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
kurtosis
}
\description{
calculates the kurtosis of a binomial distribution
}
