% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of total trials}

\item{prob}{probabilty of success in each trial}
}
\value{
kurtosis
}
\description{
returns kurtosis for binomial variable
}
\examples{
bin_kurtosis(10, 0,3)
}
