% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{kurtosis of the binomial distribution}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials, }{prob}
}
\value{
kurtosis
}
\description{
calculate the kurtosis
}
\examples{
bin_kurtosis(10, 0.3)
}
