% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_fun.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials = 1, prob = 1)
}
\arguments{
\item{trials}{trials}

\item{prob}{probability}
}
\value{
kurtosis
}
\description{
calculates the kurtosis
}
