% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_measures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{number that is the probability of success}
}
\value{
kurtosis the numerical mean
}
\description{
computes kurtosis given trials and probs
}
\examples{
#5 trials with success prob at 0.5
bin_kurtosis(trials=5, prob=0.5)
}
