% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of a success}
}
\value{
the calculated kurtosis
}
\description{
Calculates the kurtosis of a binomial random variable
}
\examples{
bin_kurtosis(7, 0.6)

bin_kurtosis(trials = 5, prob = 0.25)
}
