% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials, }{number of flips}

\item{prob, }{probability of heads}
}
\value{
kurtosis
}
\description{
Finds the kurtosis of a binomial random variable
}
