% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_func.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Bin Kurtosis}
\usage{
bin_kurtosis(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability value}
}
\value{
the kurtosis value
}
\description{
Calculates the measure of tailedness of the prob distribution of a random var
}
\examples{
# kurtosis of n=10, p=0.3
kurt <- bin_kurtosis(10, 0.3)
}
