% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{# of trials}

\item{prob}{probability of success}
}
\value{
the kurtosis
}
\description{
calculates the kurtosis of distribution
}
\examples{

# default
bin_kurtosis()

# another example
bin_kurtosis(trials = 5, prob = 0.25)

}
