% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-of-measures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{success:}{number of trials}

\item{trials:}{probability value}
}
\value{
kurtosis using trials and probability
}
\description{
calculates kurtosis
}
\examples{
bin_kurtosis(10, 0.3)
}
