% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
numeric vector with kurtosis value
}
\description{
Computes kurtosis measure for the given distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
