% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(prob, trials)
}
\arguments{
\item{prob}{probability of random variable}

\item{trials}{number of trials}
}
\value{
a numeric value
}
\description{
Finds the binomial kurtosis for a given probability and number of trials
}
\examples{
 \dontrun{
 bin_kurtosis(prob = 0.5, trials = 5)
 }
}
