% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_measures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials (integer)}

\item{prob}{probability of success in one trial (integer)}
}
\value{
Calculated kurtosis
}
\description{
Calculates the kurtosis of the distribution
}
\examples{
# distribution with 10 trials and probability 0.3
bin_kurtosis(trials = 10, prob = 0.3)
}
