% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the kurtosis of a binomal distribution
}
\description{
calculates the kurtosis of a binomal distribution
}
\examples{
What is the kurtosis of number of successes in 10 trials with prob of success = 0.3?
bin_kurtosis(10, 0.3)
}
