% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_funcs.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
the kurtosis of the binomial distribution
}
\description{
Calculates the kurtosis of a binomial distribution given the number of trials and the probability for success
}
\examples{
bin_kurtosis(10, 0.3)
}
