% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{kurtosis of a binomial distribution}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on rach trial}
}
\value{
kurtosis of a binomial distribution
}
\description{
calculate the kurtosis of a binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)

}
