% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{n}{number of trials, prob probablity of success in each trial}
}
\value{
the kurtosis of the given binomial distribution with n trials and probability prob of success for each trial
}
\description{
calculates the kurtosis of the number of successes given a number of trials and probability of success for each trial
}
