% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{n}{number of trials}
}
\value{
the kurtosis of the given binomial distribution with "trials "trials and probability "prob" of success
}
\description{
finds the kurtosis of a binomial distribution with "trials" and probability "prob"
}
