% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns_of_measures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
Kurtosis of a binomial variable
}
\description{
Calculates the kurtosis of a binomial variable
}
\examples{
bin_kurtosis(trials = 5, prob = 0.5)

}
