% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binominal kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{how many trials}

\item{prob}{probability of success on any given trial}
}
\value{
kurtosis of the specified binomal distribution
}
\description{
gives kurtosis of the binomial
}
\examples{
bin_kurtosis(6, 0.3)
}
