% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Kurtosis of binomial}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{(numeric) the number of trials}

\item{prob}{(numeric) the probability of success on each trial}
}
\value{
Kurtosis of the binomial distribution with the given trials and probability parameters.
}
\description{
Compute kurtosis of binomial distribution with the given trials and probability parameters.
}
