% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, probs)
}
\arguments{
\item{trials}{The number of trials}

\item{probs}{The probability of success}
}
\value{
kurtosis value
}
\description{
This calculates the kurtosis of the binomial experiment
}
\examples{
bin_kurtosis(trials=5,probs=0.5)

}
