% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{prob}{probablity of success in each trial}

\item{n}{number of trials}
}
\value{
kurtosis of the binomial distribution
}
\description{
computes kurtosis of the binomial distribution
}
\examples{
bin_kurtosis(10, 0.2)
}
