% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binary Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{Trials}

\item{prob}{Probability}
}
\value{
A number for kurtosis
}
\description{
Returns kurtosis of binomial variable
}
