% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on each trial}
}
\value{
the kurtosis of a random binomial variable
}
\description{
calculate the kurtosis of a random binomial variable
}
\examples{
# the kurtosis of 5 trials with the probability of success in each trial = 0.5
bin_kurtosis(5, 0.5)
}
