% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-measure.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{kurtosis of a binomial distribution}
\usage{
bin_kurtosis(n, p)
}
\arguments{
\item{n}{the number of total trials}

\item{p}{the probability of single success}
}
\value{
the calculated kurtosis
}
\description{
calculate the kurtosis of a random variable following the binomial distribution
}
\examples{
bin_kurtosis(5,0.5)
}
