% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{is the number of trials}

\item{prob}{is the probability of success}
}
\value{
the kurtosis of a binomial distribution
}
\description{
calculates the kurtosis of a binomial distribution
}
\examples{
bin_kurtosis(trials = 10, prob = 0.3)
}
