% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a number, the calculated kurtosis of a binomial random variable
}
\description{
Calculates the kurtosis ("expected number of successes in n trials"tailedness" of the probability distribution)
}
\examples{
bin_kurtosis(trials = 10, prob = 0.5)
}
