% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workout3.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{an integer with number of trials}

\item{prob}{the probability of the success occur}
}
\value{
the kurtosis value
}
\description{
Calculate the kurtosis of binomial distribution
}
\examples{
bin_kurtosis(10, 0.5)
}
