% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunc.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success within one trial}
}
\value{
the numerical value of the kurtosis of the given data
}
\description{
calculates the kurtosis of the given binomial distribution
}
\examples{

# default
bin_kurtosis()

# another example
bin_kurtosis(trials = 3, prob = 0.5)

}
