% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{numeric vector}

\item{prob}{numeric vector}
}
\value{
the kurtosis of the binomial distribution
}
\description{
calculates the kurtosis of the binomial distribution
}
\examples{

bin_kurtosis(10, 0.3)

}
