% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials (numeric)}

\item{prob}{the probability of one success (numeric)}
}
\value{
the kurtosis of the binomial distribution (numeric)
}
\description{
finds the kurtosis of the binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
