% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
computed binomial kurtosis
}
\description{
computes the kurtosis of the binomial variable
}
\examples{
bin_kurtosis(trials = 10, prob = 0.5)
}
