% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success per trial}
}
\value{
kurtosis
}
\description{
calculates the krtosis of the binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
