% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{: bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{:}{trials the number of trials}

\item{:}{prob the probability of success}
}
\value{
: the kurtosis of the binomial(trials, prob)
}
\description{
: calculates the kurtosis of the binomial with trials number of trials and probability of succeess prob
}
\examples{
: bin_kurtosis(trials = 10, prob = 0.3)
}
