% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{integer representing n, the total number of trials}

\item{prob}{float representing p, the probability of success on a given trial}
}
\value{
kurtosis of the distribution
}
\description{
Calculates the kurtosis of a binomial distribution with given trials and probability
}
\examples{
bin_kurtosis(10,.3)

bin_kurtosis(100,2/3)
}
