% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{numeric value for the number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
computed binomial kurtosis
}
\description{
computes the kurtosis of the binomial experiment
}
\examples{
bin_kurtosis(trials = 10, prob = 0.3)
}
