% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{trials}

\item{prob}{probability}
}
\value{
binkurt
}
\description{
Calculates kurtosis of a binomial distribution of a given probability in n trials
}
\examples{
bin_mean(5, 0.5)
}
