% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_kurtosis.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\value{
kurtosis
}
\description{
after checking if inputs are valid, use aux_kurtosis to return kurtosis
}
\examples{
bin_kurtosis(trials=5,prob=0.5)
}
