% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
kurtosis of a binomial distribution
}
\description{
calculates the kurtosis of a binomial distribution
}
\examples{
bin_kurt_5 <- bin_kurtosis(5,0.5)

bin_kurt_10 <- bin_kurtosis(10,0.1)

}
