% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
With valid input of parameters, return the kurtosis
}
\description{
compute the kurtosis
}
\examples{
bin_kurtosis(trials = 5, prob = 0.5)
}
