% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_summary_functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{(integer)}

\item{prob}{(real)}
}
\value{
kurtosis(tailedness) of binomial distribution
}
\description{
computes kurtosis(tailedness) of a binomial distribution
}
\examples{
bin_kurtosis(trials = 5, prob = 0.5)
bin_kurtosis(10, 0.3)
}
