% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Kurtosis of a Binomial Distribution}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{total numbeer of trials (numeric)}

\item{prob}{the probability of a success in a trial (numeric)}
}
\value{
a number indicates the kurtosis of the binomial distribution
}
\description{
Calculating the kurtosis of a binomial distribution based on inputs, trials and probability
}
\examples{
bin_kurtosis(5,0.5)
-0.4
bin_kurtosis(6,0.3)
-0.2063492
}
