% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Kurtosis of a Binomial Random Variable}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability for success in each trial}
}
\value{
the kurtosis of the binomial random variable
}
\description{
calculates the kurtosis of a binomial random variable given number of trials and prob
}
