% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_kurtosis.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{kurtosis of binomial distribution}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of a single success}
}
\value{
calculated kurtosis of binomial distribution
}
\description{
calculate the kurtosis of binomial distribution
}
\examples{

bin_kurtosis(10, 0.3)

}
