% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_summ_func.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{some integer}

\item{prob}{some real number}
}
\value{
kurtosis of binomial distribution
}
\description{
measures tailedness of probability distribution
}
\examples{
bin_kurtosis(trials=5, prob=0.5)
bin_kurtosis(10, 0.3)
}
