% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials, }{the number of trials}

\item{prob, }{the probability of a success}
}
\value{
returns a numeric kurtosis
}
\description{
gives the kurtosis of the binomial distribution
}
\examples{
bin_kurtosis(trials = 5L, prob = 0.5)
bin_kurtosis(3L, .3)
}
