% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis Function}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
computed value
}
\description{
Calculate binomial kurtosis
}
\examples{
# 10 trials, probability of 0.3
bin_kurtosis(10, 0.3)
-0.1238095
}
