% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis method}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the kurtosis of a binomial variable
}
\description{
calculates the kurtosis of a binomial variable
}
\examples{

# kurtosis of the number of heads got from tossing a fair coin 10 times
bin_mean(10, 0.5)

}
