% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-methods.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Kurtosis of Binomial Distribution}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
kurtosis value of a binomial dsitribution
}
\description{
measure the “tailedness” of the probability distribution of a random variable.
}
\examples{
bin_kurtosis(10, 0.3)
}
