% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{probability of success on each trial}
}
\value{
kurtosis of a binomial distribution
}
\description{
calculates the kurtosis of a binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
