% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_kurtosis.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
binomial kurtosis
}
\description{
binomial kurtosis
}
\examples{
bin_kurtosis(2,.5)
}
