% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial_main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{probability of success}
}
\value{
compute the kurtosis of a binomal distribution
}
\description{
calculate the kurtosis of a binomal distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
