% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_function.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability value}
}
\value{
computed kurtosis
}
\description{
calculates the kurtosis
}
\examples{
bin_kurtosis(10, 0.3)
}
