% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials, a non-negative integer}

\item{prob}{the probability of success, a number between 0 and 1 both included}
}
\value{
kurtosis of a probability distribution of the random variable
}
\description{
computes the kurtosis of a probability distribution of a random variable, which measures the tailedness of the probability distribution of random variable.
}
\examples{
# the kurtosis of a random variable in 5 trials and probability of success 0.4
bin_kurtosis(5,0.4)

# the mode of a random variable in 10 trials and probability of success 0.3
bin_kurtosis(10,0.3)
}
