% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mearsures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{an integer}

\item{prob}{a probability value}
}
\value{
kurtosis
}
\description{
a measure of the tailedness of the probability distribution of a RV
}
\examples{
bin_kurtosis(10, 0.3)
}
