% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis Function}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{total number of trials (integer)}

\item{prob}{probability of making success in one trial (decimal)}
}
\value{
the value of kurtosis of the specified binomial distribution
}
\description{
compute the value of kurtosis of a binomial distribution with specified total number of
trials and success probability
}
\examples{
bin_kurtosis(10, 0.3)
}
