% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Bin_Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{an integer or vector of numerics - number of trials}

\item{prob}{a numeric value between 0 and 1}
}
\value{
the kurtosis of the variable passed in
}
\description{
finds kurtosis of binomial given trials and prob
}
\examples{
# default
kurt <- bin_kurtosis(trials = 10, prob = 0.3)

}
