% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
computed kurtosis
}
\description{
to compute the kurtosis of a binomial distribution
}
\examples{
bin_kurtosis(trials = 10, prob = 0.3)
}
