% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{kurtosis of a binomial}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
calculated kurtosis
}
\description{
calculates the kurtosis of a binomial distribution
}
