% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{compute the binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{is the probability of success}
}
\value{
returns the binomial kurtosis of an object
#' @examples

#bin_kurtosis
bin_kurtosis(10,0.2)
}
\description{
generates binomial kurtosis
}
