% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_summary_measures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{representing # of trials (numeric value)}

\item{prob}{representing probability (numeric value)}
}
\value{
kurtosis (numeric value)
}
\description{
determines kurtosis of given probability and given number of trials
}
\examples{
bin_kurtosis(5,0.5)
bin_kurtosis(1,1)
}
