% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success for each trials}
}
\value{
tailedness
}
\description{
the function measures the "tailedness" of the probability distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
