% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials:}{number of total trials}

\item{prob:}{probability of success}
}
\value{
numeric value representing kurtosis given number of trials and prob of success
}
\description{
calculates kurtosis of binomial scenario of n trials and p prob of success using aux_kurtosis
}
