% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials:}{number of trials, a non-negative integer}

\item{prob:}{probability of success, a number between 0 and 1, inclusive}
}
\value{
kurtosis of binomial
}
\description{
calculates the kurtosis of the binomial
}
\examples{
> bin_kurtosis(10, 0.3)
[1] -0.1238095
}
