% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success in one trial}
}
\description{
The Kurtosis (from greek kurtos, meaning ¡°curved, arching¡±) is a measure of the ¡°tailedness¡± of the probability distribution of a random variable.
}
