% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
kurtosis of the binomial distribution
}
\description{
main function for summary measure "kurtosis"
}
