% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Function bin_kurtosis()}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability of success}
}
\value{
the kurtosis of the of binimial distribution
}
\description{
calculate the kurtosis of the of binimial distribution
}
\examples{
bin_kurtosis(3,0.5)
}
