% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunction.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomal kurtosis}
\usage{
bin_kurtosis(n, p)
}
\arguments{
\item{n}{the number of trials}

\item{p}{the probability of successes}
}
\value{
computer the kurtosis of a binomal distribution
}
\description{
calculate the kurtosis of a binomal distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
