% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{int}{integer trials}

\item{int}{integer prob}
}
\value{
binomial kurtosis
}
\description{
checks probability and trials, and then calculates the kurtosis
}
\examples{
bin_kurtosis(10, 0.3)
}
