% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial_Functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{: bin_kurtosis()}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials:}{the number of trials}

\item{prob:}{the probability of success}
}
\value{
: calculated kurtosis of binomial distribution
}
\description{
: kurtosis of binomial distribution
}
