% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Bin Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
number that is the kurtosis of the bin
}
\description{
main function for kurtosis of bin
}
\examples{
bin_kurtosis(10, 0.3)
}
