% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{numb of trials}

\item{prob}{prob of success}
}
\value{
the binomial kurtosis
}
\description{
calculates the kurtosis
}
\examples{
# kurtosis of 10 trials with p=.5
bin_kurtosis(10, .5)

# kurtosis of 8 trials with p=.7
bin_kurtosis(8, .7)

}
