% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis function}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of n attempts (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
the kurtosis (integer)
}
\description{
computes the kurtosis of the distribution
}
\examples{
bin_kurtosis(trials = 10, prob = 0.3)
}
