% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials, }{total number of trial}

\item{prob, }{chance of success}
}
\value{
kurtosis if input are good else error
}
\description{
checks input then calcs kurtosis
}
\examples{
bin_kurtosis(5,0.5)
}
