% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{numeric value, number of trials}

\item{prob}{numeric value between 0 and 1, probability of success}
}
\value{
a numeric value, kurtosis
}
\description{
computes the kurtosis of the binomial experiment
}
\examples{
bin_kurtosis(trials=5,prob=0.5)
}
