% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_statmanips.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Kurtosis of Binomial}
\usage{
bin_kurtosis(trials = 1, prob = 1)
}
\arguments{
\item{trials}{the number of trials in the distribution}

\item{prob}{the probability of a success}
}
\value{
The kurtosis of the binomial distribution, using aux_kurtosis
}
\description{
Finds the kurtosis of the binomial distribution.
}
