% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
kurtosis of binomial ditribution
}
\description{
calculates kurtosis of binomial ditribution
}
\examples{

bin_kurtosis(10, 0.3)
}
