% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_and_methods.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials, }{number of trials}

\item{prob, }{probablity of successes}
}
\value{
value of the kurtosis of binomial distribution
}
\description{
calculates the kurtosis of binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
