% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{integer value for number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
kurtosis of binomial
#' @examples

bin_kurtosis(5, 0.5)
bin_kurtosis(10, 0.2)
}
\description{
finds the kurtosis of the binomial
}
