% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials; non-negative integer value}

\item{prob}{probability of success; must be a numeric value between 0 and 1 (inclusive)}
}
\value{
numeric value indicating kurotis of distribution
}
\description{
Calculates the "tailedness" of the probability distribution; descriptor of the shape of the distribution
}
\examples{
kurtosis <- bin_kurtosis(10, 0.3)
kurtosis
}
