% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{The number of trials}

\item{prob}{The number of successes}
}
\value{
A number
}
\description{
Calculates the 'tailedness' of a probability distribution of a random variable
}
