% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
Calculates the kurtosis of a binomial variable
}
\description{
Calculates the kurtosis of a binomial variable
}
\examples{
# Calculate the kurtosis of a binomial distribution
bin_kurtosis(trials = 5, prob = 0.5)

}
