% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{accepts}{trials, probabilities}
}
\value{
binomial kurtosis of trials and Probabilities
}
\description{
calculates binomial kurtosis
}
