% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{prob}{}
}
\value{
The kurtosis of the distribution
}
\description{
Binomial kurtosis
}
\examples{
kurt <- bin_kurtosis(5, 0.5)
kurts <- bin_kurtosis(1:5, 0.5)

}
