% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success on each trial (numeric)}
}
\value{
kurtosis of binomial distribution
}
\description{
computes kurtosis of binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
