% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{probability of success in a single trial}
}
\value{
numeric: the kurtosis value
}
\description{
Find kurtosis given # of trials and probability of a single success
}
\examples{
:
bin_kurtosis(trials = 5, prob = 0.5)
}
