% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_main.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{Number of trials (real)}

\item{prob}{Probability of successs (real)}
}
\value{
Returns the kurtosis
}
\description{
Computes the kurtosis of a binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
-0.1238095
}
