% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability that a trial will be a success}
}
\value{
Kurtosis
}
\description{
Returns kurtosis of binomial dist
}
\examples{
bin_cumulative(5, .5)
}
