% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the inputted number of trials for a binomial distribution (numeric)}

\item{prob}{the inputted probability for a binomial distribution (numeric)}
}
\value{
returns the kurtosis of a binomial distribution with given probability and number of trials
}
\description{
Chekcks whether inputted trials and probability are valid and uses an auxiliary function to calculate the kurtosis
}
\examples{
bin_kurtosis(10, 0.3)
}
