% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis()}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the kurtosis of the binomial distribution
}
\description{
calculates the kurtosis of a binomial distribution.
}
