% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bino.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a numeric number represents the kurtosis
}
\description{
Computes kurtosis of the binomal distribution
}
\examples{
# 20 trials with probability of success 0.5
kurtosis <- bin_kurtosis(20, 0.5)

}
