% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_function.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis Function}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of getting a success in a trial (numeric)}
}
\value{
computed kurtosis of binomial distribution based on parameter values (numeric)
}
\description{
returns the kurtosis of binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
