% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of random trials}

\item{prob}{probability of successes}
}
\value{
kurtosis of a binomial distribution
}
\description{
generate the kurtosis of a binomial distrbution, a measure of the
"tailedness" of the probability distribution of a random variable
}
\examples{
bin_kurtosis(10, 0.3)

}
