% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
kurtosis if probability and trials inputs are valid. If not, throws an error.
}
\description{
calculates kurtosis of having certain trials and certain probability
}
\examples{
bin_kurtosis(5, 0.6)
bin_kurtosis(6, 0.7)
}
