% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
Kurtosis
}
\description{
Returns kurtosis of binom dist with given trials and prob
}
\examples{
bin_kurtosis(trials = 10, prob = .4)
bin_kurtosis(110, prob = .56)
}
