% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{return the kurtosis of the binomial distribution}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trails}

\item{prob}{the probability}
}
\value{
kurtosis of the binomial distribution
}
\description{
return the kurtosis of the binomial distribution
}
\examples{
bin_kurtosis(5, 0.5)
}
