% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials to attempt}

\item{prob}{probability of success}
}
\value{
returns a the kurtosis of the binomial distribution
}
\description{
Computes the kurtosis of the binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
