% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{Number of trials in the binomial distribution (numeric)}

\item{prob}{Probability of success in the binomial distribution (numeric)}
}
\value{
Kurtosis of binomial distribution
}
\description{
Computes the kurtosis of the binomial distribution
}
\examples{
bin_kurtosis(trials = 5, prob = 0.5)
}
