% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_summary.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Kurtosis.}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials.}

\item{prob}{the probability of success in any given trial}
}
\value{
the computed summary measure.
}
\description{
a measure of the “tailedness” of the probability distribution of a random variable.
}
\examples{
bin_kurtosis(trial = 5, prob = 0.5)
bin_kurtosis(10, 0.3)
bin_kurtosis(100, 0.7)
}
