% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binonmial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
binomial kurtosis
}
\description{
Creates binomial kurtosis given "trials" and "prob"
}
\examples{
# default
kurt <- bin_kurtosis(5, .5)
}
