% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis function}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{Number of (fixed) trials in the binomial distribution.}

\item{prob}{the probability of success on each trial.}
}
\value{
the kurtosis of binomial distribution
}
\description{
Compute the kurtosis of the binomial distribution
}
\examples{
bin_kurtosis(10,0.3)
}
