% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of (fixed) trials}

\item{prob}{the probability of success on each trial}
}
\value{
the kurtosis of the binomial distribution
}
\description{
calculates the kurtosis of the binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
