% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{kurtosis of binomial}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{integer prob double}
}
\value{
kurtosis double
}
\description{
return kurtosis of binomial distribution
}
\examples{
bin_kurtosis(trials=5,prob=0.5)
}
