% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis function}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trial (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
the kurtosis
}
\description{
Culculate the kurtosis of a binomial distribution
}
