% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{the kurtosis of the binomial distribution}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability (numeric)}
}
\value{
computed kurtosis value
}
\description{
calculate the kurtosis of the binomial distribution
}
\examples{
bin_kurtosis(trials=5,prob=0.5)
}
