% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success each time}
}
\value{
the number of kurtosis
}
\description{
the kurtosis of a binominal distribution
}
\examples{
bin_kurtosis(trials = 10, prob = 0.3)
}
