% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_kurtosis.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
the kurtosis of binomial
}
\description{
Calculates the kurtosis of binomial
}
\examples{
 \dontrun{
bin_kurtosis(trials=10L,prob=0.3)
 }
}
