% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis function}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the kurtosis of a binomial distribution
}
\description{
compute the kurtosis of a binomial distribution
}
\examples{
bin_kurtosis(trials = 10, prob = 0.3)
}
