% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions and methods.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis function}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{Number of trials in the binomial distribution}

\item{prob}{Probability of success in the binomial distribution}
}
\value{
the kurtosis of binomial distribution
}
\description{
: Compute the kurtosis of the binomial distribution
}
