% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Skewness}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{numeric number of trials}

\item{prob}{numeric number of probability}
}
\description{
meansure of the tailedness of the distribution of
}
\examples{
bin_kurtosis(10,0.3)
}
