% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3_Zimeng_Zheng.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability that success occurs}
}
\value{
the kurtosis of the binomial distribution
}
\description{
function to calculate the kurtosis of the binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
