% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Bin Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability}
}
\value{
kurtosis
}
\description{
calculates kurtosis of a binomial distribution
}
