% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_kurtosis.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{numeric}

\item{prob}{numeric}
}
\value{
the computed data
}
\description{
return mode given trials and prob
}
\examples{
bin_kurtosis(10,0.3)
}
