% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_of_measures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the kurtosis of a binomial variable
}
\description{
to calculates the kurtosis of a binomial variable
}
\examples{
bin_kurtosis(trials = 10, prob = 0.3)

}
