% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
The kurtosis
}
\description{
Compute Binomial Kurtosis Value
}
\examples{
bin_kurtosis(trials = 10, prob = 0.3)
}
