% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All Functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probabiltiy value (numeric)}
}
\value{
kurtosis
}
\description{
determine the kurtosis
}
\examples{
bin_kurtosis(10, 0.3)
}
