% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-adele-shen.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
mean (numeric)
}
\description{
Calculates the mean or expected value of a binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
