% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_of_measures.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of head probability}
}
\value{
computed mean value
}
\description{
calculates the mean of binomial random variable
}
\examples{

bin_mean(10 ,0.3)

}
