% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{Numeric, # of trials in the given binomial distribution}

\item{prob}{Numeric, probability of success in the given binomial distribution}
}
\value{
Expected value of binomial distribution
}
\description{
Computes E(X) of given binomial distribution
}
\examples{
bin_mean(trials = 10, prob = 0.5)
}
