% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3-divya-vijay.R
\name{bin_mean}
\alias{bin_mean}
\title{bin mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of probability}
}
\value{
mean summary measure
}
\description{
calculates mean for binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
