% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Bin Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{an integer, the number of trials}

\item{prob}{an integer of the probability of success for one trial}
}
\value{
an integer, the mean of successes in the given amount of trials
}
\description{
A function to return the mean of a binomial distribution with given trials and probability
}
\examples{
bin_mean(10,0.5)
}
