% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of random variable}
}
\value{
returns mean given by the number of trials and its probability
}
\description{
gives the mean of a binomial random variable
}
\examples{
bin_mean(10,0.3)
}
