% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_mean}
\alias{bin_mean}
\title{Bin mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{is a numeric integer that represents the number of trials}

\item{prob}{is a numeric value between 0 and 1 that represents the probability of success}
}
\value{
this function returns a numeric value that represents the expected value, or mean of the binomial distribution
}
\description{
This function gives the expected value, or mean of a binomial distribution
}
\examples{
bin_mean(trials = 5, prob = 0.5)
bin_meane(trials = 100, prob = 0.2)
bin_mean(trials = 45, prob = 0.75)
}
