% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mean}
\alias{bin_mean}
\title{Mean function}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
Mean
}
\description{
Calculates the mean
}
\examples{

#computes mean with corresponding arguments
 bin_mean(10, 0.3)

}
