% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_mean}
\alias{bin_mean}
\title{Calculate mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{integer the number of trials}

\item{prob}{the probability to success in each trials}
}
\value{
The expected mean of the binomial distribution
}
\description{
Computes the expected mean of a binomial distribution
}
\examples{
# Calculate the expected mean for 10 trials, each trial with
# independent probability 0.5
bin_mean(trials = 10, prob = 0.5)
}
