% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{(numeric)}

\item{prob}{(numeric)}
}
\value{
the expected value
}
\description{
the expect number of a binomial distribution
}
