% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mean}
\alias{bin_mean}
\title{Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a numeric value
}
\description{
Calculates the mean
}
\examples{
# mean of binomial random distribution with 10 trials and prob = 0.5
# bin_mean(10, 0.5)
}
