% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_mean}
\alias{bin_mean}
\title{: Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{:}{trials, prob}
}
\value{
: mean
}
\description{
: Calculates mean given a binomial distribution
}
\examples{
: bin_mean(10, 0.3) = 3
}
