% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{as number of trials}

\item{prob}{as probability number between 0 and 1}
}
\value{
mean as numeric value
}
\description{
calculate the mean of binomial distribution(np)
}
\examples{
bin_mean(10, 0.3)
}
