% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean Function}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{success probability}
}
\value{
mean, as a numeric value.
}
\description{
Computes binomial mean by calling auxiliary mean function.
}
\examples{
bin_mean(10, 0.3)
}
