% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_measure_functions.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the mean of a binomial variable
}
\description{
calculates the mean of a binomial variable
}
\examples{
bin_mean(trials = 10, prob = 0.3)
}
