% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
expected value of binomial distribution described by parameters
}
\description{
calulates the expected value of a binomial distribution
}
\examples{
bin_mean(5, 0.5)
}
