% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-code.R
\name{bin_mean}
\alias{bin_mean}
\title{mean of binomial distribution}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
mean of the binomial distribution
}
\description{
calculates expected value or mean of a binomial distribution
}
