% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_mean.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
mean
}
\description{
calculates the mean of a binomial distribution
}
