% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of total trials}

\item{prob}{probabilty of success in each trial}
}
\value{
expected value
}
\description{
returns expected value for binomial variable
}
\examples{
bin_mean(10, 0,3)
}
