% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{bin_mean}
\alias{bin_mean}
\title{mean of the binomial distribution}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials, }{prob}
}
\value{
mean
}
\description{
calculate the mean
}
\examples{
bin_mean(10, 0.3)
}
