% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials, }{number of flips}

\item{prob, }{probability of heads}
}
\value{
mean
}
\description{
Finds the mean of a binomial random variable
}
