% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of achieving success}
}
\value{
mean of binomial distribution
}
\description{
calculate the mean of binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
