% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{# of trials}

\item{prob}{probability of success}
}
\value{
the mean
}
\description{
calculates the mean of distribution
}
\examples{

# default
bin_mean()

# another example
bin_mean(trials = 5, prob = 0.5)

}
