% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-of-measures.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials:}{number of trials}

\item{prob:}{probability value}
}
\value{
mean using trials and probability
}
\description{
calculates mean
}
\examples{
bin_mean(10, 0.3)
}
