% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mean}
\alias{bin_mean}
\title{Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
numeric vector with mean value
}
\description{
Computes expected value (mean) of the binomial distribution
given number of trials and probability of success
}
\examples{
bin_mean(10, 0.3)
}
