% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_methods.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean Function}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success of individual trial}
}
\value{
computed mean of a binomial distribution
}
\description{
computes the expected mean of a binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
