% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(prob, trials)
}
\arguments{
\item{prob}{probability of random variable}

\item{trials}{number of trials}
}
\value{
a numeric value
}
\description{
Finds the binomial mean for a given probability and number of trials
}
\examples{
 \dontrun{
 bin_mean(prob = 0.5, trials = 5)
 }
}
