% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{bin_mean}
\alias{bin_mean}
\title{Mean of binomial random variable}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of I.I.D. Bernoulli trials}

\item{prob}{the shape parameter of each I.I.D. Bernoulli trial}
}
\value{
the mean of a binomial random variable defined by the specified shape
  specified shape parameters
}
\description{
calculates the mean of a binomial random variable defined by the
  specified shape parameters
}
