% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_measures.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials (integer)}

\item{prob}{probability of success in one trial (integer)}
}
\value{
Calculated mean
}
\description{
Calculates the mean of the distribution
}
\examples{
# distribution with 10 trials and probability 0.3
bin_mean(trials = 10, prob = 0.3)
}
