% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean function}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability}
}
\value{
a binomial mean
}
\description{
compute the auxiliary mean
}
