% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the mean of a binomal distribution
}
\description{
calculates the mean of a binomal distribution
}
\examples{
What is the mean of number of successes in 10 trials with prob of success = 0.3?
bin_mean(10, 0.3)
}
