% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_funcs.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
the mean of the binomial distribution
}
\description{
Calculates the mean of a binomial distribution given the number of trials and the probability for success
}
\examples{
bin_mean(10, 0.3)
}
