% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_mean}
\alias{bin_mean}
\title{mean of a binomial distribution}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on rach trial}
}
\value{
mean of a binomial distribution
}
\description{
calculate the mean of a binomial distribution
}
\examples{
bin_mean(10, 0.3)

}
