% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{n}{number of trials, prob probablity of success in each trial}
}
\value{
the mean of the given binomial distribution with n trials and probability prob of success for each trial
}
\description{
calculates the mean number of successes given a number of trials and probability of success for each trial
}
