% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{n}{number of trials}
}
\value{
the mean of the given binomial distribution with "trials "trials and probability "prob" of success
}
\description{
finds the mean of a binomial distribution with "trials" and probability "prob"
}
