% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns_of_measures.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
The mean of a binomial variable
}
\description{
Calculates the mean of a binomial variable
}
\examples{
bin_mean(trials = 5, prob = 0.5)

}
