% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_mean}
\alias{bin_mean}
\title{binominal mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{how many trials}

\item{prob}{probability of success on any given trial}
}
\value{
mean of the specified binomal distribution
}
\description{
gives mean of the binomial
}
\examples{
bin_mean(6, 0.3)
}
