% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, probs)
}
\arguments{
\item{trials}{This is the number of trials}

\item{probs}{This is the probability of success}
}
\value{
The mean is returned
}
\description{
This calculates the mean of the binomial distribution
}
\examples{
bin_mean(trials=5,probs=0.5)
}
