% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{prob}{probablity of success in each trial}

\item{n}{number of trials}
}
\value{
expected value of the binomial distribution
}
\description{
calculates the expected value of the binomial distribution
}
\examples{
bin_mean(10,0.2)
}
