% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binvar.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials, }{number of trials of the binomial distribution}

\item{prob, }{probability of succes of the binomial distriburion}
}
\value{
The binomial mean of given number of trials and probability of success
}
\description{
Calculate the mean for a binomial distribution
}
\examples{
bin_mean(12, 0.7)
bin_mean(5, 0.23)
}
