% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials.}

\item{prob}{probability of success.}
}
\value{
mean of binomial distribution
}
\description{
calculate mean of binomial distribution
}
\examples{

bin_mean(trials = 5, prob = 0.5)
bin_mean(trials = 10, prob = 0.3)

}
