% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Binary Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{Trials}

\item{prob}{Probability}
}
\value{
A number for mean
}
\description{
Returns mean of binomial variable
}
