% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomal.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean Function}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
Calculates the mean of a binomial distribution
}
\description{
Calculates the mean of a binomial distribution based on their parameter values
}
\examples{
bin_mean(10,0.3)
}
