% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
computed mean
}
\description{
calculates the mean success in a binomial distribution
}
