% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_mean}
\alias{bin_mean}
\title{Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mean of binomial(trials, prob)
}
\description{
calculates the mean of binomial(trials, prob)
}
\examples{
: bin_mean(trials = 10, prob = 0.3)
}
