% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of event}
}
\value{
returns a numeric mean of the binomial distribution
}
\description{
The average value of a binomial distribution given trials and prob
}
\examples{
# mean of a binomial with 10 trials and a prob of 0.3
bin_mean(10, 0.3)

}
