% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-measure.R
\name{bin_mean}
\alias{bin_mean}
\title{mean of a binomial distribution}
\usage{
bin_mean(n, p)
}
\arguments{
\item{n}{the number of total trials}

\item{p}{the probability of single success}
}
\value{
the calculated mean
}
\description{
calculate the mean of a random variable following the binomial distribution
}
