% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{parameter p to the binomial distribution}
}
\value{
mean of binomial
}
\description{
computes binomial mean
}
\examples{

# mean
bin_mean(10, 0.3)

}
