% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{is the number of trials}

\item{prob}{is the probability of success}
}
\value{
the mean of a binomial distribution
}
\description{
calculates the mean of a binomial distribution
}
\examples{
bin_mean(trials = 10, prob = 0.3)
}
