% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a number, the mean of a binomial random variable
}
\description{
calculates the expected number of successes in n trials
}
\examples{
bin_mean(trials = 10, prob = 0.5)
}
