% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workout3.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{an integer with number of trials}

\item{prob}{the probability of the success occur}
}
\value{
the mean value
}
\description{
Calculate the mean of binomial distribution
}
\examples{
bin_mean(10, 0.5)
}
