% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunc.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success within one trial}
}
\value{
the numerical mean of the given data
}
\description{
calculates the mean of the given binomial distribution
}
\examples{

# default
bin_mean()

# another example
bin_mean(trials = 3, prob = 0.5)

}
