% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{numeric vector}

\item{prob}{numeric vector}
}
\value{
the mean of the binomial distribution
}
\description{
calculates the mean of the binomial distribution
}
\examples{

bin_mean(10, 0.3)

bin1 <- bin_mean(10, 0.3)


}
