% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of trials (numeric)}

\item{prob}{the probability of one success (numeric)}
}
\value{
the mean of the binomial distribution (numeric)
}
\description{
finds the mean of the binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
