% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_functions.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
computed binomial mean
}
\description{
computes the mean of the binomial variable
}
\examples{
bin_mean(trials = 10, prob = 0.5)
}
