% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mean}
\alias{bin_mean}
\title{: bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{:}{trials the number of trials}

\item{:}{prob the probability of success}
}
\value{
: the mean of the binomial(trials, prob)
}
\description{
: calculates the mean of the binomial with trials number of trials and probability of succeess prob
}
\examples{
: bin_mean(trials = 10, prob = 0.3)
}
