% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{integer representing n, the total number of trials}

\item{prob}{float representing p, the probability of success on a given trial}
}
\value{
mean of the distribution
}
\description{
Calculates the mean of a binomial distribution with given trials and probability
}
\examples{
bin_mean(10,.3)

bin_mean(100,2/3)
}
