% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{numeric value for the number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
computed binomial mean
}
\description{
computes the mean of the binomial experiment
}
\examples{
bin_mean(trials = 10, prob = 0.3)
}
