% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{trials}

\item{prob}{probability}
}
\value{
binmean
}
\description{
Calculates mean of a binomial distribution of a given probability in n trials
}
\examples{
bin_mean(5, 0.5)
}
