% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials:}{the number of trials.}

\item{prob:}{probability of success}
}
\value{
numeric
}
\description{
gives the mean of a binomial variable
}
