% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mean of a binomial distribution
}
\description{
calculates the mean of a binomial distribution
}
\examples{
bin_mean_5 <- bin_mean(5,0.5)

bin_mean_10 <- bin_mean(10,0.1)

}
