% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_summary_functions.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{(integer)}

\item{prob}{(real)}
}
\value{
expected number of successes
}
\description{
computes mean of binomial distribution
}
\examples{
bin_mean(trials = 5, prob = 0.5)
bin_mean(10, 0.3)
}
