% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Mean of a Binomial Distribution}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{total numbeer of trials (numeric)}

\item{prob}{the probability of a success in a trial (numeric)}
}
\value{
a number indicates the mean of the binomial distribution
}
\description{
Calculating the mean of a binomial distribution based on inputs, trials and probability
}
\examples{
bin_mean(5,0.5)
2.5
bin_mean(6,0.3)
1.8
}
