% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_mean}
\alias{bin_mean}
\title{Mean of a Binomial Random Variable}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability for success in each trial}
}
\value{
the mean of the binomial random variable
}
\description{
calculates the mean of a binomial random variable given number of trials and prob
}
