% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_mean.R
\name{bin_mean}
\alias{bin_mean}
\title{mean of binomial distribution}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of a single success}
}
\value{
calculated mean of binomial distribution
}
\description{
calculate the mean of binomial distribution
}
\examples{

bin_mean(10, 0.3)

}
