% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_summ_func.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{some integer}

\item{prob}{some real number}
}
\value{
expected number of successes in n trials
}
\description{
computes expected value or mean of a binomial distribution
}
\examples{
bin_mean(trials=5, prob=0.5)
bin_mean(10, 0.3)
}
