% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{numeric, number of trials}

\item{prob}{numeric, probability of success}
}
\value{
Expected number of successes
}
\description{
computes the expected number of successes in the given number of trials
}
\examples{
bin_mean(10, 0.3)


}
