% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials, }{the number of trials}

\item{prob, }{the probability of a success}
}
\value{
returns a numeric mean
}
\description{
gives the mean of the binomial distribution
}
\examples{
bin_mean(trials = 5L, prob = 0.5)
bin_mean(3L, .3)
}
