% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean Function}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
computed value
}
\description{
Calculate binomial mean
}
\examples{
# 10 trials, probability of 0.3
bin_mean(10, 0.3)
3
}
