% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean method}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the mean of a binomial variable
}
\description{
calculates the mean of a binomial variable
}
\examples{

# expected number of heads of tossing a fair coin 10 times
bin_mean(10, 0.5)

}
