% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main function.R
\name{bin_mean}
\alias{bin_mean}
\title{: functions of treasures}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{:}{trials and prob}
}
\value{
: a numeric object
}
\description{
: calculate the attributions of a binominal distribution
}
\examples{
:
bin_mean(10, 0.3)
bin_variance(10, 0.3)
bin_mode(10, 0.3)
bin_skewness(10, 0.3)
bin_kurtosis(10, 0.3)
}
