% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-methods.R
\name{bin_mean}
\alias{bin_mean}
\title{Mean of a Binomial Distribution}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
return expected value of a binomial distribution
}
\description{
calculate the expected value of a binomial distribution
}
\examples{
bin_mean(10, 0.3)


}
