% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{the number of successes out of n trials}
}
\value{
mean of a binomial distribution
}
\description{
calculates the expected number of successes in n trials.
}
\examples{
bin_mean(10, 0.3)
}
