% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial_main.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{probability of success}
}
\value{
compute the mean of a binomal distribution
}
\description{
calculate the mean of a binomal distribution
}
\examples{
bin_mean(10, 0.3)
}
