% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Expected Value}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of trials, a non-negative integer}

\item{prob}{the probability of success, a number between 0 and 1 both included}
}
\value{
the mean of the random variable under trials number of of trials and probability prob of success per trial
}
\description{
computes the mean of a random variable under a given number of trials and a given probability of success per trial; which is the expected number of successes in a given number of trials under a given probability
}
\examples{
# the expeceted value of a random variable in 5 trials and probability of success 0.4
bin_mean(5,0.4)

# the expected value of a random variable in 10 trials and probability of success 0.3
bin_mean(10,0.3)
}
