% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mearsures.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{an integer}

\item{prob}{a probability value}
}
\value{
expected value(i.e. mean)
}
\description{
expected value of a binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
