% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_mean}
\alias{bin_mean}
\title{Bin_Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{an integer or vector of numerics - number of trials}

\item{prob}{a numeric value between 0 and 1}
}
\value{
the mean of the variable passed in
}
\description{
finds mean of binomial given trials and prob
}
\examples{
# default
mean <- bin_mean(trials = 10, prob = 0.3)

}
