% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_summary_measures.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{representing # of trials (numeric value)}

\item{prob}{representing probability (numeric value)}
}
\value{
mean (numeric value)
}
\description{
determines mean of given probability and given number of trials
}
\examples{
bin_mean(5,0.5)
bin_mean(1,1)
}
