% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success for each trials}
}
\value{
number of success
}
\description{
the function calculates mean number of success
}
\examples{
bin_mean(10, 0.3)
}
