% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_mean}
\alias{bin_mean}
\title{the average probability}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success in one trial}
}
\value{
the numeric result of the average probability of successes in trials
}
\description{
calculate the average probability of successes in trials
}
\examples{
bin_mean(10, 0.3)
#the expeted value of successes in 10 trials with 0.3 success probability in one trial

}
