% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
mean of the binomial distribution
}
\description{
main function for summary measure "mean"
}
