% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_mean}
\alias{bin_mean}
\title{Function bin_mean()}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability of success}
}
\value{
the mean of the of binimial distribution
}
\description{
calculate the mean of the of binimial distribution
}
\examples{
bin_mean(3,0.5)
}
