% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunction.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomal mean}
\usage{
bin_mean(n, p)
}
\arguments{
\item{n}{the number of trials}

\item{p}{the probability of successes}
}
\value{
computer the mean of a binomal distribution
}
\description{
calculate the mean of a binomal distribution
}
\examples{
bin_mean(10, 0.3)
}
