% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial_Functions.R
\name{bin_mean}
\alias{bin_mean}
\title{: bin_mean()}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials:}{the number of trials}

\item{prob:}{the probability of success}
}
\value{
: calculated mean of binomial distribution
}
\description{
: mean of binomial distribution
}
