% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mean of binomial distribution
}
\description{
calculates mean of binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
