% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean function}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of n attempts (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
the mean (integer)
}
\description{
computes the mean of the distribution
}
\examples{
bin_mean(trials = 10, prob = 0.3)
}
