% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials, }{total number of trial}

\item{prob, }{chance of success}
}
\value{
mean if input are good else error
}
\description{
checks input then calcs mean
}
\examples{
bin_mean(5,0.5)
}
