% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_statmanips.R
\name{bin_mean}
\alias{bin_mean}
\title{Mean of Binomial}
\usage{
bin_mean(trials = 1, prob = 1)
}
\arguments{
\item{trials}{the number of trials in the distribution}

\item{prob}{the probability of a success}
}
\value{
The mean of the binomial distribution, using aux_mean
}
\description{
Finds the mean of the binomial distribution.
}
