% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_and_methods.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials, }{number of trials}

\item{prob, }{probablity of successes}
}
\value{
value of the mean of binomial distribution
}
\description{
calculates the mean of binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
