% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{integer value for number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
mean of binomial
#' @examples

bin_mean(5, 0.5)
bin_mean(10, 0.2)
}
\description{
finds the mean of the binomial
}
