% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials; non-negative integer value}

\item{prob}{probability of success; must be a numeric value between 0 and 1 (inclusive)}
}
\value{
numeric value indicating the mean
}
\description{
Computes the expected value of a binomial distribution, or the expected number of successes in n trials
}
\examples{
mean <- bin_mean(trials = 10, prob = 0.3)
mean
}
