% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{The number of trials}

\item{prob}{The number of successes}
}
\value{
A number
}
\description{
Calculates the expected value of a binomial distribution
}
