% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
Calculates the mean of a binomial variable
}
\description{
Calculates the mean of a binomial variable
}
