% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{accepts}{trials, probabilities}
}
\value{
binomial mean of trials and Probabilities
}
\description{
calculates binomial mean
}
