% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{prob}{}
}
\value{
The mean of the distribution
}
\description{
Binomial mean
}
\examples{
mean <- bin_mean(5, 0.5)
means <- bin_mean(1:5, 0.5)

}
