% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success on each trial (numeric)}
}
\value{
mean of binomial distribution
}
\description{
computes mean of binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
