% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{probability of success in a single trial}
}
\value{
numeric: mean value
}
\description{
Calculates mean given # of trials and probability of a single success
}
\examples{
:
bin_mean(trials = 5, prob = 0.5)
}
