% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_main.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{Number of trials (real)}

\item{prob}{Probability of successs (real)}
}
\value{
Returns the mean
}
\description{
Computes the mean of a binomial distribution
}
\examples{
bin_mean(10, 0.3)
3
}
