% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability that a trial will be a success}
}
\value{
Mean
}
\description{
Returns mean of binomial dist
}
\examples{
bin_cumulative(5, .5)
}
