% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-of-measures.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mean of binomial distribution with given parameters
}
\description{
Computes the mean of a binomial distribution
}
\examples{
# mean of distribution with trials = 10 and prob = 0.3
bin_mean(10, 0.3)
}
