% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bino.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a numeric number represents the mean
}
\description{
Computes mean of the binomal distribution
}
\examples{
# 20 trials with probability of success 0.5
mean <- bin_mean(20, 0.5)

}
