% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_function.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean Function}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of getting a success in a trial (numeric)}
}
\value{
computed mean of binomial distribution based on parameter values (numeric)
}
\description{
returns the mean of binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
