% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of random trials}

\item{prob}{probability of successes}
}
\value{
mean of a binomial distribution
}
\description{
generate the expected value or mean of a binomial distrbution
}
\examples{
bin_mean(10, 0.3)

}
