% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mean if probability and trials inputs are valid. If not, throws an error.
}
\description{
calculates mean of having certain trials and certain probability
}
\examples{
bin_mean(5, 0.6)
bin_mean(6, 0.7)
}
