% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial variance}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
variance of binomal distribution
}
\description{
calculates the variance of given binomal distribution
}
\examples{
bin_variance(10, 0.3)
}
