% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mean
}
\description{
Returns mean of binom dist with given trials and prob
}
\examples{
bin_mean(trials = 10, prob = .4)
bin_mean(110, prob = .56)
}
