% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{return the mean of the binomial distribution}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trails}

\item{prob}{the probability}
}
\value{
mean of the binomial distribution
}
\description{
return the mean of the binomial distribution
}
\examples{
bin_mean(5, 0.5)
}
