% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials to attempt}

\item{prob}{probability of success}
}
\value{
returns a the mean of the binomial distribution
}
\description{
Computes the mean of the binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
