% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_summary.R
\name{bin_mean}
\alias{bin_mean}
\title{Mean.}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of trials.}

\item{prob}{the probability of success in any given trial}
}
\value{
the computed summary measure.
}
\description{
the expected number of successes in n trials.
}
\examples{
bin_mean(trial = 5, prob = 0.5)
bin_mean(10, 0.3)
bin_mean(100, 0.7)
}
