% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_mean}
\alias{bin_mean}
\title{Binonmial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
binomial mean
}
\description{
Creates binomial mean given "trials" and "prob"
}
\examples{
# default
mean <- bin_mean(5, .5)
}
