% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Max Sima .R
\name{bin_mean}
\alias{bin_mean}
\title{mean of binomial distribution}
\usage{
bin_mean(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
mean of a binomial distribution
}
\description{
calculate the mean of a binomial distribution
}
\examples{
bin_mean(trials = 10, prob = 0.3)
}
