% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials; numeric vector}

\item{prob}{probability of success; numeric vector}
}
\value{
the mean of a binomial distribution when given trials and prob
}
\description{
summary measure for finding mean of a binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
