% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of (fixed) trials}

\item{prob}{the probability of success on each trial}
}
\value{
the mean of the binomial distribution
}
\description{
calculates the mean of the binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
