% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{mean of binomial}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{integer prob double}
}
\value{
mean double
}
\description{
return mean of binomial distribution
}
\examples{
bin_mean(trials=5,prob=0.5)
}
