% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_fun.R
\name{bin_mean}
\alias{bin_mean}
\title{mean of the binomial distribution}
\usage{
bin_mean(trials = 1, prob = 1)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mean of a binomial distribution
}
\description{
calculate the mean of a binomial distribution
}
\examples{
bin_mean(1, 1)
}
