% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(n = 10, p = 0.3)
}
\arguments{
\item{n}{value of trials}

\item{k}{value of successes}
}
\description{
to calculate the mean value of a binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
