% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean function}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the number of trial (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
the mean
}
\description{
Culculate the mean of a binomial distribution
}
