% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.R
\name{bin_mean}
\alias{bin_mean}
\title{the expected value of the binomial distribution}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability (numeric)}
}
\value{
computed expected value
}
\description{
calculate the expected value of the binomial distribution
}
\examples{
bin_mean(trials=5,prob=0.5)
}
