% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{trial numbers}

\item{prob}{success probability}
}
\value{
mean of the binomial distribution given trials and probability
}
\description{
compute mean of a binomial distribution given trials and probability
}
\examples{
bmean <- bin_mean(3, 0.5)
}
