% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability of success}
}
\value{
the mean of the binomial distribution
}
\description{
calculates the mean of the binomial distribution
}
