% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout03-Zihan-Liu.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials (integer)}

\item{prob}{probability of success (numeric)}
}
\value{
the mean of a binomial distribution
}
\description{
calculates the mean of a binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
