% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{numeric number of trials}

\item{prob}{numeric number of probability}
}
\description{
Compute the binomial distribution mean
}
\examples{
bin_mean(10,0.3)
}
