% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_mean.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{numeric}

\item{prob}{numeric}
}
\value{
the computed data
}
\description{
get mean given trials and prob
}
\examples{
bin_mean(10,0.3)
}
