% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
The Mean Value
}
\description{
Compute Binomial Mean Value
}
\examples{
bin_mean(trials = 10, prob = 0.3)
}
