% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All Functions.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probabiltiy value (numeric)}
}
\value{
mean
}
\description{
determine the mean
}
\examples{
bin_mean(10, 0.3)
}
