% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{Numeric, # of trials in the given binomial distribution}

\item{prob}{Numeric, probability of success in the given binomial distribution}
}
\value{
Mode of binomial distribution
}
\description{
Computes the mode of the binomial distribution
}
\examples{
bin_mode(trials = 10, prob = 0.5)
}
