% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func_measures.R, R/workout03-ann-deng.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode()}
\usage{
bin_mode(trials, prob)

bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
double of mode

double of mode
}
\description{
returns mode of binvar with trials and prob

returns mode of binvar with trials and prob
}
\examples{
bin_mode(10, 0.3) ## returns 3
bin_mode(10, 0.3) ## returns 3
}
