% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3-divya-vijay.R
\name{bin_mode}
\alias{bin_mode}
\title{bin mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of probability}
}
\value{
mode summary measure
}
\description{
calculates mode for binomial distribution
}
\examples{
bin_mode(10, 0.3)
}
