% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{Bin Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{an integer, the number of trials}

\item{prob}{an integer of the probability of success for one trial}
}
\value{
an integer or list of integers, if there are multiple modes
}
\description{
A function to return the mode of a binomial distribution with given trials and probability
}
\examples{
bin_mode(3,0.5)
}
