% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of random variable}
}
\value{
returns mode given by the number of trials and its probability
}
\description{
gives the mode of a binomial random variable
}
\examples{
bin_mode(10,0.3)
}
