% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Bin mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{is a numeric integer that represents the number of trials}

\item{prob}{is a numeric value between 0 and 1 that represents the probability of success}
}
\value{
this function returns a numeric value that represents the mode of the probability distribution
}
\description{
This function gives the most likely number of successes in "trials" independent trials and "prob" probability of success
}
\examples{
bin_mode(trials = 5, prob = 0.5)
bin_mode(trials = 100, prob = 0.2)
bin_mode(trials = 45, prob = 0.75)
}
