% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{Mode function}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
Mode
}
\description{
Calculates the mode
}
\examples{

#computes mode with corresponding arguments
 bin_mode(10, 0.3)

}
