% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{(numeric)}

\item{prob}{(numeric)}
}
\value{
the mode value
}
\description{
the mode of the binomial distribution
}
