% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a numeric value
}
\description{
Calculates the mode
}
\examples{
mode of binomial random distribution with 10 trials and prob = 0.5
# bin_mode(10, 0.5)
}
