% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunc.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{Probability that a success occurs}
}
\value{
mode of Binomial variable
}
\description{
calculate the mode of the binomial variable
}
\examples{
bin_mode(10, 0.3)
}
