% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mode of binomial distribution described by parameters. Could
be a vector of length 2 or 1 depending on how many most likely number of
successes there are.
}
\description{
calulates the mode, or most likely number of successes, of a binomial distribution
}
\examples{
bin_mode(5, 0.5)
}
