% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-code.R
\name{bin_mode}
\alias{bin_mode}
\title{mode of binomial distribution}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
mode of the binomial distribution
}
\description{
calculates mode of a binomial distribution
}
