% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{bin_mode}
\alias{bin_mode}
\title{mode of the binomial distribution}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials, }{prob}
}
\value{
mode
}
\description{
calculate the mode
}
\examples{
bin_mode(10, 0.3)
}
