% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_measures.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{number that is the probability of success}
}
\value{
mode, the numerical mean
}
\description{
computes mode given trials and probs
}
\examples{
#5 trials with success prob at 0.5
bin_mean(trials=5, prob=0.5)
}
