% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probablility of a success}
}
\value{
a vector of mode values
}
\description{
Calculates the mode(s) of a binomial random variable
}
\examples{
bin_mode(7, 0.5)

bin_mode(trials = 3, prob = 0.3)

}
