% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials, }{number of flips}

\item{prob, }{probability of heads}
}
\value{
mode
}
\description{
Finds the mode of a binomial random variable
}
