% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_func.R
\name{bin_mode}
\alias{bin_mode}
\title{Bin Mode}
\usage{
bin_mode(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability value}
}
\value{
the mode integer value
}
\description{
Calculates the most likely number of successes in n independent trials with prob p
}
\examples{
# mode of n=10, p=0.3
mode <- bin_mode(10, 0.3)
}
