% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{# of trials}

\item{prob}{probability of success}
}
\value{
the mode
}
\description{
calculates the mode of distribution
}
\examples{

# default
bin_mode()

# another example
bin_mode(trials = 5, prob = 0.5)

}
