% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-of-measures.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{success:}{number of trials}

\item{trials:}{probability value}
}
\value{
mode using trials and probability
}
\description{
calculates mode
}
\examples{
bin_mode(10, 0.3)
}
