% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
numeric vector with mode value
}
\description{
Computes the most likely number of successes for given
number of independent trials and probability of success
}
\examples{
bin_mode(10, 0.3)
}
