% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the calculated mode
}
\description{
calculating the mode of binomial random variables
}
