% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(prob, trials)
}
\arguments{
\item{prob}{probability of random variable}

\item{trials}{number of trials}
}
\value{
a numeric value
}
\description{
Finds the binomial mode for a given probability and number of trials
}
\examples{
 \dontrun{
 bin_mode(prob = 0.5, trials = 5)
 }
}
