% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_measures.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials (integer)}

\item{prob}{probability of success in one trial (integer)}
}
\value{
Calculated mode
}
\description{
Calculates the mode of the distribution
}
\examples{
# distribution with 10 trials and probability 0.3
bin_mode(trials = 10, prob = 0.3)
}
