% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{n}{number of trials, prob probablity of success in each trial}
}
\value{
the mode of the given binomial distribution with n trials and probability prob of success for each trial
}
\description{
calculates the mode of the number of successes given a number of trials and probability of success for each trial
}
