% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{n}{number of trials}
}
\value{
the mode of the given binomial distribution with "trials "trials and probability "prob" of success
}
\description{
finds the mode of a binomial distribution with "trials" and probability "prob"
}
