% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns_of_measures.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
Mode of a binomial variable
}
\description{
Calculates the mode of a binomial variable
}
\examples{
bin_mode(trials = 5, prob = 0.5)

}
