% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_mode}
\alias{bin_mode}
\title{binominal mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{how many trials}

\item{prob}{probability of success on any given trial}
}
\value{
mode of the specified binomal distribution
}
\description{
gives mode of the binomial
}
\examples{
bin_mode(6, 0.3)
}
