% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(trials, probs)
}
\arguments{
\item{trials}{This is the number of trials}

\item{probs}{This is the probability of success}
}
\value{
Mode of the distribution
}
\description{
This calculates the mode of the binomial distribution
}
\examples{
bin_mode(trials=5,probs=0.5)
}
