% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binvar.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials, }{number of trials of the binomial distribution}

\item{prob, }{probability of succes of the binomial distriburion}
}
\value{
The binomial mode of given number of trials and probability of success
}
\description{
Calculate the mode for a binomial distribution
}
\examples{
bin_mode(12, 0.7)
bin_mode(5, 0.23)
}
