% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials.}

\item{prob}{probability of success.}
}
\value{
mode of binomial distribution
}
\description{
calculate mode of binomial distribution
}
\examples{

bin_mode(trials = 5, prob = 0.5)
bin_mode(trials = 10, prob = 0.3)

}
