% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{Binary Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{Trials}

\item{prob}{Probability}
}
\value{
An integer for mode
}
\description{
Returns mode of binomial variable
}
