% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomal.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode Function}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
Calculates the mode of a binomial distribution
}
\description{
Calculates the mode of a binomial distribution
}
\examples{
bin_mode(10,0.3)
}
