% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
computed mode
}
\description{
calculates the mode in a binomial distribution
}
