% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on each trial}
}
\value{
the mode of a random binomial variable
}
\description{
calculate the mode of a random binomial variable
}
\examples{
# the mode of 5 trials with the probability of success in each trial = 0.5
bin_mode(5, 0.5)
}
