% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of event}
}
\value{
returns a numeric mode of the binomial distribution
}
\description{
The mode of a binomial ddistribution given trials and prob
}
\examples{
# mode of a binomial with 10 trials and a prob of 0.3
bin_mode(10, 0.3)

}
