% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-measure.R
\name{bin_mode}
\alias{bin_mode}
\title{mode of a binomial distribution}
\usage{
bin_mode(n, p)
}
\arguments{
\item{n}{the number of total trials}

\item{p}{the probability of single success}
}
\value{
the calculated mode
}
\description{
calculate the mode of a random variable following the binomial distribution
}
