% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{is the number of trials}

\item{prob}{is the probability of success}
}
\value{
the mode of a binomial distribution
}
\description{
calculates the mode of a binomial distribution
}
\examples{
bin_mode(trials = 10, prob = 0.3)
}
