% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a number, the calculated mode
}
\description{
Calculates the mode (the most likely number of successes)
}
\examples{
bin_mode(trials = 10, prob = 0.5)
}
