% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workout3.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{an integer with number of trials}

\item{prob}{the probability of the success occur}
}
\value{
the mode value
}
\description{
Calculate the mode of binomial distribution
}
\examples{
bin_mode(10, 0.5)
}
