% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunc.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success within one trial}
}
\value{
the numerical mode of the given data
}
\description{
calculates the mode of the given binomial distribution
}
\examples{

# default
bin_mode()

# another example
bin_mode(trials = 3, prob = 0.5)

}
