% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{numeric vector}

\item{prob}{numeric vector}
}
\value{
the mode of the binomial distribution
}
\description{
calculates the mode of the binomial distribution
}
\examples{

bin_mode(10 , 0.3)

}
