% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_functions.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
computed binomial mode
}
\description{
computes the mode of the binomial variable
}
\examples{
bin_mode(trials = 10, prob = 0.5)
}
