% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{input number of trials}

\item{prob}{input probability}
}
\value{
computed mode of the binomial distribution
}
\description{
calculate mode of the binomial distribution
}
\examples{
bin_mode(5,0.5)
}
