% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success per trial}
}
\value{
mode of the binoial distribution
}
\description{
calculates the mode of te binomial distribution
}
\examples{
bin_mode(10, 0.3)
}
