% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{: bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{:}{trials the number of trials}

\item{:}{prob the probability of success}
}
\value{
: the mode of the binomial(trials, prob)
}
\description{
: calculates the mode of the binomial with trials number of trials and probability of succeess prob
}
\examples{
: bin_mode(trials = 10, prob = 0.3)
}
