% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{integer representing n, the total number of trials}

\item{prob}{float representing p, the probability of success on a given trial}
}
\value{
mode of the distribution; if there are two modes, return a vector containaing both modes
}
\description{
Calculates the mode of a binomial distribution with given trials and probability
}
\examples{
bin_mode(10,.3)

bin_mode(100,2/3)
}
