% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{numeric value for the number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
computed binomial mode
}
\description{
computes the mode of the binomial experiment
}
\examples{
bin_mode(trials = 10, prob = 0.3)
}
