% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{trials}

\item{prob}{probability}
}
\value{
binmode
}
\description{
Calculates mode of a binomial distribution of a given probability in n trials
}
\examples{
bin_mode(5, 0.5)
}
