% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_mode.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the mode of the input
}
\description{
after checking if input is valid, use aux_mode to give mode
}
\examples{
bin_mode(5, 0.5)
}
