% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials:}{the number of trials.}

\item{prob:}{probability of success}
}
\value{
numeric
}
\description{
gives the mode of a binomial variable
}
