% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mode of a binomial distribution
}
\description{
calculates the mode of a binomial distribution
}
\examples{
bin_mode_5 <- bin_mode(5,0.5)

bin_mode_10 <- bin_mode(10,0.1)

}
