% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
With valid input of parameters, return the mode
}
\description{
compute the mode.
}
\examples{
bin_mode(trials = 5, prob = 0.5)
}
