% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_summary_functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{(integer)}

\item{prob}{(real)}
}
\value{
most probable number of successes
}
\description{
computes mode of a binomial distribution
}
\examples{
bin_mode(trials = 5, prob = 0.5)
bin_mode(10, 0.3)
}
