% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{Mode of a Binomial Distribution}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{total numbeer of trials (numeric)}

\item{prob}{the probability of a success in a trial (numeric)}
}
\value{
a number indicates the mode of the binomial distribution
}
\description{
Calculating the mode of a binomial distribution based on inputs, trials and probability
}
\examples{
bin_mode(5,0.5)
3
bin_mode(6,0.3)
2
}
