% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_mode}
\alias{bin_mode}
\title{Mode of a Binomial Random Variable}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability for success in each trial}
}
\value{
the mode of the binomial random variable
}
\description{
calculates the mode of a binomial random variable given number of trials and prob
}
