% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_mode.R
\name{bin_mode}
\alias{bin_mode}
\title{mode of binomial distribution}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of a single success}
}
\value{
calculated mode of binomial distribution
}
\description{
calculate the mode of binomial distribution
}
\examples{

bin_mode(10, 0.3)

}
