% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_summ_func.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{some integer}

\item{prob}{some real number}
}
\value{
most probable number of successes
}
\description{
computes most likely number of successes
}
\examples{
bin_mode(trials=5, prob=0.5)
bin_mode(10, 0.3)
}
