% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{numeric, number of trials}

\item{prob}{numeric, probability of success}
}
\value{
mode of the binomial distribution
}
\description{
computes the most likely number of successes in "trials" independent trials with probaility of success "prob" on each trial
}
\examples{
bin_mode(10, 0.3)


}
