% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials, }{the number of trials}

\item{prob, }{the probability of a success}
}
\value{
returns a numeric mode
}
\description{
gives the mode of the binomial distribution
}
\examples{
bin_mode(trials = 5L, prob = 0.5)
bin_mode(3L, .3)
}
