% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode Function}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
computed value
}
\description{
Calculate binomial mode
}
\examples{
# 10 trials, probability of 0.3
bin_mode(10, 0.3)
3
}
