% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode method}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the mode of a binomial variable
}
\description{
calculates the mode of a binomial variable
}
\examples{

# most likely number of heads of tossing a fair coin 10 times
bin_mean(10, 0.5)

}
