% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Function.R
\name{bin_mode}
\alias{bin_mode}
\title{functions of mode treasure}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the total number of trials}

\item{prob}{the probability of success every time}
}
\value{
a numeric object
}
\description{
calculate the mode of a binominal distribution
}
\examples{
bin_mode(10, 0.3)
}
