% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-methods.R
\name{bin_mode}
\alias{bin_mode}
\title{Mode of a Binomial Distribution}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
mode value of a binomial distribution
}
\description{
calculate the mode of a binomial distribution
}
\examples{
bin_mode(10, 0.3)
}
