% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial_main.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{probability of success}
}
\value{
compute the mode of a binomal distribution
}
\description{
calculate the mode of a binomal distribution
}
\examples{
bin_mode(10, 0.3)
}
