% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_function.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability value}
}
\value{
computed mode
}
\description{
calculates the mode
}
\examples{
bin_mode(10, 0.3)
}
