% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the number of trials, a non-negative integer}

\item{prob}{the probability of success, a number between 0 and 1 both included}
}
\value{
the mode of a random variable
}
\description{
computes the mode of random variable which is the most likely number of success in a given number of independent trials rounded to the greatest integer less than or equal to it
}
\examples{
# the mode of a random variable in 5 trials and probability of success 0.4
bin_mode(5,0.4)

# the mode of a random variable in 10 trials and probability of success 0.3
bin_mode(10,0.3)
}
