% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mearsures.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{an integer}

\item{prob}{a probability value}
}
\value{
mode
}
\description{
the most likely number of success
}
\examples{
bin_mode(10, 0.3)
}
