% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode Function}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{total number of trials (integer)}

\item{prob}{probability of making success in one trial (decimal)}
}
\value{
the value of mode of the specified binomial distribution
}
\description{
compute the value of mode of a binomial distribution with specified total number of
trials and success probability
}
\examples{
bin_mode(10, 0.3)
}
