% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_mode}
\alias{bin_mode}
\title{Bin_Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{an integer or vector of numerics - number of trials}

\item{prob}{a numeric value between 0 and 1}
}
\value{
the mode of the variable passed in
}
\description{
finds mode of binomial given trials and prob
}
\examples{
# default
mode <- bin_mode(trials = 10, prob = 0.3)

}
