% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
computed mode
}
\description{
to compute the mode of a binomial distribution
}
\examples{
bin_mode(trials = 10, prob = 0.3)
}
