% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{(integer) the number of trials for your binomial distribution}

\item{prob}{(numeric) numeric value between 0 and 1, gives probability of success per trial}
}
\value{
numeric value for the mode of the binomial distribution with input parameters. Will return two values for mode if prob * trials + trials is an integer
}
\description{
returns the mode of the binomial distribution for the given number of trials and probability of success per trial
}
\examples{
bin_mode(20, 0.5)
bin_mode(trials = 10, prob = 0.25)
}
