% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_mode}
\alias{bin_mode}
\title{mode of a binomial}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
calculated mode
}
\description{
calculates the mode of a binomial distribution
}
