% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_functions.R
\name{bin_mode}
\alias{bin_mode}
\title{compute the binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{is the probability of success}
}
\value{
returns the binomial mode of an object
#' @examples

#bin_mode
bin_mode(10,0.2)
}
\description{
generates binomial mode
}
