% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials:}{number of total trials}

\item{prob:}{probability of success}
}
\value{
numeric value representing mode given number of trials and prob of success
}
\description{
calculates mode of binomial scenario of n trials and p prob of success using aux_mode
}
