% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials:}{number of trials, a non-negative integer}

\item{prob:}{probability of success, a number between 0 and 1, inclusive}
}
\value{
mode of binomial
}
\description{
calculates the mode of the binomial
}
\examples{
> bin_mode(10, 0.3)
[1] 3
}
