% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_mode}
\alias{bin_mode}
\title{the mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success in one trial}
}
\value{
the most likely number of success in n independent trials with probability p of success on each trial
}
\description{
calculate the most likely number of success in n independent trials with probability p of success on each trial is m, the greater integer less than or equal to np + p
}
\examples{

bin_mode(10, 0.3)
#the most likely number of success in 10 independent trials with probability 0.3 of success on each trial

}
