% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
mode of the binomial distribution
}
\description{
main function for summary measure "mode"
}
