% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_mode}
\alias{bin_mode}
\title{Function bin_mode()}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability of success}
}
\value{
the mode of the of binimial distribution
}
\description{
calculate the mode of the of binimial distribution
}
\examples{
bin_mode(3,0.5)
}
