% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunction.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomal mode}
\usage{
bin_mode(n, p)
}
\arguments{
\item{n}{the number of trials}

\item{p}{the probability of successes}
}
\value{
computer the mode of a binomal distribution
}
\description{
calculate the mode of a binomal distribution
}
\examples{
bin_mode(10, 0.3)
}
