% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{int}{integer trials}

\item{int}{integer prob}
}
\value{
binomial mode
}
\description{
checks probability and trials, and then calculates the mode
}
\examples{
bin_mode(10, 0.3)
}
