% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mode(s) of binomial distribution
}
\description{
calculates mode(s) of binomial distribution
}
\examples{
bin_mode(10, 0.3)
}
