% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Bin Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
number that is the mode of the bin
}
\description{
main function for mode of bin
}
\examples{
bin_mode(10, 0.3)
}
