% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{numb of trials}

\item{prob}{prob of success}
}
\value{
the binomial mode
}
\description{
calculates the mode
}
\examples{
# mode of 10 trials with p=.5
bin_mode(10, .5)

# mode of 8 trials with p=.7
bin_mode(8, .7)

}
