% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode function}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the number of n attempts (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
the mode (integer)
}
\description{
computes the mode of the distribution
}
\examples{
bin_mode(trials = 10, prob = 0.3)
}
