% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials, }{total number of trial}

\item{prob, }{chance of success}
}
\value{
mode if input are good else error
}
\description{
checks input then calcs mode
}
\examples{
bin_mode(5,0.5)
}
