% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_statmanips.R
\name{bin_mode}
\alias{bin_mode}
\title{Mode of Binomial}
\usage{
bin_mode(trials = 1, prob = 1)
}
\arguments{
\item{trials}{the number of trials in the distribution}

\item{prob}{the probability of a success}
}
\value{
The mode of the binomial distribution, using aux_mode
}
\description{
Finds the mode of the binomial distribution.
}
