% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mode of binomial ditribution
}
\description{
calculates mode of binomial ditribution
}
\examples{

bin_mode(10, 0.3)
}
