% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_and_methods.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials, }{number of trials}

\item{prob, }{probablity of successes}
}
\value{
value of the mode of binomial distribution
}
\description{
calculates the mode of binomial distribution
}
\examples{
bin_mode(10, 0.3)
}
