% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{integer value for number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
mode of binomial
#' @examples

bin_mode(5, 0.5)
bin_mode(10, 0.2)
}
\description{
finds the mode of the binomial
}
