% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{The number of trials}

\item{prob}{The number of successes}
}
\value{
A number
}
\description{
Calculates the most likely number of successes in a binomial distribution
}
