% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
Calculates the mode of a binomial variable
}
\description{
Calculates the mode of a binomial variable
}
