% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{accepts}{trials, probabilities}
}
\value{
binomial mode of trials and Probabilities
}
\description{
calculates binomial mode
}
