% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{prob}{}
}
\value{
The modes of the distribution
}
\description{
Binomial mode
}
\examples{
mode <- bin_mode(5, 0.5)
modes <- bin_mode(1:5, 0.5)

}
