% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{probability of success in a single trial}
}
\value{
numeric: the mode value(s)
}
\description{
Find mode given # of trials and probability of a single success. Length of result could be 1 or 2.
}
\examples{
:
bin_mode(trials = 5, prob = 0.5)
}
