% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-of-measures.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mode of binomial distribution with given parameters
}
\description{
Computes the mode of a binomial distribution
}
\examples{
# mode of distribution with trials = 10 and prob = 0.3
bin_mode(10, 0.3)
}
