% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials (n)}

\item{prob}{probability (p)}
}
\value{
mode which is based on trials n and probability p
}
\description{
calculate the greater integer less than or equal to np + p
}
\examples{
bin_mode(10, 0.3)
}
