% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bino.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a numeric number represents the mode
}
\description{
Computes mode of the binomal distribution
}
\examples{
# 20 trials with probability of success 0.5
mode <- bin_mode(20, 0.5)

}
