% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_function.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode Function}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of getting a success in a trial (numeric)}
}
\value{
computed mode of binomial distribution based on parameter values (numeric)
}
\description{
returns the mode of binomial distribution
}
\examples{
bin_mode(10, 0.3)
}
