% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of random trials}

\item{prob}{probability of successes}
}
\value{
mode of a binomial distribution
}
\description{
generate the mode of a binomial distrbution
}
\examples{
bin_mode(10, 0.3)

}
