% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mode if probability and trials inputs are valid. If not, throws an error.
}
\description{
calculates mode of having certain trials and certain probability
}
\examples{
bin_mode(5, 0.6)
bin_mode(6, 0.7)
}
