% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mode of binomal distribution
}
\description{
calculates the mode of given binomal distribution
}
\examples{
bin_mode(10, 0.3)
}
