% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
variance
}
\description{
Returns mode of binom dist with given trials and prob
}
\examples{
bin_variance(trials = 10, prob = .4)
bin_variance(110, prob = .56)
}
