% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{return the mode of the binomial distribution}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trails}

\item{prob}{the probability}
}
\value{
mode of the binomial distribution
}
\description{
return the mode of the binomial distribution
}
\examples{
bin_mode(5, 0.5)
}
