% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials to attempt}

\item{prob}{probability of success}
}
\value{
returns a the mode of the binomial distribution
}
\description{
Computes the mode of the binomial distribution
}
\examples{
bin_mode(10, 0.3)
}
