% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{Number of trials in the binomial distribution (numeric)}

\item{prob}{Probability of success in the binomial distribution (numeric)}
}
\value{
Mode of binomial distribution
}
\description{
Computes the mode of the binomial distribution
}
\examples{
bin_mode(trials = 5, prob = 0.5)
}
