% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_summary.R
\name{bin_mode}
\alias{bin_mode}
\title{Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the number of trials.}

\item{prob}{the probability of success in any given trial}
}
\value{
the computed summary measure.
}
\description{
the most likely number of success in n independent trials with probability p of success on each trial.
}
\examples{
bin_mode(trial = 5, prob = 0.5)
bin_mode(10, 0.3)
bin_mode(100, 0.7)
}
