% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binonmial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
binomial mode
}
\description{
Creates binomial mode given "trials" and "prob"
}
\examples{
# default
mode <- bin_mode(5, .5)
}
