% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binom_functions.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(n = 10, p = 0.3)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
the mode of a distribution
}
\description{
gives the mode for a given distribution
}
