% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials; numeric vector}

\item{prob}{probability of success; numeric vector}
}
\value{
the mode of a binomial distribution when given trials and prob
}
\description{
summary measure for finding mode of a binomial distribution
}
\examples{
bin_mode(10, 0.3)
}
