% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{success rate of each trials}
}
\value{
Return the mode of binomial distribution
}
\description{
Calculates the mode of a binomial variable
}
\examples{
bin_mode(trials = 5, prob = 0.5)
}
