% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{mode of binomial}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{integer prob double}
}
\value{
mode double
}
\description{
return mode of binomial distribution
}
\examples{
bin_mode(trials=5,prob=0.5)
}
