% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_fun.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode of a binomial distribution}
\usage{
bin_mode(trials = 1, prob = 1)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success on rach trial}
}
\value{
mode of a binomial distribution
}
\description{
calculate the mode of a binomial distribution
}
\examples{
bin_mode(1, 1)
}
