% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(n = 10, p = 0.3)
}
\arguments{
\item{n}{value of trials}

\item{k}{value of successes}
}
\description{
to calculate the mode value of a binomial distribution
}
\examples{
bin_mode(10, 0.3)
}
