% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode function}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the number of trial (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
the mode
}
\description{
Culculate the vaiance of a binomial distribution
}
