% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{n}{number of trials (numeric)}

\item{p}{success rate of each trials (numeric)}
}
\value{
Return the mode of binomial distribution
}
\description{
Calculates the mode of a binomial variable
}
\examples{
bin_mode(trials = 5, prob = 0.5)
}
