% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.R
\name{bin_mode}
\alias{bin_mode}
\title{the mode of the binomial distribution}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability (numeric)}
}
\value{
computed mode value
}
\description{
calculate the mode of the binomial distribution
}
\examples{
bin_mode(trials=5,prob=0.5)
}
