% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success each trial}
}
\value{
the number of mode
}
\description{
the mode of a binominal distribution
}
\examples{
bin_mode(trials = 10, prob = 0.3)
}
