% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_mode.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
the mode of binomial
}
\description{
Calculates the mode of binomial
}
\examples{
 \dontrun{
bin_mode(trials=10L,prob=0.3)
 }
}
