% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode function}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the mode of a binomial distribution
}
\description{
compute the mode of a binomial distribution
}
\examples{
bin_mode(trials = 10, prob = 0.3)
}
