% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{trial numbers}

\item{prob}{success probability}
}
\value{
mode of the binomial distribution given trials and probability
}
\description{
compute mode of a binomial distribution given trials and probability
}
\examples{
bmode <- bin_mode(3, 0.5)
}
