% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability of success}
}
\value{
the mode of the binomial distribution
}
\description{
calculates the mode of the binomial distribution
}
