% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions and methods.R
\name{bin_mode}
\alias{bin_mode}
\title{: Binomial Mode function}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{Number of trials in the binomial distribution}

\item{prob}{Probability of success in the binomial distribution}
}
\value{
the mode of binomial distribution
}
\description{
: Compute the mode of the binomial distribution
}
