% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{numeric number of trials}

\item{prob}{numeric number of probability}
}
\description{
compute the most likely number of success
}
\examples{
bin_mode(10,0.3)
}
