% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{Bin Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability}
}
\value{
mode
}
\description{
calculates mode of a binomial distribution
}
