% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_of_measures.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the mode of a binomial variable
}
\description{
to calculates the mode of a binomial variable
}
\examples{
bin_mode(trials = 10, prob = 0.3)

}
