% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
The Mode
}
\description{
Compute Binomial Mode Value
}
\examples{
bin_mode(trials = 10, prob = 0.3)
}
