% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All Functions.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probabiltiy value (numeric)}
}
\value{
mode
}
\description{
determine the mode
}
\examples{
bin_mode(10, 0.3)
}
