% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-adele-shen.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability Function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successful trials (numeric)}

\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
probability of the number of successes occurring in the number of trials (numeric)
}
\description{
Calculates the probability of the number of successes in a fixed number of random trials
performed under identical conditions (assuming a constant probability of success on each trial)
}
\examples{
Probability of getting 2 successes in 5 trials
bin_probability(success = 2, trials = 5, prob = 0.5)

Probabilities of getting 2 or less successes in 5 trials
bin_probability(success = 0:2, trials = 5, prob = 0.5)
}
