% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{prob}{Numeric, Probability of success in the binomial distribution}

\item{n}{Numeric, # of trials in the binomial distribution}

\item{k}{Numeric, # of successes in the binomial distribution}
}
\value{
Returns Binomial Probability
}
\description{
Calculates probability of the successes in the trials of a given binomial distribution
}
\examples{
bin_probability(success = 3, trials = 5, prob = 0.5)
bin_probability(success = 0.4, trials = 3, prob = 0.5)
bin_probability(success = 2, trials = 100, prob = 0.45)
}
