% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_probability.R, R/workout03-ann-deng.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability()}
\usage{
bin_probability(success, trials, prob)

bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success}

\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
double of probability

double of probability
}
\description{
uses bin_choose to compute probability of getting success in trials

uses bin_choose to compute probability of getting success in trials
}
\examples{
bin_probability(2, 5, 0.5) ## returns 0.3125
bin_probability(2, 5, 0.5) ## returns 0.3125
}
