% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes given number of trials (non-negative integer less than trials)}

\item{trials}{number of trials (non-negative integer)}

\item{prob}{probability of success (real number between 0 and 1)}
}
\value{
the binomal probability (real number between 0 and 1)
}
\description{
calculates binomial probability given success, trials, and probability of success
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5), bin_probability(success = 0:2, trials = 5, prob = 0.5)
}
