% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Bin Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{an integer or vector of numbers, the value of successes}

\item{trials}{an integer, the number of trials}

\item{prob}{an integer of the probability of success for one trial}
}
\value{
a double, the probability of getting the specified number of successes within given trials assuming the give probability of success
}
\description{
A function to calculate the probability of getting k successes in n trials given p probability.
}
\examples{
bin_probability(4,7,0.5)
}
