% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
probability value
}
\description{
Computes the binomial probability
}
\examples{
bin_probability(2, 5, 0.5)
bin_probability(2, 5, 0.5)
bin_probability(1:4, 5, 0.8)
}
\seealso{
\code{\link{bin_distribution}}, \code{\link{bin_cumulative}}
}
