% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\title{Binomial probability}
\usage{
bin1 <- bin_probability(2,5,0.5)
}
\arguments{
\item{success}{number of success}

\item{trials}{number of trials}

\item{prob}{probability between 0-1}
}
\examples{
# standard example
bin1 <- bin_probability(2,5,0.5)
}
