% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{vector of number of successes (numeric)}

\item{trials}{vector of number of trials (numeric)}

\item{prob}{vector of probabilities (numeric)}
}
\value{
calculated probability
}
\description{
calculates the probability of getting k successes in n trials
}
\examples{

bin_probability(success = 2, trials = 5, prob = 0.5)

}
