% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Bin Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{is a numeric integer or a vector of integers that represents the number of successes observed}

\item{trials}{is a numeric integer that represents the number of trials}

\item{prob}{is a numeric value between 0 and 1 that represents the probability of success}
}
\value{
this function returns a numeric value that represents the probability of "k" successes in "n" trials
}
\description{
This function calculates the probability of getting "k" successes in "n" trials given a probability of one success "prob"
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
