% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{is a variable that repesents the number of success out of n trials (0<=k <= n)}

\item{trials}{number of trials (n >= 0)}

\item{prob}{the probaility to win in oone trials}
}
\value{
the number of combinations in which the number of successes can oocur in n trials.
}
\description{
To calculate the probaility of success in numerber of trials
}
