% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_probability}
\alias{bin_probability}
\title{Calculate choose probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{integer(s) the number of successes}

\item{trials}{integer the number of trials}

\item{prob}{the probability to success in each trials}
}
\value{
The probability of getting k successes in k trials
}
\description{
Calculates the probability of getting k successes in k trials
}
\examples{
# Calculate the probability of getting 3 success out of 5 trials, each trial with
# independent probability 0.5
bin_probability(success = 3, trials = 5, prob = 0.5)

# Calculate the probability of getting 10 success out of 10 trials, each trial with
# independent probability 0.9
bin_probability(success = 10, trials = 10, prob = 0.9)
}
