% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(trials = 5, success = 1, prob = 0.5)
}
\arguments{
\item{trials}{(numeric)}

\item{success}{(numeric)}

\item{prob}{(numeric)}
}
\value{
the probability value
}
\description{
the probability in which k successes can occur in n trials with probability p for each success to happen
}
