% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_probability}
\alias{bin_probability}
\title{Probability}
\usage{
bin_probability(trials, success, prob)
}
\arguments{
\item{prob}{probability of k occurring}

\item{n}{number of trials}

\item{k}{number of successes (can be vector)}
}
\value{
a probability
}
\description{
Calculates the probability of k successes in n trials with a specified probabilit of k occurring
}
\examples{
# probability of getting 2 successes in 5 trials
# (assuming probability of success = 0.5)
bin_probability(success = 2, trials = 5, prob = 0.5)

# 55 heads in 100 tosses of a loaded coin with 45\% chance of heads
bin_probability(success = 55, trials = 100, prob = 0.45)
}
