% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of success}

\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
computed probability
}
\description{
calculates the probability of getting x number of successes in y trials.
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
}
