% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_probability}
\alias{bin_probability}
\title{: Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{:}{success, trials, prob}
}
\value{
: probability of successes
}
\description{
: Calculates the probability of number of k successes in n trials
}
\examples{
: bin_probability(2,5,0.5) = 0.3125
}
