% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability Function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{success probability}
}
\value{
Probability of successes in trials with given prob.
}
\description{
Calculates probability of given successes in given trials, using given binomial probability.
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
