% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{Number of successes of interest}

\item{trials}{Total number of trials used}

\item{prob}{Probability of success}
}
\value{
Probability of k successes in n trials with sucess probability being p
}
\description{
Calculates the probability of k successes with the specified number of trials and success probability
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
}
