% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
probability of getting success number of successes in trials number of trials,
given prob probability of success.
}
\description{
calculates the probability of getting a certain number of success
in a certain amount of trials, given a certain probability of success, for binomail
distributions.
}
\examples{
bin_probability(2, 5, 0.5)
}
