% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-code.R
\name{bin_probability}
\alias{bin_probability}
\title{Probability of a Combination}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{chance of success}
}
\value{
the probability of getting a number of successes
}
\description{
calculates the probability of a number of successes in some trials
}
\examples{
probability of getting 2 successes in 5 trials
(assuming prob of success = 0.5)
bin_probability(success = 2, trials = 5, prob = 0.5)
}
