% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success, }{trials, prob}
}
\value{
bin_prob
}
\description{
to calculate probability of getting ? successes in ? trials, assuming prob of success = ?
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5) = 0.3125
}
