% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_probability.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success = 1, trials = 1, prob = 0.5)
}
\arguments{
\item{success}{the number of success}

\item{trials}{the number of trials}

\item{prob}{the probability of success per trial}
}
\value{
the probability of getting successes in trials with probability prob per trial
}
\description{
calculating the probability of getting successes in trials with probability prob per trial
}
\examples{
# probability of getting 2 successes in 5 trials
bin_probability(success = 2, trials = 5, prob = 0.5)

# probabilities of getting 2 or less successes in 5 trials
bin_probability(success = 0:2, trials = 5, prob = 0.5)

# 55 heads in 100 tosses of a loaded coin with 45\% chance of heads
bin_probability(success = 55, trials = 100, prob = 0.45)
}
