% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_probability.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(trials, success, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{success}{the number of successes}

\item{prob}{number that is the probability of success}
}
\value{
probability a number representing probability of getting success successes in trials trials with probability prob
}
\description{
computes probability of getting k successes in n trials with specified probability of success
}
\examples{
#2 success in 5 trials
bin_choose(trials=5, success=2, prob = 0.5)

#1,2, or 3 successes in 5 trials
bin_choose(trials=5, success = 0:2, prob = 0.5)
}
