% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success, }{number of heads}

\item{trials, }{number of flips}

\item{prob, }{probability of observing heads}
}
\value{
binomial probability
}
\description{
Calculates binomail probabilities given success, trials, prob
}
