% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindist.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes, integer}

\item{trials}{total number of trials, integer}

\item{prob}{probability of success, real}
}
\value{
the probability of getting success within trials given the probability
}
\description{
calculate the probability
}
\examples{
bin_probability(2,5,0.5)
}
