% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability value}
}
\value{
probability of success in a number of trials
}
\description{
calculates probability of success in a number of trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
}
