% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success = 2, trials = 5, prob = 0.5)
}
\arguments{
\item{success}{# of successes}

\item{trials}{# of trials}

\item{prob}{probability of success per trial}
}
\value{
the probability
}
\description{
Calculates the probability of n successes in a k trials with a specific probability
}
\examples{

# default
bin_probability()

# another example
bin_probability(success = 1, trials = 6, prob = 0.3)

}
