% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{numeric number of successes}

\item{trials}{numeric number of trials}

\item{prob}{probability of success}
}
\value{
number of combinations successes can occur in trials
}
\description{
Calculates likelihood of bin
}
\examples{

# probability of getting 2 successes in 5 trials
# (assuming prob of success = 0.5)
x = bin_probability(success = 2, trials = 5, prob = .5)

# probabilities of getting 2 or less successes in 5 trials
# (assuming prob of success = 0.5)
xa = bin_probability(success = 0:2, trials = 5, prob = .5)

# 55 heads in 100 tosses of a loaded coin with 45\% chance of heads
bin_probability(success = 55, trials = 100, prob = .45)
}
