% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the number of (fixed) trials (n ≥ 0)}

\item{trials}{the probability of success on each trial (0 ≤ p ≤ 1)}

\item{prob}{the probability of success on each trial (0 ≤ p ≤ 1)}
}
\value{
the number of combinations in which k successes can occur in n trials
}
\description{
a main function that calculates probability
}
