% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{bin_probability}
\alias{bin_probability}
\title{Finds binomial probability}
\usage{
bin_probability(prob, trials, success)
}
\arguments{
\item{prob}{probability of random variable}

\item{trials}{number of trials}

\item{success}{number of successes}
}
\value{
a vector of probabilities corresponding to success
}
\description{
Finds the probability for each success
}
\examples{
 \dontrun{
 # find probability of a binomial random variable
 prob <- 0.5
 trials <- 5
 success <- c(2, 3)
 probabilty <- bin_probability(prob, trials, success)
 }
}
