% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Probability of binomial random variable}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{a numeric vector indicating the number of successes}

\item{trials}{the number of I.I.D. Bernoulli trials}

\item{prob}{the shape paramater of each I.I.D. Bernoulli trial}
}
\value{
the probability that a binomial random variable with the specified
  shape parameters take on a particular value.
}
\description{
Calculates the probability that a binomial random variable with
  the specified shape parameters takes on a particular value.
}
\examples{
  \dontrun {
    bin_probability(success = 3, trials = 10, prob = 0.5)
  }
}
