% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob = 0.5)
}
\arguments{
\item{success}{number of successes (integer)}

\item{trials}{number of trials (integer)}

\item{prob}{probability of success in one trial (integer)}
}
\value{
the overall probability of success over trials
}
\description{
Calculates probability of success over trials
}
\examples{
# probability of getting 2 successes in 5 trials
bin_probability(success = 2, trials = 5, prob = 0.5)

# 2 or less successes in 5 trials
bin_probability(success = 0:2, trials = 5, prob = 0.5)

# 55 heads in 100 tosses of a loaded coin with 45\% chance of success
bin_probability(success = 55, trials = 100, prob = 0.45)
}
