% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(k, n, p)
}
\arguments{
\item{k}{number of successes}

\item{n}{number of trials}

\item{p}{prob}
}
\value{
binomial probability
}
\description{
calculates the probablity
}
