% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_funcs.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the number of successful trials}

\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
probability of the successes
}
\description{
Calculates the probability of the number of successes in the number of trials with the probability of success prob
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)


bin_probability(success = 0:2, trials = 5, prob = 0.5)
}
