% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{a vector of number(s) of success(es)}

\item{trials}{number of trials}

\item{prob}{the probability of success}
}
\value{
a vector of probability of success(es) with given success in trials
}
\description{
calculate the probability of getting certain success in certain trials
}
\examples{
# probability of getting 2 successes in 5 trials (assume the probability
# of success is 0.5)
bin_probability(success = 2, trials = 5, prob = 0.5)
}
