% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{integer, less than or equal to the # of trials}

\item{trials}{which must be an integer or vector of integers, greater than zero}

\item{prob}{a number between 0 and 1}
}
\value{
returns the probability of a # of successes occurring for a given number of trials
}
\description{
gives the probability of something occurring over a specified number of trials
}
\examples{
bin_probability(success = 10, trials = 15, prob = 0.5)
bin_probability(5, 10, )
}