% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{prob}{probility of success in each trial}

\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
the probability of k successes in n trials with probability prob of success in each trial
}
\description{
calculates the probability of k successes in n trials with probability prob of success
}
\examples{
# Calculates probability of 4 successes in 5 trials with probability of success .8
bin_probability(success = 4, trials = 5, prob = 0.8)
}
