% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_prob.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probabality}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{Number of Successes}

\item{trials}{Number of Trials}

\item{prob}{Probabality of Success}
}
\value{
Probablity of sucesses in trials with a probablity
}
\description{
Calculate the probablity of sucesses in trials with a probablity
}
\examples{
#Probability of getting 2 successes in 5 trials
bin_probability(success = 2, trials = 5, prob = 0.5)
#Probability of getting 2 or less sucesses in 5 trials
bin_probability(success = 0:2, trials = 5, prob = 0.5)
}
