% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, probs)
}
\arguments{
\item{success}{: Number of successes}

\item{trials}{: This is the number of trials}

\item{probs}{: This is the probability of success in a single tried}
}
\value{
The calculated binomial probability
}
\description{
This computes the probability of getting the desired successes from a given number of trials
}
\examples{
#probability of getting 2 successes in 5 trials (assuming prob of success = 0.5)
bin_probability(success = 2, trials = 5, probs = 0.5)

#probability of getting 0 to 2 successes in 5 trials (assuming prob of success = 0.5)
bin_probability(0:2, trials = 5, probs = 0.5)

#probability of getting 55 successes in 100 trials (assuming prob of success = 0.45)
bin_probability(55, trials = 100, probs = 0.45)

}
