% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{prob}{probility of success in each trial}

\item{n}{number of trials}

\item{k}{number of success}
}
\value{
returns probability, given k successes in n trials with probability of success
}
\description{
Calculates probability, given k successes in n trials with probability of success
}
\examples{
bin_probability(2, 5, 0.5)
bin_probability(45, 100, 0.55)
}
