% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin probability}
\usage{
bin_probability(k, n, p)
}
\arguments{
\item{k}{success}

\item{n}{trials}

\item{p}{prob}
}
\value{
probability
}
\description{
calculates the probability
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
