% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindis.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success, }{number of successes}

\item{trials, }{number of trials}

\item{prob, }{the probability of success}
}
\value{
The probability of having the given number of success from the given number of trials with given probability
}
\description{
Calculate the probability of a binomial random variable for desire number of successes with probability from a number of trials
}
\examples{
bin_probability(3, 7, 0.3)
bin_probability(2, 5, 0.5)
}
