% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Binary Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{Successes}

\item{trials}{Trials}

\item{prob}{Probability}
}
\value{
A float
}
\description{
Returns the probability of k successes from n trials.
}
