% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomal.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability Function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{Total number of successes}

\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
The probability of getting k successes in n trials with prob of success p
}
\description{
Computes the probability of getting k successes in n trials with prob of success p
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
