% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success = 1, trials = 2, prob = 0.5)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
numeric vector of probabilities
}
\description{
calculates the probability of success in a certain amount of trials
}
