% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-probability.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probabilities}
\usage{
bin_probability(n, k, p)
}
\arguments{
\item{n}{the total times of trials}

\item{k}{the success times}

\item{the}{probabilty of success of one single trial}
}
\value{
the probability of succeding n times
}
\description{
calculate the probabilities of certain number of times of success
}
\examples{
bin_probability(10,5,0.3)
}
