% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{is the number of successes}

\item{trials}{is the number of trials}

\item{prob}{is the probability of success}
}
\value{
probability of getting (success) successes in (trials) trials
}
\description{
calculates probability of getting (success) successes in (trials) trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = .5)
bin_probability(success = 0:2, trials = 5, prob = .5)
bin_probability(success = 55, trials = 100, prob = .45)
}
