% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{: bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{:success, }{trials, and prob}
}
\value{
: the probability
}
\description{
:calculates the probability of getting the given number of successes in given number of trials
}
