% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workout3.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{a list with number of success}

\item{trials}{an integer with number of trials}

\item{prob}{the probability of the success occur}
}
\value{
a list containging the probability of success given trials and success
}
\description{
the function used to calculate the probablity of the binomial random variable
}
\examples{
bin_probability(3, 5, 0.5)
}
