% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{numeric vector}

\item{trials}{numeric vector}

\item{prob}{numeric vector}
}
\value{
the resulted binomial probability depending on the param
}
\description{
calculates the binomial probability
}
\examples{

bin_probability(success = 2, trails = 5, prob = 0.5)

bin_probability(success = 0:2, trials = 5, prob = 0.5)


}
