% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes (numeric)}

\item{trials}{number of trials (numeric)}

\item{prob}{probabililty of success (numeric)}
}
\value{
computed probability
}
\description{
computes probability of getting successes in trials
}
\examples{
bin_probability(2, 5, 0.5)
bin_probability(0:2, 5, 0.5)
}
