% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_probability}
\alias{bin_probability}
\title{Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{input number of successes}

\item{trials}{input number of trials}

\item{prob}{input probability}
}
\value{
computed probability in which k successes can occur in n trials
}
\description{
calculate the probability in which k successes can occur in n trials
}
\examples{
bin_probability(3,6,0.5)
}
