% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_probability}
\alias{bin_probability}
\title{: bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{:}{success the number of successes}

\item{:}{trials the number of trials}

\item{:}{prob the probability of success}
}
\value{
: double: the probability in which k successes can occur in n trials
}
\description{
: computes the probability in which k successes can occur in n trials
}
\examples{
: bin_probability(success = 2,trials = 5, prob = 0.5)
: bin_probability(success = 0:2,trials = 5, prob = 0.5)
: bin_probability(success = 55,trials = 100, prob = 0.45)
}
