% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{integer or vector of integers representing k, the total number of successes}

\item{trials}{integer representing n, the total number of trials}

\item{prob}{float representing p, the probability of success on a given trial}
}
\value{
float or a vector of floats, probability X = k
}
\description{
Calculates the probability X = k for a binomial random variable X = Binomial(n,p)
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)

bin_probability(success = 0:2, trials = 5, prob = 0.5)

bin_probability(success = 55, trials = 100, prob = 0.45)
}
