% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{successes}

\item{trials}{trials}

\item{prob}{probability}
}
\value{
binprob
}
\description{
Calculates probability of given number of successes in given number of trials
}
\examples{
bin_probability(trials=5, success=2, prob=0.5)
}
