% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(trials, prob, success)
}
\arguments{
\item{n:}{the number of total objects}

\item{k:}{number of draws}
}
\value{
numeric
}
\description{
gives the probability of the number of successes within fixed trials of a random variable following a binomial distribution
}
