% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
With valid input of parameters, return the computed probability.
}
\description{
compute the probability of binomial distribution with given trial and success input.
}
\examples{
bin_probability(2, 5, 0.5)
}
