% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{(integer)}

\item{trials}{(integer)}

\item{prob}{(real)}
}
\value{
probability of successes in trials
}
\description{
computes probability of successes
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(0:2, 5, 0.5)
bin_probability(55, 100, 0.45)
}
