% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial's Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{total number of trials}

\item{prob}{probability of success in each trial}
}
\value{
the calculated probability according to the success, trials and prob inputs
}
\description{
calculate the probability of a binomial distribution which has success, trials, and prob.
}
