% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_probability}
\alias{bin_probability}
\title{Probability of a Binomial Random Variable}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of success}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
probability that the number of successes occur given the number of trials and probability of success
}
\description{
Calculates probability that the number of successes occur given the number of trials and probability of success
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
