% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_probability.R
\name{bin_probability}
\alias{bin_probability}
\title{the probability of binomial distribution}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the number of successes}

\item{trials}{the number of trials}

\item{prob}{the probability of success in a single trial}
}
\value{
calculated proability of k successes in n trials
}
\description{
calculate the probability of having k successes in n trials
}
\examples{

bin_probability(success = 2, trials = 5, prob = 0.5)

}
