% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_bin_func.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{some integer}

\item{trials}{some integer}

\item{prob}{some real number}
}
\value{
probability of k success in n trials
}
\description{
computes probability of k successes in n trials
}
\examples{
bin_probability(success=2, trials=5, prob=0.5)
bin_probability(0:2, 5, 0.5)
bin_probability(55, 100, 0.45)
}
