% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{numeric, number of successes}

\item{trials}{numeric, number of trials}

\item{prob}{numeric, probability of success}
}
\value{
the binomial probability
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)


}
