% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_probability.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(successes, trials, prob)
}
\arguments{
\item{successes}{the number of successes}

\item{trials}{the number of trials}

\item{prob}{the probability of a success}
}
\value{
returns the binomial probability
}
\description{
calculates binomial probability
}
\examples{
bin_probability(successes = 2, trials = 5L, prob = 0.5)
bin_probability(successes = 0:2, trials = 5L, prob = 0.5)
bin_probability(successes = 55, trials = 100L, prob = 0.45)
}
