% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_probability.R
\name{bin_probability}
\alias{bin_probability}
\title{Binary Probability Function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
one or more computed probabilities
}
\description{
Calculate a value using the formula of the binary probability
}
\examples{
# probability of getting 2 succeses in 5 trials
# (assuming prob of success = 0.5)
bin_probability(success = 2, trials = 5, prob = 0.5)
0.3125

# probabilities of getting 2 or less successes in 5 trials
# (assuming prob of success = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
0.03125 0.15625 0.31250

# 55 heads in 100 tosses of a loaded coin with 45\% chance of heads
bin_probability(success = 55, trials = 100, prob = 0.45)
0.01075277
}
