% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability method}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of sucesses}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the probability of getting given number of successes out of given trials with given probability
}
\description{
calculates the probability of getting given number of successes out of given trials with given probability
}
\examples{

# probability of getting 2 successes in 5 trials with prob success of 0.5
bin_probability(2, 5, 0.5)

# probability of getting 1,2,3 successes in 6 trials respectively with prob success of 0.25
bin_probability(1:3, 6, 0.25)

}
