% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main function.R
\name{bin_probability}
\alias{bin_probability}
\title{: function bin_probability()}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{:}{success,trials,prob}
}
\value{
: the probability of getting success in trials
}
\description{
: evaluate the probability of getting success in trials
}
\examples{
:
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
