% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial probability function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the number of successes}

\item{trials}{the number of trials}

\item{prob}{the probability of success on each trial}
}
\description{
the probability of a given outcome of a binomial task
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
