% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial_main.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of trialsnumber of success out of the number of trials}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
probability of k success in n trials given p probability
}
\description{
caculates the binomial probability of k success in n trials given p probability
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
