% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{numeric vaule of number of success}

\item{trials}{numeric value of number of trials}

\item{prob}{numeric value of probability of success of each trial}
}
\description{
this function calculates the probabilty of getting k successes in n trails with prob p of success
}
\examples{
Calculate the probability of have 2 successes in 5 trials with a probabiity 0.5 of success.
bin_probability(success = 2, trials = 5 prob = 0.5)
}
