% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Probability of a given number of success}
\usage{
bin_probability(success, trials, prob)
}
\value{
the calculated probability of getting success number of successes in trials number of trials
}
\description{
calculates the probability of getting a given number of successes in a given number of trials
}
\examples{
# the probability of obtaining 2 successes in 7 trials with success probability of 0.3 per trial
bin_probability(2,7,0.3)

# the probability of obtaining no successes in 4 trials with success probability of 0.8 per trial
bin_probability(0,4,0.8)
}
