% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{success numbers}

\item{trials}{trial numbers}

\item{prob}{success probability}
}
\value{
the probability of k successes in n trials
}
\description{
get probability of k successes in n trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
}
