% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability Function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{vector of number of successes (integer)total number of trials (integer)}

\item{trials}{total number of trials (integer)}

\item{prob}{probability of making success in one trial (decimal)}
}
\value{
probability of making successes in several trials
}
\description{
compute the probability of making specified number of successes in several trials
with corressponding binomial probablity
}
\examples{
# probability of getting 2 successes in 5 trials (assuming prob of success = 0.5)
bin_probability(success = 2, trials = 5, prob = 0.5)

# probabilities of getting 2 or less successes in 5 trials (assuming prob of success = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)

#55 heads in 100 tosses of a loaded coin with 45\% chance of heads
bin_probability(success = 55, trials = 100, prob = 0.45)
}
