% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_probability}
\alias{bin_probability}
\title{Bin_Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{an integer or vector of numerics - number of successes}

\item{trials}{an integer or vector of numerics - number of trials}

\item{prob}{a numeric value between 0 and 1}
}
\value{
probability of getting success successes in trials given prob of success
}
\description{
Computes the probability of a certain amount of success in a certain amount
of trials, given the probability of a single success
}
\examples{
# integers 
bin_probability(success = 2, trials = 5, prob = 0.5)

# vectorized
bin_probability(success = 0:2, trials = 5, prob = 0.5)

}
