% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probabilities}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{(integer) the number of successes whose probability you'd like to calculate}

\item{trials}{(integer) the total number of trials for yoru binomial distribution}

\item{prob}{(numeric) probability value between 0 and 1 that gives the probability of success per trial}
}
\value{
(numeric) returns the probability from a binomial distribution PMF value for each number of successes inputted
}
\description{
bin_probability() takes one or more number of successes, one value of trials and one value for probability and returns the probability of that number of successes in that many trials for a binomial distribution
}
\examples{
bin_probability(success = 8, trials = 10, prob = 0.6)
bin_probability(success = c(3, 6 ,9), trials = 12, prob = 0.25)
bin_probability(6, 15, 0.5)
}
