% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability()}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the number of successes}

\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
the binomial probability
}
\description{
calculates the binomial probability
}
