% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{probability function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
the computed probability
}
\description{
computes the probability of getting a certain amount of successes in a number of trials with probability prob
}
