% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{function used to compute the bin probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{success rate}

\item{trials}{is the number of trials}

\item{prob}{is the probability of success}
}
\value{
returns the bin probability
}
\description{
computes the probability of getting # successes in # trials
}
\examples{
#binomial probability with success, trials and probability

bin_probability(success= 2, trials = 5, prob = 0.5)
bin_probability(success=0:2, trials=5, prob=0.5)
bin_probability(success=55, trials = 100, prob= 0.45)
}
