% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_chose_bin_probability.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{representing # of successes (numeric vector)}

\item{trials}{representing # of trials (numeric value)}

\item{prob}{representing probability (numeric value)}
}
\value{
numeric vector of same length as success
}
\description{
determines probability of a given # of successes occuring in a given # of trials
}
\examples{
bin_probability(3,5,0.5)
bin_probability(c(1,2,3),5,0.5)
}
