% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success:}{number of successes}

\item{trials:}{number of total trials}

\item{prob:}{probability of success}
}
\value{
numeric value representing probability of k successes within n trials total
}
\description{
calculates the probability that a certain number of successes occur within some number of trials
}
