% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{bin_probability}
\alias{bin_probability}
\title{probability of binomial}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success:}{number of successes, a non-negative integer}

\item{trials:}{number of trials, a non-negative integer}

\item{prob:}{probability of success, a number between 0 and 1, inclusive}
}
\value{
probability of k successes in n trials with prob p
}
\description{
calculates the probability of getting [success] successes in [trials] trials assuming the probabiity of success is [prob]
}
\examples{
> bin_probability(success = 2, trials = 5, prob = 0.5)
[1] 0.3125
> bin_probability(success = 0:2, trials = 5, prob = 0.5)
[1] 0.03125 0.15625 0.31250
> bin_probability(success = 55, trials = 100, prob = 0.45)
 [1] 0.01075277
}
