% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{desired number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
numeric value(s) describing binomial probability for each success
}
\description{
calculates binomial probability of
success with probability and number of trials
}
