% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_probability}
\alias{bin_probability}
\title{Function bin_probability()}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of success}

\item{trials}{number of trials}

\item{prob}{the probability of success}
}
\value{
the probability when such situation occurs
}
\description{
calculates the probability when k successes can occur in n trials.
}
\examples{
#probability of getting 2 successes in 5 trials
# (assuming prob of success = 0.5)
bin_probability(success = 2, trials = 5, prob = 0.5)
}
