% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{int}{integer success}

\item{int}{integer trials}

\item{int}{integer prob}
}
\value{
binomial probability
}
\description{
calculated binomial probability of having k successes in n trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)

bin_probability(success = 0:2, trials = 5, prob = 0.5)

bin_probability(success = 55, trials = 100, prob = 0.45)
}
