% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{: bin_probability()}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the number of success out of trials}

\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
calcualted probability of binomial when success, trials, and probability of success are given
}
\description{
: calculate the probability of binomial distribution
}
