% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{vector or single number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
binomial probability
}
\description{
calculates probability of `k` successes in `n` trials given a binomial distribution
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)

bin_probability(success = 0:2, trials = 5, prob = 0.5)
}
