% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{bin_probability}
\alias{bin_probability}
\title{Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{denotes succeses}

\item{trials}{denotes trials}

\item{prob}{denotes probability of success}
}
\value{
the probability ranging 0 to 1
}
\description{
finds a probability given values of trials, success, and probability
}
\examples{
bin2 <- bin_probability()


}
