% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Bin Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes you want to calculate the probability of}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
probability of k successes in n trials
}
\description{
calculates probability of getting k number of successes in n trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
}
