% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
probability of successs
}
\description{
computes probability of successes
}
\examples{
# chance of 5 successes in 10 trials with p=.5
bin_probability(5, 10, .5)

# chance of 2 successes in 8 trials with p=.7
bin_probability(2, 8, .7)

}
