% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the number of k successes (integer)}

\item{trials}{the number of n attempts (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
a numeric probability (numeric)
}
\description{
calculates the probability of getting k successes in n trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
}
