% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success, }{number of successes}

\item{trials, }{total number of trials}

\item{prob, }{chance of success}
}
\value{
result of bin formula
}
\description{
evaluates the binomial formula
}
\examples{
bin_probability(3,5,0.3)
}
