% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{numeric vector, number of successes}

\item{trials}{numeric value, number of trials}

\item{prob}{numeric value between 0 and 1, probability of success}
}
\value{
computed binomial probability
}
\description{
computes the probability of getting the desired successes
}
\examples{
# probability of getting 2 successes in 5 trials (assuming prob of success = 0.5)
bin_probability(success = 2, trials = 5, prob = 0.5)
}
