% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_statmanips.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability Function}
\usage{
bin_probability(success = 1, trials = 1, prob = 1)
}
\arguments{
\item{success}{the number of successes out of all the trials}

\item{trials}{the number of trials for the Binomial Distribution}

\item{prob}{the probability for each independent Bernoulli trial}
}
\value{
The robability of having a certain amount of successes in a certain amount of trials with a certain probability
}
\description{
Takes success, trials, and prob, and checks if they are valid, then uses bin_choose and the Binomial Distribution
formula to solve for the probability
}
\examples{
bin_probability(2,5,0.5)
}
