% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of success}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
number of combinations
}
\description{
finds number of combinations in which k success can occur in n trials
}
