% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_and_methods.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success, }{number of successes, less than or equal to n}

\item{trials, }{number of trials}

\item{prob, }{probablity of successes}
}
\value{
returns the probabilty in which k successes occur in n trials with probablity of success p.
}
\description{
calculates the probability in which k successes occur in n trials with probablity of success p.
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
}
