% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{integer value for number of successes}

\item{trials}{integer value for number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
computed binomial probability
#' @examples

bin_probability(5, 2, 0.5) returns 10/32
bin_probability(5, 1:3, 0.5) returns c(5/32, 10/32, 10/32)
}
\description{
computes probability of success for binomial distribution
}
