% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_prob}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{a number}

\item{trials}{a number}

\item{prob}{a number}
}
\value{
a number, binomial probability
}
\description{
calculates prob
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
