% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_probability.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability Function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes given n trials (Could be a vector longer than 1)}

\item{trials}{number of fixed trials}

\item{prob}{probability of a success in each trial}
}
\value{
The computed binomial probability given the inputs are valid.
}
\description{
Function computes the binomial probability after taking in 3 arguments. Also checks if the 
inputs entered are valid.
}
\examples{
bin_probability(success = 2, trials = 5,prob = 0.5) 
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
