% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes; non-negative integer value}

\item{trials}{number of trials; non-negative integer value}

\item{prob}{probability of success; must be a numeric value between 0 and 1 (inclusive)}
}
\value{
numeric probability value between 0 and 1 (inclusive)
}
\description{
Calculates the probability of getting k successes in n trials assuming a specified probability value
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)

bin_probability(0:2, 5, 0.5)

bin_probability(55, 100, .45)
}
