% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{The number of successes}

\item{trials}{The number of trials}

\item{prob}{A number between 0 and 1 that expresses the probability of success}
}
\value{
A numeric value of probability itself
}
\description{
Calculates the likelihood of a binomial distribution under a certian times of success
}
