% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{Number of successes}

\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
Calculates the probability of having k successes in n trials with probability p
}
\description{
Calculates the probability of having k successes in n trials with probability p
}
