% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{accepts}{success, trials, probability}
}
\value{
returns probability for successes in number of trials
}
\description{
calculate the probability of getting number of successes in trials
}
