% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes (numeric vector)}

\item{trials}{number of trials (numeric)}

\item{prob}{probability of success on each trial (numeric)}
}
\value{
computed probability
}
\description{
computes the probability of getting k successes in n trials with probability of success prob
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(0:2, 5, 0.5)
bin_probability(55, 100, 0.45)
}
