% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{Number of successes}

\item{trials}{Number of trials}

\item{prob}{probability of success in a single trial}
}
\value{
numeric: probability of getting k successes out of n trials
}
\description{
calculates the binomial probability distribution given # of successes, # of
trials and the probability of getting a success in one trial
}
\examples{
:
bin_probability(10,15,0.4)
}
