% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R, R/hello.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)

bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of desired successes}

\item{trials}{number of trials}

\item{prob}{probability that a trial will be a success}

\item{success}{number of desired successes}

\item{trials}{number of trials}

\item{prob}{probability that a trial will be a success}
}
\value{
probability of getting number of successes from trials

probability of getting number of successes from trials
}
\description{
Calculates the probability of getting a number of successes from  number of trials with a probability of a success rate

Calculates the probability of getting a number of successes from  number of trials with a probability of a success rate
}
\examples{
bin_probability(2, 5, .5) = .3125
bin_probability(2, 5, .5) = .3125
}
