% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-functions-methods.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
binomial probability
}
\description{
Calculates the probability of getting successes in trials
}
\examples{
# probability of getting 2 successes in 5 trials
bin_probability(success = 2, trials = 5, prob = 0.5)

# probability of 2 or less successes in 5 trials
bin_probability(success = 0:2, trials = 5, prob = 0.5)
}
