% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{is the specified number of success for this binomial probability (numeric)}

\item{trials}{is the specified number of trials for this binomial probability (numeric)}

\item{prob}{is the specified probability of success for this binomial probability (numeric)}
}
\value{
the binomial probability for specified values of success, trials, and probability
}
\description{
Calculates the probability of a specified number of successes, with a specified number of trials, with a specified success probability, under binomial assumptions
}
\examples{
bin_probability (success = 2, trials = 5, prob = 0.5)
bin_probability (success = 0:2, trials = 5, prob = 0.5)
bin_probability (success = 55, trials = 100, prob = 0.45)
}
