% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Bin Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes (numeric)}

\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
probability of getting that number of successes
}
\description{
computes the probability of getting a given number of successes in a given number of trials with a probability of success
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
