% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability()}
\usage{
bin_probability(successes, trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}

\item{success}{number of successes}
}
\value{
binomial probability
}
\description{
calculates binomial probability given number of successes, trials, and probability
}
