% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of success (k)}

\item{trials}{number of trials (n)}

\item{prob}{probability (p)}
}
\value{
P(X = k)
}
\description{
calculate Pr(X = success)
}
\examples{
bin_probability(0:2, 5, 0.5)
bin_probability(2, 5, 0.5)
bin_probability(55, 100, 0.45)
}
