% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bino.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_proability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes, can be a vector}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
binomial probability for each success
}
\description{
Computes the binomial probability
}
\examples{
# 10 success in 20 trials with probability of success 0.5
probability <- bin_probability(10, 20, 0.5)

}
