% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_function.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability Function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes (numeric)}

\item{trials}{number of trials (numeric)}

\item{prob}{probability of getting a success in a trial (numeric)}
}
\value{
computed probability of getting number of success in number of trials with prob of success prob (numeric)
}
\description{
Computes the probability of getting k successes in n trials with prob of success p
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
