% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of success}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
probability if success, probability, and trials inputs are valid. If not, throws an error.
}
\description{
calculates probability of getting k successes in n trials
}
\examples{
bin_probability(2, 5, 0.6)
bin_probability(0:4, 6, 0.1)
}
