% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes or a vector (represents several times) of number of successes}

\item{trials}{number of trials}

\item{prob}{probability of successes}
}
\value{
probability of certain successes in certain number of trials with certain probability
}
\description{
calculates the probability of certain successes in certain number of trials with certain probability
}
\examples{
bin_probability(2, 5, 0.5)
}
