% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{trials}{number of trials, must be non negative int}

\item{prob}{probability of success, must be numeric between 0 and 1}

\item{successs}{number of successes, must be non negative int less than num trials}
}
\value{
computed probability
}
\description{
Computes the probability of obtaining given number of successesin given
trials with probability of success prob
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(0:2, 5, 0.5)
}
