% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{calculates the probability in which k successes can occur in n trials}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of success}

\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
the probability
}
\description{
calculates the probability in which k successes can occur in n trials
}
\examples{
bin_probability(3, 5, 0.5)
}
