% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{describe the number of success}

\item{trials}{describe the number of trials}

\item{prob}{describe probability}
}
\value{
the probability that such an event happens
}
\description{
describe the probability of having certain number of successes in certain number of trials with some probability
}
\examples{
bin_probability(3,5,0.3)
}
