% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes in the trials}

\item{trials}{number of trials to attempt}

\item{prob}{probability of success}
}
\value{
The probability of success number of successes over trials attemps given prob probabiltiy
}
\description{
computes the probability of a binomial distribution
}
\examples{
bin_probability(5,7,0.5)
}
