% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_prob,distr, cumulative, choose.R
\name{bin_probability}
\alias{bin_probability}
\title{bin probability.}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the number of successes.}

\item{trials}{the number of trials.}

\item{prob}{the probability of success in any given trial.}
}
\value{
the computed probability.
}
\description{
computes the probability of getting a number of success in a given number of trials with a probability percentage.
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)

bin_probability(7, 8, 0.1)
}
