% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{successes of trial}

\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
binomial probability
}
\description{
that calculates the binomial probability
}
\examples{
# default
prob <- bin_probability(3, 4, 0.5)

}
