% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin probability function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of success (intger)}

\item{trials}{number of trials (intger)}

\item{prob}{probability of success on each trial (numeric)}
}
\value{
the calculated binomial probability
}
\description{
calculate the binomial probability
}
