% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binom_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(k = 2, n = 5, p = 0.5)
}
\arguments{
\item{k}{number of successes}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
binomial probability that given parameters would be true integer
}
\description{
computes the binomial probability
}
\examples{
0.3125
}
