% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Max Sima .R
\name{bin_probability}
\alias{bin_probability}
\title{Function bin_probability()}
\usage{
bin_probability(k, n, p)
}
\arguments{
\item{k}{number of successes}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
the probability of k successes in n trials
}
\description{
calculate the probability of k successes in n trials
}
\examples{
bin_probability(2, 5, 0.5)
bin_probability(0:2, 5, 0.5)
bin_probability(55, 100, 0.45)
}
