% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes out of the trials; numeric vector}

\item{trials}{number of trials; numeric vector}

\item{prob}{probability of success; numeric vector}
}
\value{
binomial probability of k successes in n trials given p probability
}
\description{
takes in three arguments and caculates the binomial probability of k successes in n trials given p probability
}
\examples{
bin_probability(2, 5, 0.5)

bin_probability(success = 0:2, trials = 5, prob = 0.5)

bin_probability(success = 55, trials = 100, prob = 0.45)
}
