% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workout3_Yeongyu_Shin.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\value{
binary probability
}
\description{
compute binary probability
}
\examples{
bin_choose(2, 5, 0.5) = 0.3125
}
