% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_probability}
\alias{bin_probability}
\title{Bin probability function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{a variable that represents the number of successes out of n trials.}

\item{trials}{Number of (fixed) trials in the binomial distribution .}

\item{prob}{the probability of success on each trial.}
}
\value{
the probability of getting k successes in n trials
}
\description{
calculate the probability of getting k successes in n trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
