% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{probability of binomial distribution}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{vector; trials integer prob double}
}
\value{
probability vector
}
\description{
calculate a binomial probability
}
\examples{
bin_probability(success=2,trials=5,prob=0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
