% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_probability.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability()}
\usage{
bin_probability(success = 2, trials = 5, prob = 0.5)
}
\arguments{
\item{success}{the number of success in given trials}

\item{trials}{the number of the total trials}

\item{prob}{the probability of a succeess}
}
\value{
the probability of getting successs in trials with probability per trial
}
\description{
bin_prpbability is a function that compute each probability under given numbers of trials and success
}
\examples{
# binomial probability
bin_probability()
bin_probability(2,5,0.5)
}
