% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success = 2, trials = 5, prob = 0.5)
}
\arguments{
\item{success}{value of successes}

\item{trials}{value of trials}

\item{prob}{probability of one independent success}
}
\description{
calculate the probability of k successes occurring in n trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
0.3125
}
