% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the number of success (integer)}

\item{trials}{the number of totoal trials (integer)
#' @param prob the possibility of each success (numeric)}
}
\value{
the probability
}
\description{
calculates the probabilities of the specific times of successes can occur in trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
