% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{Number of successes (numeric)}

\item{trials}{Number of trials  (numeric)}

\item{prob}{Probability of success in one trial (numeric)}
}
\value{
Returns the probability given certain number of successes, probability and trials
}
\description{
Calculates probability given certain number of successes, probability and trials
}
\examples{
bin_probability(2,  5, 0.5)
bin_probability( 0:2,  5, 0.5)
bin_probability(55, 100, 0.45)
}
