% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{a vector of number of successes (numeric)}

\item{trials}{number of trials (numeric)}

\item{prob}{probability (numeric)}
}
\value{
computed combinations
}
\description{
calculates the binomial probability
}
\examples{
bin_probability(success=2,trials=5,prob=0.5)
bin_probability(success=0:2,trials=5,prob=0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
