% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of success}

\item{trials}{number of trials}

\item{prob}{the probability of success on each trial}
}
\value{
number of probability
}
\description{
finding the probability of getting the number of successs in trials with a probability of success on each trial
}
