% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_probability.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of success}

\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
the probability of number of combinations
}
\description{
Calculates the probability of number of combinations
}
\examples{
 \dontrun{
bin_probability(success=2L,trials=5L,prob=0.5)
bin_probability(success=0L:2L,trials=5L,prob=0.5)
bin_probability(success=55L,trials=100L,prob=0.45)
 }
}
