% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{total number of trials}

\item{prob}{probability of success in each trial}
}
\value{
a numerical value of the probability in which k successes can occur in n trials.
}
\description{
calculates the probability in which k successes can occur in n trials.
}
