% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout03-Zihan-Liu.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes (integer)}

\item{trials}{number of trials (integer)}

\item{prob}{probability of success (numeric)}
}
\value{
the probability of exactly k successes on n repeated trials
}
\description{
calculates the probability of exactly k successes on n repeated trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
}
