% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions and methods.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of success}

\item{trials}{number of trials}

\item{prob}{Probability of success in one trial}
}
\value{
the binomial probability given certain number of successes, probability and trials
}
\description{
computes the binomial probability given certain number of successes, probability and trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
