% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability Function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the numeric vector of success}

\item{trials}{the numeric vector of total trials}

\item{prob}{the numeric vector of probability for success of one single trial}
}
\value{
computed probabilities
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_probability(3, 10, 0.5)
}
