% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{vector of number of success}

\item{trials}{number of trials}

\item{prob}{the probability if success}
}
\value{
a vector of probability of success with given success in trials
}
\description{
calculate the probability of success
}
\examples{
# probability of getting 2 successes in 5 trials and assume the probability
# of success is 0.5
bin_probability(success = 2, trials = 5, prob = 0.5)
}
