% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Bin Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successful trials}

\item{trials}{total number of trials}

\item{prob}{probablity}
}
\value{
binomial probability
}
\description{
calculates binomial probability
}
