% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_probability.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{numeric vector}

\item{trials}{numeric}

\item{prob}{numeric}
}
\value{
calculated result
}
\description{
probability of getting success successes or less successes in trials trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)

bin_probability(success = 0:2, trials = 5, prob = 0.5)

bin_probability(success = 55, trials = 100, prob = 0.45)
}
