% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All Functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{vector of non-negative integer(s) less than or equal to trials (numeric)}

\item{trials}{number of trials (numeric)}

\item{prob}{probability value (numeric)}
}
\value{
probability value
}
\description{
check for validity and calculuate probability
}
\examples{
bin_probability(2, 5, 0.5)
bin_probability(0:2, 5, 0.5)
bin_probability(55, 100, 0.45)
}
