% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_of_measures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of head probability}
}
\value{
computed skewness value
}
\description{
calculates the skewness of binomial random variable
}
\examples{

bin_skewness(10 ,0.3)

}
