% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{Numeric, # of trials in the given binomial distribution}

\item{prob}{Numeric, probability of success in the given binomial distribution}
}
\value{
Skewness of binomial distribution
}
\description{
Computes the skewness of the binomial distribution
}
\examples{
bin_skewness(trials = 10, prob = 0.5)
}
