% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func_measures.R, R/workout03-ann-deng.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness()}
\usage{
bin_skewness(trials, prob)

bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
double of skewness

double of skewness
}
\description{
returns skewness of binvar with trials and prob

returns skewness of binvar with trials and prob
}
\examples{
bin_skewness(10, 0.3) ## returns 0.2760262
bin_skewness(10, 0.3) ## returns 0.2760262
}
