% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3-divya-vijay.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of probability}
}
\value{
skewness summary measure
}
\description{
calculates skewness for binomial distribution
}
\examples{
bin_skewness(10, 0.3)
}
