% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Bin Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{an integer, the number of trials}

\item{prob}{an integer of the probability of success for one trial}
}
\value{
an integer, the skewness
}
\description{
A function to return the skewness of a binomial distribution with given trials and probability
}
\examples{
bin_skewness(10,0.5)
}
