% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of random variable}
}
\value{
returns skewness given by the number of trials and its probability
}
\description{
gives the skewness of a binomial random variable
}
\examples{
bin_skewness(10,0.3)
}
