% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skew}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{vector of number of trials (numeric)}

\item{prob}{vector of probabilities (numeric)}
}
\value{
calculated skewness
}
\description{
lists binomial skewness
}
