% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Skewness function}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
Skewness
}
\description{
Calculates the skewness
}
\examples{

#computes skewness with corresponding arguments
 bin_skewness(10, 0.3)

}
