% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Calculate skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{integer the number of trials}

\item{prob}{the probability to success in each trials}
}
\value{
The skewness of a binomial distribution
}
\description{
Computes the skewness of a binomial distribution
}
\examples{
# Calculate the skewness for 10 trials, each trial with
# independent probability 0.5
bin_skewness(trials = 10, prob = 0.5)
}
