% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a numeric value
}
\description{
Calculates the skewness
}
\examples{
skeness of binomial random distribution with 10 trials and prob = 0.5
# bin_skewness(10, 0.5)
}
