% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Mode}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
skewness
}
\description{
calculates the skewness
}
\examples{
bin_mode(10, 0.3)

}
