% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
skewness value
}
\description{
calculates skewness of probability distribution
}
\examples{
bin_skewness(10, 0.3)
}
