% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{: Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{:}{trials, prob}
}
\value{
: skewness
}
\description{
: Calculates skewness given a binomial distribution
}
\examples{
: bin_skewness(10, 0.3) = 0.2760202
}
