% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{as number of trials}

\item{prob}{as probability number between 0 and 1}
}
\value{
skewness as numeric value
}
\description{
calculate asymmetry of the probability dist. of a random variable about its mean.
}
\examples{
bin_skewness(10, 0.3)
}
