% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness Function}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{success probability}
}
\value{
skewness, as a numeric value.
}
\description{
Computes skewness for binomial variable with given trials and prob. Uses auxiliary function.
}
\examples{
bin_skewness(10, 0.3)
}
