% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_measure_functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the skewness of a binomial variable
}
\description{
calculates the skewness of a binomial variable
}
\examples{
bin_skewness(trials = 10, prob = 0.3)
}
