% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
Skewness of the binomial distribution with a specified number of trials and success probability
}
\description{
Calculates the skewness of a binomial distribution
}
\examples{
bin_skewness(10, 0.3)
bin_skewness(100, 0.1)
}
