% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
skewness value of binomial distribution described by parameters
}
\description{
calulates the skewness value of a binomial distribution
}
\examples{
bin_skewness(5, 0.5)
}
