% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-code.R
\name{bin_skewness}
\alias{bin_skewness}
\title{skewness of binomial distribution}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
skewness of the binomial distribution
}
\description{
calculates the skewness of a binomial distribution
}
