% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of total trials}

\item{prob}{probabilty of success in each trial}
}
\value{
skew value
}
\description{
returns skewness binomial variable
}
\examples{
bin_skewness(10, 0,3)
}
