% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{bin_skewness}
\alias{bin_skewness}
\title{skewness of the binomial distribution}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials, }{prob}
}
\value{
skewness
}
\description{
calculate the skewness
}
\examples{
bin_skewness(10, 0.3)
}
