% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_fun.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials = 1, prob = 1)
}
\arguments{
\item{trials}{trials}

\item{prob}{probability}
}
\value{
skewness
}
\description{
calculates the skewness
}
