% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_measures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{number that is the probability of success}
}
\value{
skewness the numerical mean
}
\description{
computes skewness given trials and probs
}
\examples{
#5 trials with success prob at 0.5
bin_skewness(trials=5, prob=0.5)
}
