% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of a success}
}
\value{
the calculated skewness
}
\description{
Calculates the skewness of a binomial random variable
}
\examples{
bin_skewness(5, 0.4)

bin_skewness(trials = 2, prob = 0.9)
}
