% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials, }{number of flips}

\item{prob, }{probability of heads}
}
\value{
skewness
}
\description{
Finds the skewness of a binomial random variable
}
