% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_func.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Bin Skewness}
\usage{
bin_skewness(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability value}
}
\value{
the skewness value that is postive, negative or undefined
}
\description{
Calculates the asymmetry of the prob distribution of a random var about its mean
}
\examples{
# skewness of n=10, p=0.3
skew <- bin_skewness(10, 0.3)
}
