% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{# of trials}

\item{prob}{probability of success}
}
\value{
the numerical skewness value
}
\description{
calculates the skewness of distribution
}
\examples{

# default
bin_skewness()

# another example
bin_skewness(trials = 5, prob = 0.5)

}
