% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_measures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{number of trials. Must be a positive integer.}

\item{prob}{probability of success on each trial. Must be between 0 and 1.}
}
\value{
a measure of the skewness of the binomial distribution
}
\description{
computes the skewness of the given binomial distribution.
skewness is a measure of the assymetry of the distribution about its mean.
}
\examples{
bin_skewness()
bin_skewness(5)
bin_skewness(100, .3)
}
