% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-of-measures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{success:}{number of trials}

\item{trials:}{probability value}
}
\value{
skewness using trials and probability
}
\description{
calculates skewness
}
\examples{
bin_skewness(10, 0.3)
}
