% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
numeric vector with skewness value
}
\description{
Computes skewness measure for the given distribution
}
\examples{
bin_skewness(10, 0.3)
}
