% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the calculated skewness
}
\description{
calculating the skewness of binomial random variables
}
