% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_methods.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness Function}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success of individual trial}
}
\value{
computed skewness of a binomial distribution
}
\description{
computes the expected skewness of a binomial distribution
}
\examples{
bin_skewness(10, 0.3)
}
