% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(prob, trials)
}
\arguments{
\item{prob}{probability of random variable}

\item{trials}{number of trials}
}
\value{
a numeric value
}
\description{
Finds the binomial skewness for a given probability and number of trials
}
\examples{
 \dontrun{
 bin_skewness(prob = 0.5, trials = 5)
 }
}
