% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_measures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials (integer)}

\item{prob}{probability of success in one trial (integer)}
}
\value{
Calculated skewness
}
\description{
Calculates the skewness of the distribution
}
\examples{
# distribution with 10 trials and probability 0.3
bin_skewness(trials = 10, prob = 0.3)
}
