% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the skewness of a binomal distribution
}
\description{
calculates the skewness of a binomal distribution
}
\examples{
What is the skewness of number of successes in 10 trials with prob of success = 0.3?
bin_skewness(10, 0.3)
}
