% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_funcs.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
the skewness of the binomial distribution
}
\description{
Calculates the skewness of a binomial distribution given the number of trials and the probability for success
}
\examples{
bin_skewness(10, 0.3)
}
