% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{n}{number of trials, prob probablity of success in each trial}
}
\value{
the skewness of the given binomial distribution with n trials and probability prob of success for each trial
}
\description{
calculates the skewness of the number of successes given a number of trials and probability of success for each trial
}
