% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{n}{number of trials}
}
\value{
the skewness of the given binomial distribution with "trials "trials and probability "prob" of success
}
\description{
finds the skewness of a binomial distribution with "trials" and probability "prob"
}
