% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns_of_measures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
Skewness of a binomial variable
}
\description{
Calculates the skewness of a binomial variable
}
\examples{
bin_skewness(trials = 5, prob = 0.5)

}
