% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binominal skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{how many trials}

\item{prob}{probability of success on any given trial}
}
\value{
skewness of the specified binomal distribution
}
\description{
gives skewness of the binomial
}
\examples{
bin_skewness(6, 0.3)
}
