% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, probs)
}
\arguments{
\item{trials}{This is the number of trials}

\item{probs}{This is the probability of success}
}
\value{
skewness of the binomial distribution
}
\description{
This calculates the skewness of the given binomial experiment
}
\examples{
bin_skewness(trials=5,probs=0.5)
}
