% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{prob}{probablity of success in each trial}

\item{n}{number of trials}
}
\value{
the skewness of the binomial distribution
}
\description{
computes the skewness of the binomial distribution
}
\examples{
bin_skewness(10, 0.2)
}
