% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials.}

\item{prob}{probability of success.}
}
\value{
skewness of binomial distribution
}
\description{
calculate skewness, a measure of the asymmetry of the probability distribution of a random variable about its mean of a binomial distribution
}
\examples{

bin_skewness(trials = 5, prob = 0.5)
bin_skewness(trials = 10, prob = 0.3)

}
