% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binary Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{Trials}

\item{prob}{Probability}
}
\value{
A number for skewness
}
\description{
Returns skewness of binomial variable
}
