% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomal.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness Function}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
Calculates the skewness of a binomial distribution
}
\description{
Calculates the skewness of a binomial distribution
}
\examples{
bin_skewness(10, 0.3)
}
