% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
computed skewness
}
\description{
calculates the skewness of a binomial distribution
}
