% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of event}
}
\value{
returns a numeric skewness of the binomial distribution
}
\description{
The skewness of a binomial distribution given trials and prob
}
\examples{
# skewness of a binomial with 10 trials and a prob of 0.3
bin_skewness(10, 0.3)

}
