% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-measure.R
\name{bin_skewness}
\alias{bin_skewness}
\title{skewness of a binomial distribution}
\usage{
bin_skewness(n, p)
}
\arguments{
\item{n}{the number of total trials}

\item{p}{the probability of single success}
}
\value{
the calculated skewness
}
\description{
calculate the skewness of a random variable following the binomial distribution
}
