% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{is the number of trials}

\item{prob}{is the probability of success}
}
\value{
the skewness of a binomial distribution
}
\description{
calculates the skewness of a binomial distribution
}
\examples{
bin_skewness(trials = 10, prob = 0.3)
}
