% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a number, the calculated skewness of a binomial random variable
}
\description{
Calculates the sknewness of the graph (the assymetry of the probability distribution about its mean)
}
\examples{
bin_skewness(trials = 10, prob = 0.5)
}
