% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workout3.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{an integer with number of trials}

\item{prob}{the probability of the success occur}
}
\value{
the skewness value
}
\description{
Calculate the skewness of binomial distribution
}
\examples{
bin_skewness(10, 0.5)
}
