% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunc.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success within one trial}
}
\value{
the numerical skewness value of the given data
}
\description{
calculates the skewness of the given binomial distribution
}
\examples{

# default
bin_skewness()

# another example
bin_skewness(trials = 3, prob = 0.5)

}
