% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{numeric vector}

\item{prob}{numeric vector}
}
\value{
the skewness of the binomial distribution
}
\description{
calculates the skewness of the binomial distribution
}
\examples{

bin_skewness(10, 0.3)

}
