% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{input number of trials}

\item{prob}{input probability}
}
\value{
computed skewness of the binomial distribution
}
\description{
calculate skewness of the binomial distribution
}
\examples{
bin_skewness(5,0.5)
}
