% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success per trial}
}
\value{
skewness
}
\description{
calculates the skewness of the binomial distribution
}
\examples{
bin_skewness(10, 0.3)
}
