% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
the skewness of the binomial(trials, prob)
}
\description{
calculates the skewness of the binomial with trials number of trials and probability of succeess prob
}
\examples{
bin_skewness(trials = 10, prob = 0.3)
}
