% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{numeric value for the number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
computed binomial skewness
}
\description{
computes the skewness of the binomial experiment
}
\examples{
bin_skewness(trials = 10, prob = 0.3)
}
