% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{trials}

\item{prob}{probability}
}
\value{
binskew
}
\description{
Calculates skewness of a binomial distribution of a given probability in n trials
}
\examples{
bin_skewness(5, 0.5)
}
