% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_skewness.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_variance}
\usage{
bin_skewness(trials, prob)
}
\value{
skewness
}
\description{
after checking if input is valid, use aux_skewness to return skewness
}
\examples{
bin_skewness(trials=5,prob=0.75)
}
