% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials:}{the number of trials.}

\item{prob:}{probability of success}
}
\value{
numeric
}
\description{
gives the skewness of a binomial variable
}
