% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
skewness of a binomial distribution
}
\description{
calculates the skewness of a binomial distribution
}
\examples{
bin_skew_5 <- bin_skewness(5,0.5)

bin_skew_10 <- bin_skewness(10,0.1)

}
