% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
With valid input of parameters, return the skewness
}
\description{
compute the skewness
}
\examples{
bin_skewness(trials = 5, prob = 0.5)
}
