% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_summary_functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{(integer)}

\item{prob}{(real)}
}
\value{
skewness of binomial distribution
}
\description{
computes skewness of a binomial distribution
}
\examples{
bin_skewness(trials = 5, prob = 0.5)
bin_skewness(10, 0.3)
}
