% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Skewness of a Binomial Distribution}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{total numbeer of trials (numeric)}

\item{prob}{the probability of a success in a trial (numeric)}
}
\value{
a number indicates the skewness of the binomial distribution
}
\description{
Calculating the skewness of a binomial distribution based on inputs, trials and probability
}
\examples{
bin_skewness(5,0.5)
0
bin_skewness(6,0.3)
0.3563483
}
