% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Skewness of a Binomial Random Variable}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability for success in each trial}
}
\value{
the skewness of the binomial random variable
}
\description{
calculates the skewness of a binomial random variable given number of trials and prob
}
