% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_skewness.R
\name{bin_skewness}
\alias{bin_skewness}
\title{skewness of binomial distribution}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of a single success}
}
\value{
calculated skewness of binomial distribution
}
\description{
calculate the skewness of binomial distribution
}
\examples{

bin_skewness(10, 0.3)

}
