% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_summ_func.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{some integer}

\item{prob}{some real number}
}
\value{
skewness of binomial distribution
}
\description{
measures degree of asymmetry of probability distribution around mean
}
\examples{
bin_skewness(trials=5, prob=0.5)
bin_skewness(10, 0.3)
}
