% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{prob}{numeric, probability of success}
}
\value{
skewness of the binomial distribution
}
\description{
computes the measure of the asymmetry of the probability distribution of a random variable about its mean
}
\examples{
bin_skewness(10, 0.3)


}
