% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials, }{the number of trials}

\item{prob, }{the probability of a success}
}
\value{
returns a numeric skewness
}
\description{
gives the skewness of the binomial distribution
}
\examples{
bin_skewness(trials = 5L, prob = 0.5)
bin_skewness(3L, .3)
}
