% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness Function}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
computed value
}
\description{
Calculate binomial skewness
}
\examples{
# 10 trials, probability of 0.3
bin_skewness(10, 0.3)
0.2760262
}
