% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness method}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the skewness of a binomial variable
}
\description{
calculates the skewness of a binomial variable
}
\examples{

# skewness of the number of heads got from tossing a fair coin 10 times
bin_mean(10, 0.5)

}
