% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Function.R
\name{bin_skewness}
\alias{bin_skewness}
\title{functions of skewness treasure}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the total number of trials}

\item{prob}{the probability of success every time}
}
\value{
a numeric object
}
\description{
calculate the skewness of a binominal distribution
}
\examples{
bin_skewness(10, 0.3)
}
