% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main function.R
\name{bin_skewness}
\alias{bin_skewness}
\title{: functions of treasures}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{:}{trials and prob}
}
\value{
: a numeric object
}
\description{
: calculate the attributions of a binominal distribution
}
\examples{
:
}
