% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-methods.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Skewness of Binomial Distribution}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
skewness value of a binomial distribution
}
\description{
the measure the asymmetry of the probability distribution of a random variable about its mean.
}
\examples{
bin_skewness(10, 0.3)
}
