% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{probability of success on each trial}
}
\value{
skewness of a binomial distribution
}
\description{
calculates the skewness of a binomial distribution
}
\examples{
bin_skewness(10, 0.3)
}
