% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_skewness.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
binomial skewness
}
\description{
binomial skewness
}
\examples{
bin_skewness(2,.5)
}
