% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{probability of success}
}
\value{
compute the skewness of a binomal distribution
}
\description{
calculate the skewness of a binomal distribution
}
\examples{
bin_skewness(10, 0.3)
}
