% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_function.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability value}
}
\value{
computed skewness
}
\description{
calculates the skewness
}
\examples{
bin_skewness(10, 0.3)
}
