% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{numeric value of number of trials}

\item{prob}{numeric value of probability of success of each trial}
}
\description{
this function returns the skewness of the binomial distribution
}
\examples{
Calculate the skewness of the binomial distribution with 5 trials and probability of
0.5 of success.
bin_skewness(trials = 5, prob = 0.5)
}
