% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the number of trials, a non-negative integer}

\item{prob}{the probability of success, a number between 0 and 1 both included}
}
\value{
skewness of a probability distribution of the random variable
}
\description{
computes the skewness of a probability distribution of a random variable which measures the asymmetry of the probability distribution of a random variable about its mean.
}
\examples{
# the skewness of a random variable in 5 trials and probability of success 0.4
bin_skewness(5,0.4)

# the skewness of a random variable in 10 trials and probability of success 0.3
bin_skewness(10,0.3)
}
