% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mearsures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{an integer}

\item{prob}{a probability value}
}
\value{
skewness
}
\description{
a measure of the asymmetry of the probability distribution
}
\examples{
bin_skewness(10, 0.3)
}
