% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness Function}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{total number of trials (integer)}

\item{prob}{probability of making success in one trial (decimal)}
}
\value{
the value of skewness of the specified binomial distribution
}
\description{
compute the value of skewness of a binomial distribution with specified total number of
trials and success probability
}
\examples{
bin_skewness(10, 0.3)
}
