% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Bin_Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{an integer or vector of numerics - number of trials}

\item{prob}{a numeric value between 0 and 1}
}
\value{
the skewness of the variable passed in
}
\description{
finds skewness of binomial given trials and prob
}
\examples{
# default
skew <- bin_skewness(trials = 10, prob = 0.3)

}
