% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{compute the binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{is the probability of success}
}
\value{
returns the binomial skewness of an object
#' @examples

#bin_skewness
bin_skewness(10,0.2)
}
\description{
generates binomial skewness
}
