% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_summary_measures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{representing # of trials (numeric value)}

\item{prob}{representing probability (numeric value)}
}
\value{
skewness (numeric value)
}
\description{
determines skewness of given probability and given number of trials
}
\examples{
bin_skewness(5,0.5)
bin_skewness(1,1)
}
