% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials:}{number of total trials}

\item{prob:}{probability of success}
}
\value{
numeric value representing skewness given number of trials and prob of success
}
\description{
calculates mode of binomial scenario of n trials and p prob of success using aux_skewness
}
