% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials:}{number of trials, a non-negative integer}

\item{prob:}{probability of success, a number between 0 and 1, inclusive}
}
\value{
skewness of binomial
}
\description{
calculates the skewness of the binomial
}
\examples{
> bin_skewness(10, 0.3)
[1] 0.2760262
}
