% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_skewness}
\alias{bin_skewness}
\title{skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success in one trial}
}
\value{
The skewness of a binomial random variable
}
\description{
Skewness is a measure of the asymmetry of the probability distribution of a random variable about its mean. The skewness value can be positive or negative, or undefined.
}
\examples{

bin_skewness(10, 0.3)

}
