% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
skewness of the binomial distribution
}
\description{
main function for summary measure "skewness"
}
