% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Function bin_skewness()}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability of success}
}
\value{
the skewness of the of binimial distribution
}
\description{
calculate the skewness of the of binimial distribution
}
\examples{
bin_skewness(3,0.5)
}
