% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunction.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomal skewness}
\usage{
bin_skewness(n, p)
}
\arguments{
\item{n}{the number of trials}

\item{p}{the probability of successes}
}
\value{
computer the skewness of a binomal distribution
}
\description{
calculate the skewness of a binomal distribution
}
\examples{
bin_skewness(10, 0.3)
}
