% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{int}{integer trials}

\item{int}{integer prob}
}
\value{
binomial skewness
}
\description{
checks probability and trials, and then calculates the skewness
}
\examples{
bin_skewness(10, 0.3)
}
