% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial_Functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{: bin_skewness()}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials:}{the number of trials}

\item{prob:}{the probability of success}
}
\value{
: calculated skewness of binomial distribution
}
\description{
: skewness of binomial distribution
}
