% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Bin Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
number that is the skewness of the bin
}
\description{
main function for skewness of bin
}
\examples{
bin_skewness(10, 0.3)
}
