% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness function}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the number of n attempts (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
the skewness (integer)
}
\description{
computes the skewness of the distribution
}
\examples{
bin_skewness(trials = 10, prob = 0.3)
}
