% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials, }{total number of trial}

\item{prob, }{chance of success}
}
\value{
skewness if input are good else error
}
\description{
checks input then calcs skewness
}
\examples{
bin_skewness(5,0.5)
}
