% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_statmanips.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Skewness of Binomial}
\usage{
bin_skewness(trials = 1, prob = 1)
}
\arguments{
\item{trials}{the number of trials in the distribution}

\item{prob}{the probability of a success}
}
\value{
The skewness of the binomial distribution, using aux_skewness
}
\description{
Finds the skewness of the binomial distribution.
}
