% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
skewness of binomial ditribution
}
\description{
calculates skewness of binomial ditribution
}
\examples{

bin_skewness(10, 0.3)
}
