% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{integer value for number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
skewness of binomial
#' @examples

bin_skewness(5, 0.5)
bin_skewness(10, 0.2)
}
\description{
finds the skewness of the binomial
}
