% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials; non-negative integer value}

\item{prob}{probability of success; must be a numeric value between 0 and 1 (inclusive)}
}
\value{
numeric value indicating skewness of distribution
}
\description{
Calculates the asymmetry of the probabiltiy distribution about its mean; can be positive, negative, or undefined
}
\examples{
skew <- bin_skewness(10, 0.3)
skew
}
