% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{The number of trials}

\item{prob}{The number of successes}
}
\value{
A number
}
\description{
Calculates the asymmetry of the probability distribution of a random variable about its mean
}
