% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
Calculates the skewness of a binomial variable
}
\description{
Calculates the skewness of a binomial variable
}
