% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{accepts}{trials, probabilities}
}
\value{
binomial skewness of trials and Probabilities
}
\description{
calculates binomial skewness
}
