% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{prob}{}
}
\value{
The skewness of the distribution
}
\description{
Binomial skewness
}
\examples{
skew <- bin_skewness(5, 0.5)
skews <- bin_skewness(1:5, 0.5)

}
