% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{probability of success in a single trial}
}
\value{
numeric: the skewness value
}
\description{
Find skewness given # of trials and probability of a single success
}
\examples{
:
bin_skewness(trials = 5, prob = 0.5)
}
