% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-of-measures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
skewness of binomial distribution with given parameters
}
\description{
Computes the skewness of a binomial distribution
}
\examples{
# skewness of distribution with trials = 10 and prob = 0.3
bin_skewness(10, 0.3)
}
