% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{chance of success}
}
\value{
the skewness of a binomial with given parameters
}
\description{
Computes the skewness of a binomial
}
\examples{
# skewness of 10 trials with p=.5
bin_skewness(10, .5)

# skewness of 8 trials with p=.7
bin_skewness(8, .7)

}
