% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials (n)}

\item{prob}{probability (p)}
}
\value{
skewness of binomial distribution
}
\description{
calculate skewness of binomial distribution
}
\examples{
bin_skewness(10, 0.3)
}
