% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_function.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness Function}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of getting a success in a trial (numeric)}
}
\value{
computed skewness of binomial distribution based on parameter values (numeric)
}
\description{
returns the skewness of binomial distribution
}
\examples{
bin_skewness(10, 0.3)
}
