% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of random trials}

\item{prob}{probability of successes}
}
\value{
skewness of a binomial distribution
}
\description{
generate the skewness of a binomial distrbution
}
\examples{
bin_skewness(10, 0.3)

}
