% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
skewness if probability and trials inputs are valid. If not, throws an error.
}
\description{
calculates skewness of having certain trials and certain probability
}
\examples{
bin_skewness(5, 0.6)
bin_skewness(6, 0.7)
}
