% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
skewness of binomal distribution
}
\description{
calculates the skewness of given binomal distribution
}
\examples{
bin_skewness(10, 0.3)
}
