% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
skewness
}
\description{
Returns skewness of binom dist with given trials and prob
}
\examples{
bin_skewness(trials = 10, prob = .4)
bin_skewness(110, prob = .56)
}
