% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{return the skewness of the binomial distribution}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trails}

\item{prob}{the probability}
}
\value{
skewness of the binomial distribution
}
\description{
return the skewness of the binomial distribution
}
\examples{
bin_skewness(5, 0.5)
}
