% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials to attempt}

\item{prob}{probability of success}
}
\value{
returns a the skewness of the binomial distribution
}
\description{
Computes the skewness of the binomial distribution
}
\examples{
bin_skewness(10, 0.3)
}
