% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{Number of trials in the binomial distribution (numeric)}

\item{prob}{Probability of success in the binomial distribution (numeric)}
}
\value{
Skewness of binomial distribution
}
\description{
Computes the skewness of the binomial distribution
}
\examples{
bin_skewness(trials = 5, prob = 0.5)
}
