% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_summary.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Skewness.}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the number of trials.}

\item{prob}{the probability of success in any given trial}
}
\value{
the computed summary measure.
}
\description{
a measure of the asymmetry of the probability distribution of a random variable about its mean.
}
\examples{
bin_skewness(trial = 5, prob = 0.5)
bin_skewness(10, 0.3)
bin_skewness(100, 0.7)
}
