% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binonmial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
binomial skewness
}
\description{
Creates binomial skewness given "trials" and "prob"
}
\examples{
# default
skew <- bin_skewness(5, .5)
}
