% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Max Sima .R
\name{bin_skewness}
\alias{bin_skewness}
\title{skewness of binomial distribution}
\usage{
bin_skewness(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
computed skewness
}
\description{
calculate the skewness of a binomial distribution
}
\examples{
bin_skewness(trials = 10, prob = 0.3)
}
