% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic Functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success in each trials}
}
\value{
Return the skewness of binomial distribution
}
\description{
Calculates the skewness of a binomial variable
}
\examples{
bin_skewness(trials = 5, prob = 0.5)
}
