% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{skewness of binomial}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{integer}

\item{prob}{double}
}
\value{
skewness double
}
\description{
return skewness of binomial distribution
}
\examples{
bin_skewness(trials=5,prob=0.5)
}
