% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_fun.R
\name{bin_skewness}
\alias{bin_skewness}
\title{skewness of a binomial distribution}
\usage{
bin_skewness(trials = 1, prob = 1)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on rach trial}
}
\value{
skewness of a binomial distribution
}
\description{
calculate the skewness of a binomial distribution
}
\examples{
bin_skewness(1,1)
}
