% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(n = 10, p = 0.3)
}
\arguments{
\item{n}{value of trials}

\item{k}{value of successes}
}
\description{
to calculate the skewness value of a binomial distribution
}
\examples{
bin_skewness(10, 0.3)
}
