% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness function}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the number of trial (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
the skewness
}
\description{
Culculate the skewness of a binomial distribution
}
