% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{n}{number of trials (numeric)}

\item{p}{success rate of each trials (numeric)}
}
\value{
Return the skewness of binomial distribution
}
\description{
Calculates the skewness of a binomial variable
}
\examples{
bin_skewness(trials = 5, prob = 0.5)
}
