% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.R
\name{bin_skewness}
\alias{bin_skewness}
\title{the skewness of the binomial distribution}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability (numeric)}
}
\value{
computed skewness value
}
\description{
calculate the skewness of the binomial distribution
}
\examples{
bin_skewness(trials=5,prob=0.5)
}
