% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success each trial}
}
\value{
the number of skewness
}
\description{
the skewness of a binominal distribution
}
\examples{
bin_skewness(trials = 10, prob = 0.3)
}
