% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_skewness.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
the skewness of binomial
}
\description{
Calculates the skewness of binomial
}
\examples{
 \dontrun{
bin_skewness(trials=10L,prob=0.3)
 }
}
