% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{trial numbers}

\item{prob}{success probability}
}
\value{
skewness of the binomial distribution given trials and probability
}
\description{
compute skewness of a binomial distribution given trials and probability
}
\examples{
bskew <- bin_skewness(3, 0.5)
}
