% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability of success}
}
\value{
the skewness of the binomial distribution
}
\description{
calculates the skewness of the binomial distribution
}
