% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout03-Zihan-Liu.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials (integer)}

\item{prob}{probability of success (numeric)}
}
\value{
the skewness of a binomial distribution
}
\description{
calculates the skewness of a binomial distribution
}
\examples{
bin_skewness(10, 0.3)
}
