% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions and methods.R
\name{bin_skewness}
\alias{bin_skewness}
\title{: Binomial Skewness function}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{Number of trials in the binomial distribution}

\item{prob}{Probability of success in the binomial distribution}
}
\value{
the skewness of binomial distribution
}
\description{
: Compute the skewness of the binomial distribution
}
