% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Bin Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability}
}
\value{
skewness
}
\description{
calculates skewness of a binomial distribution
}
