% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_skewness.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{numeric}

\item{prob}{numeric}
}
\value{
the computed data
}
\description{
return mode given trials and prob
}
\examples{
bin_skewness(10,0.3)
}
