% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
The Skewness
}
\description{
Compute Binomial Skewness Value
}
\examples{
bin_skewness(trials = 10, prob = 0.3)
}
