% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All Functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probabiltiy value (numeric)}
}
\value{
skewness
}
\description{
determine the skewness
}
\examples{
bin_skewness(10, 0.3)
}
