% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number indicating the number of trials}

\item{prob}{number indicating teh probability of getting success}
}
\value{
the skewness
}
\description{
calculate the skewness with certain trials and prob
}
\examples{
bin_skewness(10, 0.3)
}
