% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-adele-shen.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Cumulative Distribution Function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
object of class "binvar", a binomial random variable object with elements "trials" and "prob"
}
\description{
Returns a list that displays the number of trials and probability of success
}
\examples{
# 10 trials and probability of success 0.3
bin_variable(10, 0.3)
}
