% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of head probability}
}
\value{
an object of class "binvar", that is, a binomial random variable object
}
\description{
creates an object of class "binvar", which is a list with named elements trials and prob
}
\examples{

bin1 <- bin_variable(trials = 10, p = 0.3)
bin1

binsum1 <- summary(bin1)
binsum1

}
