% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials = 10, prob = 0.4)
}
\arguments{
\item{trials}{Numeric, # of trials}

\item{prob}{Numeric, probability from 0 to 1 inclusive}
}
\value{
a list containing the two parameters of the binomial RV
}
\description{
Function that returns a list of named elements trials and prob
trials and probability.
}
\examples{
bin_variable(10, .5)
bin_variable(100, .25)
}
