% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_random_var.R, R/workout03-ann-deng.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable()}
\usage{
bin_variable(trials, prob)

bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
object of class "binvar"

object of class "binvar"
}
\description{
returns object of class "binvar" (binomial random variable)

returns object of class "binvar" (binomial random variable)
}
\examples{
bin_variable(5, 0.5) ## returns list(5, 0.5)
bin_variable(5, 0.5) ## returns list(5, 0.5)
}
