% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (non-negative integer)}

\item{prob}{probability of success (real number between 0 and 1)}
}
\value{
generates a "binvar" class object that is a list of trials and prob
summary will generate
}
\description{
function used to generate summary stats for the binomial probability
}
