% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3-divya-vijay.R
\name{bin_variable}
\alias{bin_variable}
\title{bin variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of probability}
}
\value{
an list of class binvar
}
\description{
returns a list of trials and probability
}
\examples{
bin_variable(trials = 10, p = 0.3)
bin_variable(trials = 5, p = 0.1)
}
