% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Bin Variables}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{an integer, the number of trials}

\item{prob}{an integer of the probability of success for one trial}
}
\value{
a named list of the "binvar" class with the trials and the probability
}
\description{
A function to create a binvar object with the given trials and probability
}
\examples{
bin_variable(10,0.3)
}
