% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
an object of class \code{"binvar"}
}
\description{
Creates a binomial random variable object
}
\examples{
binex <- bin_variable(10, 0.3)
binex

summary(binex)
}
