% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variability}
\title{Binomial variability}
\usage{
bin1 <- bin_variable(10,0.3)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability between 0-1}
}
\examples{
# standard example
bin1 <- bin_variable(10,0.3)
}
