% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of random variable}
}
\value{
returns an object of class \code{c("binvar")}
}
\description{
creates an object of class \code{c("binvar")}
}
\examples{
bin_variable(5,0.2)
}
