% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{vector of number of trials (numeric)}

\item{prob}{vector of probabilities (numeric)}
}
\value{
named list
}
\description{
creates a list of binomial random variables
}
\examples{

bin_variable(trials = 10, prob = 0.3)

bin1 <- bin_variable(trials = 10, prob = 0.3)
print(bin1)
summary(bin1)

binsum1 <- summary(bin1)
print(binsum1)

}
