% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Bin variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{is a numeric integer that represents the number of trials}

\item{prob}{is a numeric value between 0 and 1 that represents the probability of success}
}
\value{
a list with two elements: "trials" and "prob". All values in both elements will be numeric.
}
\description{
This function returns a list of class binvar with elements "trials" and "prob", where "trials" is the number of trials and "prob" is the corresponding probability of success.
}
\examples{
bin_variable(trials = 5, prob = 0.5)
bin_variable(trials = 100, prob = 0.2)
bin_variable(trials = 45, prob = 0.75)
}
