% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variable}
\alias{bin_variable}
\title{Variable function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
Printed content of the parameters of a "binvar" object: trials and prob
}
\description{
Generates information of a variable's parameters
}
\examples{

#List with named elements, trials and prob respectively
 bin_variable(trials = 10, p = 0.3)

}
