% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable object}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (n >= 0)}

\item{prob}{the probaility to win in oone trials}
}
\value{
the object of class "binvar", that is, a binomial random variable object
}
\description{
To list out the number of trails and the probaility of success
}
