% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_variable}
\alias{bin_variable}
\title{Create a binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{integer the number of trials}

\item{prob}{the probability to success in each trials}
}
\value{
an object of class \code{'binvar'}
}
\description{
Creates an object of class \code{'binvar'}
}
\examples{
# A bin variable with 10 trials and success rate 0.5
bin_variable(trials = 10, prob = 0.5)
}
