% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{(numeric)}

\item{prob}{(numeric)}
}
\value{
a list of information about the object
}
\description{
the binomial random variable object
}
