% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_variable}
\alias{bin_variable}
\title{: bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{:}{trials, prob}
}
\value{
: bincum object
}
\description{
: creates a binomial variable
}
\examples{
: bin_variable(5, 0.5)
}
