% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin variable function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
return an object of class "binvar"
}
\description{
returns variables of the binomial distribution
}
\examples{
bin_variable(3, .5)
}
