% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a list of binomial variable
}
\description{
gives the binomial variables
}
\examples{
bin_variable(trials = 10, p = 0.3)

}
