% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
binvar object
}
\description{
create a binvar object which has useful methods such as print
and summary which provide a lot of informaton.
}
\examples{
bin_variable(5, 0.5)
}
