% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-code.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable Object}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
lists number of trials, probability, and object attribute
}
\description{
lists the number of trials, probability, and attribute of the object
}
