% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials, }{prob}
}
\value{
an object of class "binvar", that is, a binomial random variable object
}
\description{
a list that displaces the number of trials and the probability of success
}
\examples{
bin_variable(trials = 10, p = 0.3)
}
