% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success per trial}
}
\value{
a list with class "binvar
}
\description{
showing the information about variables
}
\examples{
# binomial random variable
bin_variable(trials = 10, p = 0.3)
}
