% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{number that is the probability of success}
}
\value{
bindis a last with trials and prob
}
\description{
computes binvar attributes
}
\examples{
#5 trials with success prob at 0.5
bin_variable(trials=5, prob=0.5)

lst <- bin_variable(trials=5, prob = 0.5)
print(lst)
summary(lst)
}
