% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials, }{number of flips}

\item{prob, }{probability of observing heads}
}
\value{
binvar object that is a list with trials and prob information
}
\description{
Creates a list with trials and prob information
}
