% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func.R
\name{bin_variable}
\alias{bin_variable}
\title{Bin Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a list with named elements trials and prob
}
\description{
Creates an object of class binvar that is a binomial random variable
}
\examples{
# bin variable of five trials with a half probability
var <- bin_variable(5, 0.5)
}
