% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindist.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of total trials, integer}

\item{prob}{probability of success}
}
\value{
list with number of trials and probability of success
}
\description{
binomial distribution variables
}
\examples{
bin_variable(10,0.5)
}
