% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability value}
}
\value{
a list with named elements, class "binvar"
}
\description{
returns list of number of trials and probability of success
}
\examples{
bin_variable(10, 0.3)
}
