% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{# of trials}

\item{prob}{probability of success}
}
\value{
a list with # of trials and probability of success of the binomial variable
}
\description{
creates a binomial variable object
}
\examples{

# default
bin_variable()

# another example
bin_variable(trials = 4, prob = 0.3)

}
