% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
return a list with named elements: trials: number of trials,
prob: probability of success
}
\description{
return a list with named elements: trials: number of trials,
prob: probability of success
}
\examples{
bin_variable(trials = 10, p = 0.3)

}
