% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability value of a binomial variable}
}
\value{
an object of class "binvar", which is a list with named elements: trials, prob
}
\description{
creates an object of class "binvar"
}
\examples{
bin_variable(trials = 10, p = 0.3)
}
