% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials:}{total number of trials}

\item{prob:}{probability of success}
}
\value{
list of trials and prob values
}
\description{
extracts trials and probability values after using bin_cumulative
}
