% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variable}
\alias{bin_variable}
\title{Create Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
binvar object containing trials and prob
}
\description{
creates a binomial random variable object (class "binvar"),
essentially a list with named elements for trials and probability of success
}
\examples{
bin_variable(trials = 10, p = 0.3)
}
