% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(prob = 0.5, trials = 10)
}
\arguments{
\item{prob}{probability of random variable}

\item{trials}{number of trials}
}
\value{
an object of class \code{"binvar"}
}
\description{
Creates an object of class \code{"binvar"}
}
\examples{
 \dontrun{
 bin <- bin_variable(prob = 0.5, trials = 5)
 binsum <- summary(bin)
 bin
 binsum 
 }
}
