% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial random variable constructor}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of I.I.D. Bernoulli trials}

\item{prob}{the shape parameter of each I.I.D. Bernoulli trial}
}
\value{
an object representing a binomial random variable
}
\description{
Creates a binomial random variable with the specified shape
  parameters.
}
\examples{
binvar = bin_variable(trials = 10, prob = 0.5)
cdf = bin_cumulative(binvar$trials, binvar$prob)
}
