% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (integer)}

\item{prob}{probability of success in one trial (integer)}
}
\value{
'binvar' object with attributes trials and prob
}
\description{
Creates an object of class 'binvar' (binomial random variable)
}
\examples{
# Random Variable with 10 trials and probability 0.3
bin_variable(trials = 10, prob = 0.3)
}
