% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable()}
\usage{
bin_variable(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{prob}
}
\value{
return a list with named elements: trials: number of trials, prob: probability of success
}
\description{
create a object named binvar
}
