% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
An object of class "binvar"
}
\description{
generates a binomial random variable object
}
\examples{
Bulid a binomial variable
bin1 <- bin_variable(trials = 10, prob = 0.3)
Summarize a binomial variable with statistics
summary(bin1)
}
