% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_funcs.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
an object of class binvar that contains a list of trials and prob
}
\description{
Creates a binomial random variable
}
\examples{
bin_variable(10, 0.3)
}
