% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability if success}
}
\value{
an object of class "binvar", that is, a binomial random variable object
}
\description{
compute the binomial Variable
}
\examples{
bin1 <- bin_variable(trials = 10, p = 0.3)
bin1
}
