% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{which must be an integer or vector of integers, greater than zero}

\item{prob}{which is a number between 0 and 1}
}
\value{
outputs your number of trials, and the probability of success in a formatted way
}
\description{
function which allows you to format your trials and prob, as well as summarizing using the print() or summary(), which will give you more information on your binomial distribution
}
\examples{
bin_variable(trials = 5, prob = 0.5)
bin_variable(10, 0.3)
bin_variable(1:10, 0.4)
}