% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a list object of class binvar
}
\description{
make a list of class binvar with number of trials and probability of success
}
\examples{
bin_variable(trials = 10, prob = 0.3)
}
