% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial random variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on rach trial}
}
\value{
a list containing binomial random variables
}
\description{
store binomial random variables
}
\examples{
var1 = bin_variable(5,0.5)

}
