% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{prob}{probablity of success}

\item{n}{number of trials}
}
\value{
an object of class "binvar"
}
\description{
creates a binomial random variable object with attributes "trials" and "prob"
}
\examples{
# Creates a binomial random variable object with 12 trials and a probability of success for each trial of .1
bin1 <- bin_variable(trials = 12, p = 0.1)
}
