% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{n}{number of trials, prob probablity of success in each trial}
}
\value{
an object of class binvar represented by a list
}
\description{
creates a binomial random variable object with stored values of trials n and probability prob
}
\examples{
# Creates a binomial random variable object with 10 trials and a probability of success for each trial of .3
bin1 <- bin_variable(trials = 10, p = 0.3)
}
