% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_variable}
\alias{bin_variable}
\title{binominal variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{how many trials}

\item{prob}{probability of a success in any given trial}
}
\value{
object of class "binvar", list with named elements trials and prob
}
\description{
gives data frame with the corresponding probability distribution with cumulative stats
}
\examples{
bin_variable(5, 0.5)
}
