% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial random variable object}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{(numeric) the number of trials}

\item{prob}{(numeric) the probability of success on each trial}
}
\value{
An object of class \code{"binvar"}
}
\description{
Creates an object of class \code{"binvar"}
}
\examples{
# a binomial random variable with 20 trials and a 45\% chance of success on each trial
bin_variable(20, 0.45)
}
