% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial random variable}
\usage{
bin_variable(trials, probs)
}
\arguments{
\item{trials}{: This is the number of trials}

\item{probs}{: This is the probability of success}
}
\value{
an object of class "binvar"
}
\description{
This verifies if the trials and the probability are valid
and returns an object of class "binvar"
}
\examples{
#build a binomial distribution
bin1 <- bin_variable(trials = 10, p = 0.3)
#summarizes it
binsum1 <- summary(bin1)
#print the summary
binsum1
}
