% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{prob}{probablity of success in each trial}

\item{n}{number of trials}
}
\value{
a binomial random variable object
}
\description{
calculates the binomial random variable
}
\examples{
bin_variable(5, 0.5)
}
