% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin variable}
\usage{
bin_variable(n, p)
}
\arguments{
\item{n}{trials}

\item{p}{prob}
}
\value{
object of class "binvar"
}
\description{
calculates the variable
}
