% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binary Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{Trials}

\item{prob}{Probability}
}
\value{
A list with named elements of class "binvar"
}
\description{
Constructs a list with named elements trials and prob.
}
