% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomal.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable Function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
A binomial random variable object
}
\description{
Creates an object that is a binomial variable object with class "binvar"
}
\examples{
# Create a binomial variable
bin_variable(trials = 5, prob = 0.5)
# Get summary statistics of the binomial variable
summary_var <- summary(var)
# Print summary statistics
summary_var
}
