% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
object of class "binvar"
}
\description{
creates object of class "binvar"
}
\examples{
bin_variable(5, 0.3)
}
