% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
binomial random variable
}
\description{
make a binomial random variable
}
\examples{
bin_variable(trials = 10, prob = 0.3)
}
