% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on each trial}
}
\value{
a list with named elements trials and prob
}
\description{
create a list with named elements trials and prob
}
\examples{
# the binomial probability of 5 trials
# assume the probability of success in each trial is 0.5
bin_variable(trials =5, prob = 0.5)
}
