% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{A intiger that is the number of trials}

\item{prob}{The probability of getting a success, must be between 0 and 1}
}
\value{
A binomial random variable object of class binvar
}
\description{
Return a list of the number of trials and the probability of success
}
