% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of event}
}
\value{
Returns an object of class "binvar", that is, a binomial random variable object.
}
\description{
A Binomial Random Variable, can be used with print() and summary()
}
\examples{
# default
k <- bin_variable

# print
k

# summary
k.summary <- summary(k)

# print summary
k.summary

}
