% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-variable.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variables}
\usage{
bin_variable(n, p)
}
\arguments{
\item{n}{the number of total trials}

\item{p}{the probability of single success}
}
\value{
a list consisting of the number of trials and the probability of single success
}
\description{
form a binomial object
}
