% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{parameter p to the binomial distribution}
}
\value{
object of class binvar
}
\description{
generates an object of class binvar, a list with named elements trials and prob
}
\examples{

# binvar object
bin_variable(trials = 5, prob = 0.5)

}
