% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{is the number of trials}

\item{prob}{is the probability of success}
}
\value{
a binomial random variable object
}
\description{
calculates a binomial random variable object
}
\examples{
bin_variable(trials = 10, prob = 0.3)
}
