% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
list, a list of the number of trials and probability of success associated with the random variable
}
\description{
this function describes the random variable
}
\examples{
bin_variable(trials = 10, prob = 0.2)
}
