% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workout3.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{an integer with number of trials}

\item{prob}{the probability of the success occur}
}
\value{
a 'binvar' class object containing the number of trials and the probability of success
}
\description{
the function used to generate a binomial variable with class 'binvar'
}
\examples{
bin_variable(10, 0.5)
}
