% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{numeric vector}

\item{prob}{numeric vector}
}
\value{
binvar object
}
\description{
binomial random variable object
}
\examples{

bin_variable(trials = 3, prob = 0.4)


bin1 <- bin_variable(trials = 10, prob = 0.3)


}
