% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of trials (numeric)}

\item{prob}{the probability of one success (numeric)}
}
\value{
a binomial random variable object (binvar)
}
\description{
creates a binvar object with the given parameters
}
\examples{
bin1 <- bin_variable(5, 0.5)
binsum1 <- summary(bin1)
binsum1
}
