% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probabililty of success (numeric)}
}
\value{
list with trials and prob
}
\description{
prints parameters of the binomial variable
}
\examples{
bin_variable(10, 0.3)
}
