% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{input number of trials}

\item{prob}{input probability}
}
\value{
a list with named elements trials and prob and class "binvar"
}
\description{
make an object class "binvar"
}
\examples{
bin_variable(5,0.5)
}
