% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success}
}
\value{
binvar object
}
\description{
makes a binomial random variable
}
\examples{
: bin_variable(trials = 10, prob = 0.3)
}
