% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_var.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{integer representing n, the total number of trials}

\item{prob}{float representing p, the probability of success on a given trial}
}
\value{
a named list of class binvar, containing the number of trials and the probability
}
\description{
Creates an object of class binvar, a list containing the number of trials and the probability of success on each trial
}
\examples{
bin_variable(10,.5)

bin_variable(100,2/3)
}
