% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial random variable object}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{numeric value for number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
computed binomial random variable object
}
\description{
computes the cumulative binomial distribution given number of trials and probability of success
}
\examples{
# bin1 <- bin_variable(trials = 10, p = 0.3)
bin1

# another one
# bin1 <- bin_variable(trials = 10, p = 0.3)
binsum1 <- summary(bin1)
binsum1
}
