% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{trials}

\item{prob}{probability}
}
\value{
output
}
\description{
Generates a binomial random variable with given trials and probability of success
}
\examples{
bin_variable(5, 0.5)
}
