% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_var.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_var}
\usage{
bin_variable(trials, prob)
}
\value{
list with number of trials and probability of success
}
\description{
returns binomial random variable
}
\examples{
#Create a binomial distribution
bin <- bin_variable(10,0.3)
# summarize
summary(bin)
}
