% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{the probability of success in each trial}
}
\value{
return an object of class"binvar", a binomial random variable
}
\description{
Create a binomial random variable object based on number of trials and probability of success
}
