% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable, variable creator}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
Binomial Random Variable object (named binvar) that represents a binomial random variable
}
\description{
Creates a Binomial Random Variable given number of trials and prob
}
\examples{
bin_variable(trials = 10, prob = 0.3)
}
