% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variabable.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of tirals}

\item{prob}{the probability of a single success}
}
\value{
a list that contains the number of tirals and the probability of a single success
}
\description{
return the binomial variables
}
\examples{

 bin1 <- bin_variable(trials = 10, p = 0.3)

}
