% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_bin_func.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{some integer}

\item{prob}{some real number}
}
\value{
a list with named elements, trials and prob
}
\description{
gives the parameters of the inputs
}
\examples{
bin_variable(trials=5, prob=0.5)
bin_variable(7, 0.4)
}
