% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob, class = "binvar")
}
\arguments{
\item{trials}{numeric, number of trials}

\item{prob}{numeric, probability of success}
}
\value{
a list of class binvar
}
\description{
gets us a list with named elements (number of trials and probability of success)
}
\examples{
bin_variable(trials = 10, prob = 0.3)


}
