% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable Function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a list with named elements: trials, prob
}
\description{
Shows the binomial random variables
}
\examples{
# 10 trials with a probability 0.3
bin_variable(trials = 10, prob = 0.3)
$trials
[1] 10

$prob
[1] 0.3
}
