% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_var}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a binvar object
}
\description{
creates a binary variable (binvar) object
}
\examples{

# a binary variable of tossing a fair coin five times to get heads
bin_variable(5, 0.5)

# a binary variable of tossing a loaded coin with 25\% chance of heads ten times to get heads
bin_cumulative(10, 0.25)

}
