% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Function.R
\name{bin_variable}
\alias{bin_variable}
\title{function bin_variable()}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the total number of the trials}

\item{prob}{the probability of success every time}
}
\value{
an object of class \code{"binvar"}
}
\description{
return a binomial random variable object
}
\examples{
bin_variable(trials = 10, p = 0.3)
}
