% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main function.R
\name{bin_variable}
\alias{bin_variable}
\title{: function bin_variable()}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{:}{trials and prob}
}
\value{
: an object of class \code{"binvar"}
}
\description{
: return a binomial random variable object
}
\examples{
:
bin_variable(trials = 10, p = 0.3)
}
