% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{probability of success on each trial}
}
\value{
an object of class \code{"binvar"}
}
\description{
takes in trials and prob and returns a list with trials: 
number of trials and prob: probability of success.
}
\examples{
bin_variable(trials = 10, p = 0.3)
}
