% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a list of binomial random objects
}
\description{
generate a list of binomial random objects
}
\examples{
bin_variable(2,.5)
summary(bin_variable(2,.5))
}
