% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial variable function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on each trial}
}
\description{
returns a list of trials and the probability of success
}
\examples{
bin_variable(trials = 10, prob = 0.3)
}
