% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_main.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
return a list
}
\description{
list the class binomial random variable with the named of trials and prob
}
\examples{
bin_variable(trials = 5, prob = 0.5)
}
