% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{numeric value of number of trials}

\item{prob}{numeric value of probability of success of each trial}
}
\description{
this function returns a list with named elements trials and prob
}
\examples{
Generate a list of the number of trials and the probability of success.
bin_variable(trials = 5, prob = 0.5)
}
