% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{a random variable with a number of trials and a probability of success}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of trials, a non-negative integer}

\item{prob}{the probability of success, a number between 0 and 1 both included}
}
\value{
a list of class binvar and containing a numeric vector of trials and numeric vector of probability
}
\description{
creates an object of class binvar
}
\examples{
# the binomial variable with 6 trials and probability of success 0.45
bin_variable(6,0.45)

# the binomial variable with 3 trials and probability of success 0.67
bin_variable(3,0.67)
}
