% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{an integer}

\item{prob}{a numeric value}
}
\value{
created object
}
\description{
Create an object of class *binvar*
}
\examples{
bin_variable(trials = 5, prob = 0.5)
bin_variable(100, 0.75)
}
