% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable Function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (integer)}

\item{prob}{probability of making success in one trial (decimal)}
}
\value{
a list of class "binvar" which contains the numbers of trials and the corresponding success probability
}
\description{
create a binomial random variable object wih specified number of trials and success probability
}
\examples{
bin_variable(trials = 10, p = 0.3)
}
