% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable}
\usage{
bin_variable(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a list with named elements trials and prob
}
\description{
to create an ojbect of class binvar
}
\examples{
bin_variable(trials = 10, prob = 0.3)
}
