% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_var_and_related.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{representing # of trials (numeric value)}

\item{prob}{representing probability (numeric value)}
}
\value{
a list of class \code{"binvar"}
}
\description{
prints a list of # of trials and probability for a random variable
}
\examples{
bin_variable(5,0.5)
bin_variable(1,1)
}
