% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success for each trials}
}
\value{
an object of class "binvar", that is, a binomial random variable object.
}
\description{
This is a main function that takes two arguments: trials and prob.The function should return a binomial random variable object
}
