% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial random variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials:}{number of trials, a non-negative integer}

\item{prob:}{probability of success, a number between 0 and 1, inclusive}
}
\value{
list of trials and prob
}
\description{
returns a list with the named elements trials and prob
}
\examples{
> bin_variable(trials = 5, prob = 0.5)
}
