% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials = c(trials), prob = c(prob))
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
list with named elements "trials" and "prob" of class "binvar"
}
\description{
creates a binomial random variable list of the elements
trials and probability
}
