% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_variable}
\alias{bin_variable}
\title{Function bin_variable()}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability of success}
}
\value{
the variable of binimial distribution
}
\description{
dispaly the variable of binimial distribution
}
\examples{
bin_variable(3,0.5)
}
