% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunction.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomal variable}
\usage{
bin_variable(n, p)
}
\arguments{
\item{n}{the number of trials}

\item{p}{the probability of successes}
}
\value{
create a summary of a binomal random variable
}
\description{
summarize a binomal random variable
}
\examples{
# Bulid a binomial variable
 bin1 <- bin_variable(trials = 10, p = 0.3)
# Summarize a binomial variable
binsum1 <- summary(bin1)
# Print the statistics
binsum1
}
