% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{int}{integer trials}

\item{int}{integer prob}
}
\value{
binomial random variable object
}
\description{
returns an object with a list of named elements
}
\examples{
bin_variable(trials = 10, p = 0.3)
}
