% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{: bin_variable()}
\usage{
bin_variable(trials, prob)
}
\value{
: input trials and prob
}
\description{
: check numbers of input trials and probability and shows input trials and prob
}
