% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial random variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
an object of class \code{"binvar"}
}
\description{
creates an object of class \code{"binvar"}
}
\examples{
bin_variable(trials = 5, prob = 0.5)
}
