% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{bin_variable}
\alias{bin_variable}
\title{trials and success}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{denotes trials}

\item{succeses}{denotes succeses}
}
\value{
values of trials and success
}
\description{
Creates an object of trials and success
}
\examples{
bin7 <- bin_variable()


}
