% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{numb of trials}

\item{prob}{prob of success}
}
\value{
binvar object
}
\description{
creates a binomial random variable
}
\examples{
# binvar with 10 trials with p=.5
bin_variable(10, .5)

# binvar with 8 trials with p=.7
bin_variable(8, .7)

}
