% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of n attempts (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
a list with named elements and object class "binvar"
}
\description{
describes the given variables
}
\examples{
bin_variable(trials = 10, prob = 0.3)
}
