% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials, }{total number of trials}

\item{prob, }{chance of success}
}
\value{
a binvar object
}
\description{
makes a binomial random variable
}
\examples{
bin_variable(5,0.5)
}
