% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial random variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{numeric value, number of trials}

\item{prob}{numeric value between 0 and 1, probability of success}
}
\value{
an object of class "binvar"
}
\description{
check if trials and probability are valid
and return an object of class "binvar"
}
\examples{
#build a binomial distribution
bin1 <- bin_variable(trials = 10, p = 0.3)
#summarize it
binsum1 <- summary(bin1)
#print the summary
binsum1
}
