% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_statmanips.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Object Creator}
\usage{
bin_variable(trials = 1, prob = 1)
}
\arguments{
\item{trials}{the number of trials for the Binomial Distribution}

\item{prob}{the probability for each independent Bernoulli trial}
}
\value{
an object binvar which holds the trials value in trials and prob value in prob
}
\description{
Creates an object which holds the values of trials and prob
}
\examples{
bin_variable(5,0.5)
}
