% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
data.frame with two classes: c("bindis", "data.frame")
}
\description{
lists the number of trials and probability of success
}
\examples{

bin_variable(trials = 10, p = 0.3)
}
