% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_and_methods.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials, }{number of trials}

\item{prob, }{probablity of successes}
}
\value{
a list with named elements: trials: number of trials, prob: probability of success
}
\description{
creates a binomial random variable
}
\examples{
bin_variable(trials = 5, prob = 0.5)
}
