% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{integer value for number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
binomial random variable object
#' @examples

bin_variable(5, 0.5)
bin_variable(5, 0.2)
}
\description{
creates a binomial random variable object
}
