% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_workout3.R
\name{bin_variable}
\alias{bin_variable}
\title{}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{how many times was the experiment repeated}

\item{prob}{probability to get a success}
}
\value{
a certain probability
}
\description{
}
\examples{
}
