% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_var}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{prob}{a number}

\item{trialsa}{number}
}
\value{
returns a binvar
}
\description{
returns a data frame of probs
}
\examples{
bin1 <- bin_variable(trials = 10, p = 0.3)
}
