% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable Function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{trials number of fixed trials}

\item{prob}{probability of success in each trial}
}
\value{
A list containing the following named elements: Trials and Probability of success.
}
\description{
The function takes in two arguments and returns a list containing the elements trials 
and prob. The list will contain the number of trials and the probability of success. The returned object
will be of class binvar which is a binomial variable oject. The function will also check whether the inputs
are valid and will return an error if they are not.
}
\examples{
bin_variable(trials = 5, prob = 0.5)
}
