% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials; non-negative integer value}

\item{prob}{probability of success; must be a numeric value between 0 and 1 (inclusive)}
}
\value{
list of class "binvar" (binomial random variable)
}
\description{
Returns a binomial random variable list displaying the number of trials and probability of success
}
\examples{
bin1 <- bin_variable(10, 0.3)
bin1

# display summary of bin_variable

summary(bin1)
}
