% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{for the time of success}

\item{prob}{for the probability of success}
}
\value{
it returns an object of class "binvar"
}
\description{
it invokes check funtion
}
\examples{
bin_variable(5,0.5)
}
