% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
An object of binomial variable
}
\description{
Creates an object of binomial variable
}
\examples{
# Create a binomial variable
var <- bin_variable(trials = 5, prob = 0.5)
# Get summary statistics of the binomial variable
sum_var <- summary(var)
# Print summary statistics
sum_var

}
