% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{accepts}{trials, probability}
}
\value{
returns "binvar" object
}
\description{
creates a variable with trials and probability as features
}
