% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{- number of trials}

\item{prob}{- probability of success}
}
\value{
A simple description of the binomial random variable.
}
\description{
Binomial variable
}
\examples{
bin_variable(trials = 10, p = 0.3)

}
