% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-functions.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success on each trial (numeric)}
}
\value{
list with number of trials & probability of success
}
\description{
return list with number of trials (trials) & probability of success (prob), of class "binvar"
}
\examples{
bin_variable(trials = 10, prob = 0.3)
}
