% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{probability of success in a single trial}
}
\value{
binvar: binvar object
}
\description{
Creates and returns an object of "binvar"
}
\examples{
:
bin_variable(trials = 5, prob = 0.5)
}
