% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_main.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{Number of trials (numeric)}

\item{prob}{Probability of successs (real)}
}
\value{
Binomial Random Variable object
}
\description{
An object which represents a binomial random variable
}
\examples{
bin1 <- bin_variable(trials = 10, p = 0.3)
bin1
"Binomial variable"
Paramaters
Number of Trials: 10
Probability of Success : 0.3

binsum1 <- summary(bin1)
binsum1

"Summary Binomial"
Parameters
Number of Trials: 10
Probability of Successs : 0.3
Measures
Mean : 3
Variance: 2.1
Mode: 3
Skewness: 0.2760262
Kurtosis: -0.1238095
}
