% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R, R/hello.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)

bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability that a trial will be a success}

\item{trials}{number of trials}

\item{prob}{probability that a trial will be a success}
}
\value{
A binomial random variable object

A binomial random variable object
}
\description{
Finds the binomial variable of a binomial distribution

Finds the binomial variance of a binomial distribution
}
\examples{
bin_variable(5, .5)
bin_variable(5, .5)
}
