% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-functions-methods.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
object of class "binvar", a binomial random variable
}
\description{
Creates a binomial random variable object given number of trials and probability
}
\examples{
# print bin_variable
bin1 = bin_variable(trials = 10, p = 0.3)
bin1
}
