% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (n)}

\item{prob}{probability (p)}
}
\value{
list of number of trials and probability of success
}
\description{
get binomial random variable object
}
\examples{
bin_variable(10, 0.3)
}
