% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bino.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
an object of class \code{"binvar"}
}
\description{
prints the binomial variable with the parameters
}
\examples{
# 10 trials with probability of success 0.5
fac <- bin_choose(20, 10)

}
