% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_function.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable Function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of getting a success in a trial (numeric)}
}
\value{
object of class binvar that is a binomial random variable object ("binvar")
}
\description{
Returns an object of class "binvar"that is a binomial random variable object.
}
