% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
an object of class "binvar", a binomial random variable object
}
\description{
generate binomial random variables
}
\examples{
bin_variable(trials = 10, p = 0.3)

}
