% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
object of class binvar if trials and probability inputs are valid if probability and trials are valid
}
\description{
generates an object that is a list with named elements of trial and probability of success
}
\examples{
bin_variable(5, 0.5)
bin_variable(4, 0.1)
}
