% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
An object of class "binvar"
}
\description{
generates a binomial random variable object
}
\examples{
bin_variable(5, 0.5)
}
