% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binvar.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
binomial random variable object
}
\description{
returns a list with named elements trials and probability of
success
}
\examples{
bin_variable(trials = 10, prob = .4)
bin_variable(110, prob = .56)
}
