% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{return the binomial variables}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trails}

\item{prob}{the probability}
}
\value{
a list of trials and probability
}
\description{
return the binomial variables
}
\examples{
bin_variable(5, 0.5)
}
