% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{describe number of trials}

\item{prob}{describe probability}
}
\value{
a list containing parameter
}
\description{
stores parameter information
}
\examples{
bin_variable(3,0.3)
}
