% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial random variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials to attempt}

\item{prob}{probability of success}
}
\value{
returns a binvar variable that represents the binomial random variable
}
\description{
Creates a binomial random variable with trials and prob as parameters
}
\examples{
bin_variable(10, 0.3)
}
