% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{Number of trials in the binomial distribution (numeric)}

\item{prob}{Probability of success in the binomial distribution (numeric)}
}
\value{
A Binomial Random Variable object with attributes trials and probability
}
\description{
Create a Binomial Random Variable
}
\examples{
bin_variable(trials = 5, prob = 0.5)
}
