% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_var.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of trials.}

\item{prob}{the probability of success in any given trial.}
}
\value{
returns an object of class "binvar". A list with named elements.
}
\description{
creates a list of variables.
}
\examples{
bin1 <- bin_variable(trials = 10, p = 0.3)
binsum1 <- summary(bin1)
binsum1

bin2 <- bin_variable(30, 0.7)
bin2
}
