% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binonmial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
an object of class \code{c("binvar")} which is a list with named elements trials and prob
}
\description{
Creates an object of class \code{c("binvar", "data.frame")}
}
\examples{
# default
binvar <- bin_variable(5, 0.5)
}
