% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin variable function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (intger)}

\item{prob}{number of success (intger)}
}
\value{
an object of class "binvar" , a binomial random variable object
}
\description{
get an object of class "binvar" , a binomial random variable object
}
