% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binom_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(n, p)
}
\arguments{
\item{n}{trials}

\item{p}{probability}
}
\value{
a list of named elements 'trials' and 'prob'
}
\description{
returns a bin variable
}
