% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Max Sima .R
\name{bin_variable}
\alias{bin_variable}
\title{Function bin_variable()}
\usage{
bin_variable(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
an ojbect of class binvar
}
\description{
return an ojbect of class binvar
}
\examples{
bin_variable(trials = 10, prob = 0.3)
}
