% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials; numeric vector}

\item{prob}{probability of success; numeric vector}
}
\value{
a list with named elements: trials (a number of trials) and prob (probability of success)
}
\description{
an object of class "binvar", that is, a binomial random variable object
}
\examples{
var <- bin_variable(trials = 5, prob = 0.5)
var

summary(var)
}
