% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_variable}
\alias{bin_variable}
\title{Bin variable function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{Number of (fixed) trials in the binomial distribution.}

\item{prob}{the probability of success on each trial.}
}
\value{
a list with named elements:trials: number of trials and prob: probability of success
}
\description{
a list of binomial random variable object
}
\examples{
bin_variable(trials = 5, prob = 0.5)
}
