% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{a random vinomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{integer prob double}
}
\value{
object  binvar
}
\description{
return a random binomial variable based on the given trials and prob
}
\examples{
bin_variable(trials=5,prob=0.5)
}
