% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable()}
\usage{
bin_variable(trials = 1, prob = 1)
}
\arguments{
\item{trials}{the number of the total trials}

\item{prob}{the probability of a succeess}
}
\value{
it returns a list with class "binvar"
}
\description{
bin_variable() is a function that check the probability and trials to see if its valid.
}
\examples{
#binomial variable trials and probability
 bin_variable()
 bin_variable(trials = 10, p = 0.3)
}
