% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{value of trials}

\item{prob}{probability of one independent success}
}
\description{
obtain a binomial random variable object.(class "binvar")
}
\examples{
bin_var(trials = 5, prob = 0.5)

}
