% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial distribution function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of totoal trials (integer)}

\item{prob}{the possibility of each success (numeric)}
}
\value{
an object of class "binvar", that is, a binomial random variable object.
}
\description{
summary methods of binomial distribution
}
\examples{
bin1 <- bin_variable(trials = 10, p = 0.3)
}
