% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Random Variable Function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{n}{number of trials (numeric)}

\item{p}{success rate of each trials (numeric)}
}
\value{
a list with a list of trials and probs
}
\description{
find summary of variable and its distribution
}
\examples{
bin_variable(trials = 5, prob = 0.5)
}
