% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.R
\name{bin_variable}
\alias{bin_variable}
\title{object of class "binvar"}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability (numeric)}
}
\value{
ls (binvar)
}
\description{
return an object of class "binvar"
}
\examples{
bin_variable(trials=5,prob=0.5)
}
