% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of the trials}

\item{prob}{the probability of success on each trial}
}
\value{
an object of class \code{"binvar"}
}
\description{
Create an object of class \code{"binvar"}
}
\examples{
bin_variable(trials = 10, prob = 0.3)
}
