% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability of success in each trial}
}
\value{
an object of class 'binvar'
}
\description{
creates the binomial random variable object
}
