% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout03-Zihan-Liu.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial random variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (integer)}

\item{prob}{probability of success (numeric)}
}
\value{
a binomial random variable object of class "binvar"
}
\description{
output a binomial random variable object of class "binvar"
}
\examples{
bin_variable(trials = 10, p = 0.3)
}
