% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions and methods.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{Number of trials (numeric)}

\item{prob}{Probability of success (numeric)}
}
\value{
A Binomial Random Variable object with attributes of trials and probability
}
\description{
Calculate and form a Binomial Random Variable
}
\examples{
bin_variable(trials = 5, prob = 0.5)
}
