% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable Function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the numeric vector of total trials}

\item{prob}{the numeric vector of probability for success of one single trial}
}
\value{
returned object of a list with named elements: trials and prob
}
\description{
output an object of class "binvar" with both the probability distribution and the cumulative probabilities: sucesses in the first column, probability in the second column, and cumulative in the third column
}
\examples{
bin_variable(10, 0.3)
}
