% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability if success}
}
\value{
an object of class \code{"binvar"} with the following elements:

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\description{
compute the binomial random variable
}
\examples{
# binomial random variable with parameters 10 trials and 0.3 probability
bin_variable(trials = 10, prob = 0.3)
}
