% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3_Zimeng_Zheng.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability that success occurs}
}
\value{
a binomial random variable object
}
\description{
function to generate a binomial variable with class binvar
}
\examples{
bin_variable(trials=5,prob=0.5)
}
