% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Bin Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability}
}
\value{
result
}
\description{
forms a binvar type list with trials and probability
}
