% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
binomial random variable object
}
\description{
to create a binomial random variable object
}
\examples{
bin1 <- bin_variable(trials = 10, p = 0.3)
bin1
binsum1 <- summary(bin1)
binsum1

}
