% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
A Binomial Random Variable
}
\description{
Compute Binomial Variable
}
\examples{
bin1 <- bin_variable(trials = 5, prob = 0.5)
summary(bin1)

}
