% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All Functions.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability value (numeric)}
}
\value{
binomial random variable project
}
\description{
create a binomial random variable object
}
\examples{
bin_variable(5, 0.5)
}
