% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number indicating the number of trials}

\item{prob}{number indicating teh probability of getting success}
}
\value{
a binvar object
}
\description{
get a binvar object with the provided trials and prob
}
\examples{
bin_variable(trials = 10, p = 0.3)
}
