% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-adele-shen.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
variance (numeric)
}
\description{
Calculates the variance of a binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
