% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_of_measures.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of head probability}
}
\value{
computed variance value
}
\description{
calculates the variance of binomial random variable
}
\examples{

bin_variance(10 ,0.3)

}
