% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func_measures.R, R/workout03-ann-deng.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance()}
\usage{
bin_variance(trials, prob)

bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
double of variance

double of variance
}
\description{
returns variance of binvar with trials and prob

returns variance of binvar with trials and prob
}
\examples{
bin_variance(10, 0.3) ## returns 2.1
bin_variance(10, 0.3) ## returns 2.1
}
