% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3-divya-vijay.R
\name{bin_variance}
\alias{bin_variance}
\title{bin variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{value of probability}
}
\value{
variance summary measure
}
\description{
calculates variance for binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
