% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{Bin Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{an integer, the number of trials}

\item{prob}{an integer of the probability of success for one trial}
}
\value{
an integer, the variance
}
\description{
A function to return the variance of a binomial distribution with given trials and probability
}
\examples{
bin_variance(10,0.5)
}
