% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of random variable}
}
\value{
returns variance given by the number of trials and its probability
}
\description{
gives the variance of a binomial random variable
}
\examples{
bin_variance(10,0.3)
}
