% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Var}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{vector of number of trials (numeric)}

\item{prob}{vector of probabilities (numeric)}
}
\value{
calculated variance
}
\description{
lists binomial variance
}
