% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{Variance function}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
Variance
}
\description{
Calculates the variance
}
\examples{

#computes variance with corresponding arguments
 bin_variance(10, 0.3)

}
