% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_variance}
\alias{bin_variance}
\title{Calculate variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{integer the number of trials}

\item{prob}{the probability to success in each trials}
}
\value{
The variance of the binomial distribution
}
\description{
Computes the variance of a binomial distribution
}
\examples{
# Calculate the variance for 10 trials, each trial with
# independent probability 0.5
bin_variance(trials = 10, prob = 0.5)
}
