% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{(numeric)}

\item{prob}{(numeric)}
}
\value{
the variance value
}
\description{
the variance of the binomial distribution
}
