% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
a numeric value
}
\description{
Calculates the variance
}
\examples{
# variance of binomial random distribution with 10 trials and prob = 0.5
# bin_variance(10, 0.5)
}
