% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_variance}
\alias{bin_variance}
\title{: Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{:}{trials, prob}
}
\value{
: variance
}
\description{
: Calculates variance given a binomial distribution
}
\examples{
: bin_variance(10, 0.3) = 2.1
}
