% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{as number of trials}

\item{prob}{as probability number between 0 and 1}
}
\value{
variance as numeric value
}
\description{
calculate variance of binomial dist. (np(1-p))
}
\examples{
bin_variance(10, 0.3)
}
