% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance Function}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{success probability}
}
\value{
variance, as a numeric value.
}
\description{
Computes variance for binomial variable with given trials and prob. Uses auxiliary function.
}
\examples{
bin_variance(10, 0.3)
}
