% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_measure_functions.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the variance of a binomial variable
}
\description{
calculates the variance of a binomial variable
}
\examples{
bin_variance(trials = 10, prob = 0.3)
}
