% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
variance of binomial distribution described by parameters
}
\description{
calulates the variance of a binomial distribution
}
\examples{
bin_variance(5, 0.5)
}
