% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial-code.R
\name{bin_variance}
\alias{bin_variance}
\title{variance of binomial distribution}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
variance of the binomial distribution
}
\description{
calculates variance of a binomial distribution
}
