% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variance.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
variance
}
\description{
calculates the variance of a binomial distribution
}
