% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of total trials}

\item{prob}{probabilty of success in each trial}
}
\value{
variance
}
\description{
returns variance of binomial variable
}
\examples{
bin_variance(10, 0,3)
}
