% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{bin_variance}
\alias{bin_variance}
\title{variance of the binomial distribution}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials, }{prob, bbb}
}
\value{
variance
}
\description{
calculate the variance
}
\examples{
bin_variance(10, 0.3)
}
