% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_fun.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials = 1, prob = 1)
}
\arguments{
\item{trials}{trials}

\item{prob}{probability}
}
\value{
variance
}
\description{
calculates the variance
}
