% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_measures.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{number that is the probability of success}
}
\value{
variance, the numerical variance
}
\description{
computes variance given trials and probs
}
\examples{
#5 trials with success prob at 0.5
bin_variance(trials=5, prob=0.5)
}
