% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials, }{number of flips}

\item{prob, }{probability of heads}
}
\value{
variance
}
\description{
Finds the variance of a binomial random variable
}
