% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_func.R
\name{bin_variance}
\alias{bin_variance}
\title{Bin Variance}
\usage{
bin_variance(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability value}
}
\value{
the variance value of np(1-p)
}
\description{
Calculates the variance of the binomial distribution
}
\examples{
# variance of n=10, p=0.3
var <- bin_variance(10, 0.3)
}
