% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials = 5, prob = 0.5)
}
\arguments{
\item{trials}{# of trials}

\item{prob}{probability of success}
}
\value{
the variance
}
\description{
calculates the variance of distribution
}
\examples{

# default
bin_variance()

# another example
bin_variance(trials = 5, prob = 0.5)

}
