% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_measures.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{number of trials. Must be a positive integer.}

\item{prob}{probability of success on each trial. must be
between 0 and 1}
}
\value{
the variance of the binomial distribution
}
\description{
computes the variance of the given binomial distribution.
Variance is a representation of the average spread about the mean.
}
\examples{
bin_variance()
bin_variance(5)
bin_variance(100, .3)
}
