% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-of-measures.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{success:}{number of trials}

\item{trials:}{probability value}
}
\value{
variance using trials and probability
}
\description{
calculates variance
}
\examples{
bin_variance(10, 0.3)
}
