% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
numeric vector with variance value
}
\description{
Computes variance of the distribution given number of trials
and probability of success
}
\examples{
bin_variance(10, 0.3)
}
