% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_methods.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance Function}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success of individual trial}
}
\value{
computed variance of a binomial distribution
}
\description{
computes the expected variance of a binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
