% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(prob, trials)
}
\arguments{
\item{prob}{probability of random variable}

\item{trials}{number of trials}
}
\value{
a numeric value
}
\description{
Finds the binomial variance for a given probability and number of trials
}
\examples{
 \dontrun{
 bin_variance(prob = 0.5, trials = 5)
 }
}
