% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_measures.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials (integer)}

\item{prob}{probability of success in one trial (integer)}
}
\value{
Calculated variance
}
\description{
Calculates the variance of the distribution
}
\examples{
# distribution with 10 trials and probability 0.3
bin_variance(trials = 10, prob = 0.3)
}
