% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability}
}
\value{
a binomial variance
}
\description{
compute the auxiliary variance
}
