% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunctions.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the variance of a binomal distribution
}
\description{
calculates the variance of a binomal distribution
}
\examples{
What is the variance of number of successes in 10 trials with prob of success = 0.3?
bin_variance(10, 0.3)
}
