% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{Number of trials in the binomial distribution}

\item{prob}{Probability of success in the binomial distribution}
}
\value{
the variance of binomial distribution
}
\description{
Compute the variance of the binomial distribution
}
\examples{
bin_variance(10,0.3)
}
