% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_variance}
\alias{bin_variance}
\title{variance of a binomial distribution}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on rach trial}
}
\value{
variance of a binomial distribution
}
\description{
calculate the variance of a binomial distribution
}
\examples{
bin_variance(10, 0.3)

}
