% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{n}{number of trials, prob probablity of success in each trial}
}
\value{
the variance of the given binomial distribution with n trials and probability prob of success for each trial
}
\description{
calculates the variance of the number of successes given a number of trials and probability of success for each trial
}
