% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{n}{number of trials}
}
\value{
the variance of the given binomial distribution with "trials "trials and probability "prob" of success
}
\description{
finds the variance of a binomial distribution with "trials" and probability "prob"
}
