% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns_of_measures.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
Calculates the variance of a binomial variable
}
\description{
Calculates the variance of a binomial variable
}
\examples{
bin_variance(trials = 5, prob = 0.5)

}
