% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_variance}
\alias{bin_variance}
\title{binominal variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{how many trials}

\item{prob}{probability of success on any given trial}
}
\value{
variance of the specified binomal distribution
}
\description{
gives variance of the binomial
}
\examples{
bin_variance(6, 0.3)
}
