% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, probs)
}
\arguments{
\item{trials}{This is the number of trials}

\item{probs}{This is the probability of success}
}
\value{
The variance is returned
}
\description{
This calculates the variance of the binomial experiment
}
\examples{
bin_variance(trials=5,prob=0.5)
}
