% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{Bin Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{prob}{probablity of success in each trial}

\item{n}{number of trials}
}
\value{
the variance of the binomial distribution
}
\description{
calculates the variance of the binomial distribution
}
\examples{
bin_variance(10, 0.2)
}
