% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binvar.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials, }{number of trials of the binomial distribution}

\item{prob, }{probability of succes of the binomial distriburion}
}
\value{
The binomial variance of given number of trials and probability of success
}
\description{
Calculate the variance for a binomial distribution
}
\examples{
bin_variance(12, 0.7)
bin_variance(5, 0.23)
}
