% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials.}

\item{prob}{probability of success.}
}
\value{
variance of binomial distribution
}
\description{
calculate variance of binomial distribution
}
\examples{

bin_variance(trials = 5, prob = 0.5)
bin_variance(trials = 10, prob = 0.3)

}
