% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{Binary Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{Trials}

\item{prob}{Probability}
}
\value{
A number for variance
}
\description{
Returns variance of binomial variable
}
