% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomal.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance Function}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
Calculates the variance of a binomial distribution
}
\description{
Calculates the variance of a binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
