% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_variance}
\alias{bin_variance}
\title{Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
variance of binomial(trials, prob)
}
\description{
calculates the variance of binomial(trials, prob)
}
\examples{
: bin_variance(trials = 10, prob = 0.3)
}
