% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success on each trial}
}
\value{
the variance of a random binomial variable
}
\description{
calculate the variance of a random binomial variable
}
\examples{
# the variance of 5 trials with the probability of success in each trial = 0.5
bin_variance(5, 0.5)
}
