% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-measure.R
\name{bin_variance}
\alias{bin_variance}
\title{variance of a binomial distribution}
\usage{
bin_variance(n, p)
}
\arguments{
\item{n}{the number of total trials}

\item{p}{the probability of single success}
}
\value{
the calculated variance
}
\description{
calculate the variance of a random variable following the binomial distribution
}
