% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{is the number of trials}

\item{prob}{is the probability of success}
}
\value{
the variance of a binomial distribution
}
\description{
calculates the variance of a binomial distribution
}
\examples{
bin_variance(trials = 10, prob = 0.3)
}
