% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a number, the variance of a binomial random variable
}
\description{
Calculates the variance of a binomial distribution
}
\examples{
bin_variance(trials = 10, prob = 0.5)
}
