% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workout3.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{an integer with number of trials}

\item{prob}{the probability of the success occur}
}
\value{
the variance value
}
\description{
Calculate the variance of binomial distribution
}
\examples{
bin_variance(10, 0.5)
}
