% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_functions.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success (numeric)}
}
\value{
computed binomial variance
}
\description{
computes the variance of the binomial variable
}
\examples{
bin_variance(trials = 10, prob = 0.5)
}
