% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{input number of trials}

\item{prob}{input probability}
}
\value{
computed variance of the binomial distribution
}
\description{
calculate variance of the binomial distribution
}
\examples{
bin_variance(5,0.5)
}
