% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{: bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{:}{trials the number of trials}

\item{:}{prob the probability of success}
}
\value{
: the variance of the binomial(trials, prob)
}
\description{
: calculates the variance of the binomial with trials number of trials and probability of succeess prob
}
\examples{
: bin_variance(trials = 10, prob = 0.3)
}
