% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{integer representing n, the total number of trials}

\item{prob}{float representing p, the probability of success on a given trial}
}
\value{
variance of the distribution
}
\description{
Calculates the variance of a binomial distribution with given trials and probability
}
\examples{
bin_variance(10,.3)

bin_variance(100,2/3)
}
