% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{numeric value for the number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
computed binomial variance
}
\description{
computes the variance of the binomial experiment
}
\examples{
bin_variance(trials = 10, prob = 0.3)
}
