% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{trials}

\item{prob}{probability}
}
\value{
binvariance
}
\description{
Calculates varaince of a binomial distribution of a given probability in n trials
}
\examples{
bin_variance(5, 0.5)
}
