% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variance.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the variance of the input
}
\description{
first check if input is valid, then use aux_variance to create variance
}
\examples{
bin_variance(5,0.5)
}
