% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials:}{the number of trials.}

\item{prob:}{probability of success}
}
\value{
numeric
}
\description{
gives the variance of a binomial variable
}
