% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
With valid input of parameters, return the variance
}
\description{
compute the variance.
}
\examples{
bin_variance(trials = 5, prob = 0.5)
}
