% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_summary_functions.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{(integer)}

\item{prob}{(real)}
}
\value{
variance of the expected value
}
\description{
computes variance of a binomial distribution
}
\examples{
bin_variance(trials = 5, prob = 0.5)
bin_variance(10, 0.3)
}
