% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{Variance of a Binomial Distribution}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{total numbeer of trials (numeric)}

\item{prob}{the probability of a success in a trial (numeric)}
}
\value{
a number indicates the variance of the binomial distribution
}
\description{
Calculating the variance of a binomial distribution based on inputs, trials and probability
}
\examples{
bin_variance(5,0.5)
1.25
bin_variance(6,0.3)
1.26
}
