% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_variance}
\alias{bin_variance}
\title{Variance of a Binomial Random Variable}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability for success in each trial}
}
\value{
the variance of the binomial random variable
}
\description{
calculates the variance of a binomial random variable given number of trials and prob
}
