% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variance.R
\name{bin_variance}
\alias{bin_variance}
\title{variance of binomial distribution}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of a single success}
}
\value{
calculated variance of binomial distribution
}
\description{
calculate the variance of binomial distribution
}
\examples{

bin_variance(10, 0.3)

}
