% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_summ_func.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{some integer}

\item{prob}{some real number}
}
\value{
variance of the expected value of a binomial distribution
}
\description{
computes variance of a binomial distribution
}
\examples{
bin_variance(trials=5, prob=0.5)
bin_variance(10, 0.3)
}
