% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials, }{the number of trials}

\item{prob, }{the probability of a success}
}
\value{
returns a numeric variance
}
\description{
gives the variance of the binomial distribution
}
\examples{
bin_variance(trials = 5L, prob = 0.5)
bin_variance(3L, .3)
}
