% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance Function}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
computed value
}
\description{
Calculate binomial variance
}
\examples{
# 10 trials, probability of 0.3
bin_variance(10, 0.3)
2.1
}
