% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance method}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the variance of a binomial variable
}
\description{
calculates the variance of a binomial variable
}
\examples{

# variance of heads of tossing a fair coin 10 times
bin_variance(10, 0.5)

}
