% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variance.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
binomial variance
}
\description{
binomial variance
}
\examples{
bin_variance(2,.5)
}
