% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_main.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{probability of success}
}
\value{
compute the variance of a binomal distribution
}
\description{
calculate the variance of a binomal distribution
}
\examples{
bin_variance(10, 0.3)
}
