% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_function.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability value}
}
\value{
computed variance
}
\description{
calculates the variance
}
\examples{
bin_variance(10, 0.3)
}
