% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trials, a non-negative integer}

\item{prob}{the probability of success, a number between 0 and 1 both included}
}
\value{
the variance of a random variable
}
\description{
computes the variance of a random variable under a given number of trials and a given probability of success per trial
}
\examples{
# the variance of a random variable in 5 trials and probability of success 0.4
bin_variance(5,0.4)

# the variance of a random variable in 10 trials and probability of success 0.3
bin_variance(10,0.3)
}
