% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mearsures.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{an integer}

\item{prob}{a probability value}
}
\value{
variance
}
\description{
variance of a binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
