% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{trial numbers}

\item{prob}{success probability}
}
\value{
variance of the binomial distribution
}
\description{
compute variance of a binomial
}
\examples{
bin_variance(10, 0.3)
}
