% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance Function}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{total number of trials (integer)}

\item{prob}{probability of making success in one trial (decimal)}
}
\value{
the variance of the specified binomial distribution
}
\description{
compute the variance of a binomial distribution with specified total number of
trials and success probability
}
\examples{
bin_variance(10, 0.3)
}
