% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{bin_variance}
\alias{bin_variance}
\title{Bin_Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{an integer or vector of numerics - number of trials}

\item{prob}{a numeric value between 0 and 1}
}
\value{
the variance of the variable passed in
}
\description{
finds variance of binomial given trials and prob
}
\examples{
# default
var <- bin_variance(trials = 10, prob = 0.3)

}
