% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
computed variance
}
\description{
to compute the variance of a binomial distribution
}
\examples{
bin_variance(trials = 10, prob = 0.3)
}
