% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{(integer) the number of trials for your binomial distribution}

\item{prob}{(numeric) numeric value between 0 and 1, gives probability of success per trial}
}
\value{
numeric value for the variance of the binomial distribution with input parameters
}
\description{
returns the variance of the binomial distribution for the given number of trials and probability of success per trial
}
\examples{
bin_variance(20, 0.5)
bin_variance(trials = 10, prob = 0.25)
}
