% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_variance}
\alias{bin_variance}
\title{variance of a binomial}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
calculated variance
}
\description{
calculates the variance of a binomial distribution
}
