% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_summary_measures.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{representing # of trials (numeric value)}

\item{prob}{representing probability (numeric value)}
}
\value{
variance (numeric value)
}
\description{
determines variance of given probability and given number of trials
}
\examples{
bin_variance(5,0.5)
bin_variance(1,1)
}
