% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin_variance}
\alias{bin_variance}
\title{the variance of probabilities}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success in one trial}
}
\value{
the numeric result of the variance of probabilities of successes in trials
}
\description{
calculate the variance of probabilities of successes in trials
}
\examples{
bin_variance(10, 0.3)
#the variance of successes in 10 trials with 0.3 success probability in one trial

}
