% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
variance of the binomial distribution
}
\description{
main function for summary measure "variance"
}
