% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_variance}
\alias{bin_variance}
\title{Function bin_variance()}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability of success}
}
\value{
the variance of the of binimial distribution
}
\description{
calculate the variance of the of binimial distribution
}
\examples{
bin_variance(3,0.5)
}
