% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialfunction.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomal variance}
\usage{
bin_variance(n, p)
}
\arguments{
\item{n}{the number of trials}

\item{p}{the probability of successes}
}
\value{
computer the variance of a binomal distribution
}
\description{
calculate the variance of a binomal distribution
}
\examples{
bin_variance(10, 0.3)
}
