% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{int}{integer trials}

\item{int}{integer prob}
}
\value{
binomial variance
}
\description{
checks probability and trials, and then calculates the variance
}
\examples{
bin_variance(10, 0.3)
}
