% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial_Functions.R
\name{bin_variance}
\alias{bin_variance}
\title{: bin_variance()}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials:}{the number of trials}

\item{prob:}{the probability of success}
}
\value{
: calculated variance of binomial distribution
}
\description{
: variance of binomial distribution
}
