% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance function}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of n attempts (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
the variance (integer)
}
\description{
computes the variance of the distribution
}
\examples{
bin_variance(trials = 10, prob = 0.3)
}
