% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials, }{total number of trial}

\item{prob, }{chance of success}
}
\value{
variance if input are good else error
}
\description{
checks input then calcs variance
}
\examples{
bin_variance(5,0.5)
}
