% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{numeric value, number of trials}

\item{prob}{numeric value between 0 and 1, probability of success}
}
\value{
a numeric value, variance
}
\description{
computes the variance of the binomial experiment
}
\examples{
bin_variance(trials=5,prob=0.5)
}
