% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_statmanips.R
\name{bin_variance}
\alias{bin_variance}
\title{Variance of Binomial}
\usage{
bin_variance(trials = 1, prob = 1)
}
\arguments{
\item{trials}{the number of trials in the distribution}

\item{prob}{the probability of a success}
}
\value{
The variance of the binomial distribution, using aux_variance
}
\description{
Finds the variance of the binomial distribution.
}
