% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
variance of binomial ditribution
}
\description{
calculates variance of binomial ditribution
}
\examples{

bin_variance(10, 0.3)
}
