% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{integer value for number of trials}

\item{prob}{numeric value for the probability of success}
}
\value{
variance of binomial
#' @examples

bin_variance(5, 0.5)
bin_variance(10, 0.2)
}
\description{
finds the variance of the binomial
}
