% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_functions.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance function}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{probability of a success in each trial}
}
\value{
the variance of the binomial distribution
}
\description{
Function that will take in 2 arguments and return the variance of the binomial distribution.
It will also check whether the parameters entered are valid and will return an error if either variable
entered is invalid.
}
\examples{
bin_variance(10, 0.3)
}
