% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{The number of trials}

\item{prob}{The number of successes}
}
\value{
A number
}
\description{
Calculates the variance of a binomial distribution
}
