% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{Total number of trials}

\item{prob}{Probability of success}
}
\value{
Calculates the variance of a binomial variable
}
\description{
Calculates the variance of a binomial variable
}
