% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{accepts}{trials, probabilities}
}
\value{
binomial variance of trials and Probabilities
}
\description{
calculates binomial variance
}
