% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{prob}{}
}
\value{
The variance of the distribution
}
\description{
Binomial variance
}
\examples{
var <- bin_variance(5, 0.5)
vars <- bin_mean(1:5, 0.5)

}
