% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-functions.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability of success on each trial (numeric)}
}
\value{
variance of binomial distribution
}
\description{
computes variance of binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
