% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{probability of success in a single trial}
}
\value{
numeric: the variance value
}
\description{
Calculates variance given # of trials and probability of a single success
}
\examples{
:
bin_variance(trials = 5, prob = 0.5)
}
