% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_main.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{Number of trials (real)}

\item{prob}{Probability of successs (real)}
}
\value{
Returns the variance
}
\description{
Computes the variance of a binomial distribution
}
\examples{
bin_variance(10, 0.3)
2.1
}
