% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability that a trial will be a success}
}
\value{
Variance
}
\description{
Returns variance of binomial dist
}
\examples{
bin_cumulative(5, .5)
}
