% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{chance of success}
}
\value{
the variance of a binomial with given parameters
}
\description{
Computes the variance of a binomial
}
\examples{
# variance of 10 trials with p=.5
bin_variance(10, .5)

# variance of 8 trials with p=.7
bin_variance(8, .7)

}
