% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of random trials}

\item{prob}{probability of successes}
}
\value{
variance of a binomial distribution
}
\description{
generate the variance of a binomial distrbution
}
\examples{
bin_variance(10, 0.3)

}
