% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
variance if probability and trials inputs are valid. If not, throws an error.
}
\description{
calculates variance of having certain trials and certain probability
}
\examples{
bin_variance(5, 0.6)
bin_variance(6, 0.7)
}
