% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{return the variance of the binomial distribution}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trails}

\item{prob}{the probability}
}
\value{
variance of the binomial distribution
}
\description{
return the variance of the binomial distribution
}
\examples{
bin_variance(5, 0.5)
}
