% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials to attempt}

\item{prob}{probability of success}
}
\value{
returns a the variance of the binomial distribution
}
\description{
Computes the variance of the binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
