% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_summary.R
\name{bin_variance}
\alias{bin_variance}
\title{Variance.}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trials.}

\item{prob}{the probability of success in any given trial}
}
\value{
the computed summary measure.
}
\description{
the variance is given by: np(1 − p).
}
\examples{
bin_variance(trial = 5, prob = 0.5)
bin_variance(10, 0.3)
bin_variance(100, 0.7)
}
