% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_variance}
\alias{bin_variance}
\title{Binonmial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
binomial variance
}
\description{
Creates binomial variance given "trials" and "prob"
}
\examples{
# default
variance <- bin_variance(5, .5)
}
