% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Max Sima .R
\name{bin_variance}
\alias{bin_variance}
\title{variance of binomial distribution}
\usage{
bin_variance(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
computed variance
}
\description{
calculate variance of a binomial distribution
}
\examples{
bin_variance(trials = 10, prob = 0.3)
}
