% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of (fixed) trials}

\item{prob}{the probability of success on each trial}
}
\value{
the variance of the binomial distribution
}
\description{
calculates the variance of the binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
