% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{success rate of each trials}
}
\value{
Return the variance of binomial distribution
}
\description{
Calculates the variance of a binomial variable
}
\examples{
bin_variance(trials = 5, prob = 0.5)
}
