% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{variance of binomial}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{integer prob double}
}
\value{
variance double
}
\description{
return variance of binomial distribution
}
\examples{
bin_variance(trials=5,prob=0.5)
}
