% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_fun.R
\name{bin_variance}
\alias{bin_variance}
\title{variance of a binomial distribution}
\usage{
bin_variance(trials = 1, prob = 1)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
variance of a binomial distribution
}
\description{
calculate the variance of a binomial distribution
}
\examples{
bin_variance(1, 1)
}
