% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(n = 10, p = 0.3)
}
\arguments{
\item{n}{value of trials}

\item{k}{value of successes}
}
\description{
to calculate the variance value of a binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
