% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance function}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trial (integer)}

\item{prob}{the probability of success (integer)}
}
\value{
the vaiance
}
\description{
Culculate the vaiance of a binomial distribution
}
