% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.R
\name{bin_variance}
\alias{bin_variance}
\title{the variance of the binomial distribution}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probability (numeric)}
}
\value{
computed variance value
}
\description{
calculate the variance of the binomial distribution
}
\examples{
bin_variance(trials=5,prob=0.5)
}
