% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability of success each trial}
}
\value{
the number of variance
}
\description{
the variance of a binominal distribution
}
\examples{
bin_variance(trials = 10, prob = 0.3)
}
