% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability of success}
}
\value{
the variance of the binomial distribution
}
\description{
calculates the variance of the binomial distribution
}
