% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{numeric number of trials}

\item{prob}{numeric number of probability}
}
\description{
Compute the binomial distribution variance
}
\examples{
bin_variance(10,0.3)
}
