% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3_Zimeng_Zheng.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trials}

\item{prob}{the probability that success occurs}
}
\value{
the variance of the binomial distribution
}
\description{
function to calculate the variance of the binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
