% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{Bin Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{probability}
}
\value{
variance
}
\description{
calculates variance of a binomial distribution
}
