% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variance.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{numeric}

\item{prob}{numeric}
}
\value{
the computed data
}
\description{
return variance given trials and prob
}
\examples{
bin_variance(10,0.3)
}
