% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials; numeric vector}

\item{prob}{probability}
}
\value{
The Variance Value
}
\description{
Compute Binomial Variance Value
}
\examples{
bin_variance(trials = 10, prob = 0.3)
}
