% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All Functions.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials (numeric)}

\item{prob}{probabiltiy value (numeric)}
}
\value{
variance
}
\description{
determine the variance
}
\examples{
bin_mean(10, 0.3)
}
