% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number indicating the number of trials}

\item{prob}{number indicating teh probability of getting success}
}
\value{
the variance
}
\description{
calculate the variance with certain trials and prob
}
\examples{
bin_variance(10, 0.3)
}
