% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker_functions.R
\name{check_prob}
\alias{check_prob}
\title{name: check_prob
description: checks if the given prob is valid
param: prob a probability
return: boolean: prob is between 0 and 1}
\usage{
check_prob(prob)
}
\description{
name: check_prob
description: checks if the given prob is valid
param: prob a probability
return: boolean: prob is between 0 and 1
}
