% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker_functions.R
\name{check_prob}
\alias{check_prob}
\title{title check_prob
description checks if given probability value is valid in a probability context
param prob representing a probability (numeric value)
return answer to "is it valid?" (TRUE or error message)}
\usage{
check_prob(prob)
}
\description{
title check_prob
description checks if given probability value is valid in a probability context
param prob representing a probability (numeric value)
return answer to "is it valid?" (TRUE or error message)
}
