% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker_functions.R
\name{check_success}
\alias{check_success}
\title{name: check_success
description: checks if the given number of successes is valid
param: success: the number of successes
param: trials: the number of trials to run
return: boolean: number of trials is non-negative}
\usage{
check_success(success, trials)
}
\description{
name: check_success
description: checks if the given number of successes is valid
param: success: the number of successes
param: trials: the number of trials to run
return: boolean: number of trials is non-negative
}
