% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker_functions.R
\name{check_success}
\alias{check_success}
\title{title check_successes
description Checks if a given success value is valid in a probability context
param success (numeric vector) representing # of successes
param trials (numeric value) representing # of trials
return answer to "is it valid?" (TRUE or error message)}
\usage{
check_success(success, trials)
}
\description{
title check_successes
description Checks if a given success value is valid in a probability context
param success (numeric vector) representing # of successes
param trials (numeric value) representing # of trials
return answer to "is it valid?" (TRUE or error message)
}
