% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker_functions.R
\name{check_trials}
\alias{check_trials}
\title{name: check_trials
description: checks if the given number of trials is valid
param: trials: number of trials
return: boolean: number of trials is positive}
\usage{
check_trials(trials)
}
\description{
name: check_trials
description: checks if the given number of trials is valid
param: trials: number of trials
return: boolean: number of trials is positive
}
