% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker_functions.R
\name{check_trials}
\alias{check_trials}
\title{title check_trials
description Checks if given trial value is valid in a probability context
param trials (numeric value) representing a given number of trials
return answer to "is it valid?" (TRUE or error message)}
\usage{
check_trials(trials)
}
\description{
title check_trials
description Checks if given trial value is valid in a probability context
param trials (numeric value) representing a given number of trials
return answer to "is it valid?" (TRUE or error message)
}
