% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bincum}
\alias{plot.bincum}
\title{Plot of cumulative binomial distribution}
\usage{
\method{plot}{bincum}(x)
}
\arguments{
\item{x}{an object of class \code{"bincum"}}

\item{prob}{probability of random variable}

\item{trials}{number of trials}
}
\description{
Plots a line for the cumulative probability over number of successes
}
\examples{
 \dontrun{
 # plot cumulative distribution of a binomial random variable
 prob <- 0.5
 trials <- 5
 cdf <- bin_cumulative(prob, trials)
 plot(cdf)
 }
}
