% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-cumulative.R
\name{plot.bincum}
\alias{plot.bincum}
\title{binomial cumulative distribution plot}
\usage{
\method{plot}{bincum}(y, ...)
}
\arguments{
\item{...}{arguments from other functions}

\item{x}{a bincum dataframe recording the cumulative distribution information}
}
\value{
the cumulative distribution graph
}
\description{
plot the cumulative distribution graph
}
\examples{
plot(bin_cumulative(10, 0.2))
}
