% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cum_and_related.R
\name{plot.bincum}
\alias{plot.bincum}
\title{title plot.cum
description plots a line graph with points of a cumulative distribution
param dist an object of class "bincum"
return a line graph with points}
\usage{
\method{plot}{bincum}(cum)
}
\description{
title plot.cum
description plots a line graph with points of a cumulative distribution
param dist an object of class "bincum"
return a line graph with points
}
