% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{plot.bincum}
\alias{plot.bincum}
\title{plots cumulative binomial distribution}
\usage{
\method{plot}{bincum}(bincum)
}
\description{
graphs the cumulative distribution of an object
}
\examples{
> dis2 <- bin_cumulative(trials = 5, prob = 0.5)
> plot(dis2)
}
