% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.bincum}
\alias{plot.bincum}
\title{plot.bincum}
\usage{
\method{plot}{bincum}(dis)
}
\arguments{
\item{dis}{(a bincum)}
}
\value{
a ggplot
}
\description{
given a cumulative distribution, plots it
}
\examples{
plot(bin_cumulative(5,0.5))
}
