% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-distribution.R
\name{plot.bindis}
\alias{plot.bindis}
\title{binomial distribution plot}
\usage{
\method{plot}{bindis}(x, ...)
}
\arguments{
\item{x}{a bindis dataframe recording the distribution information}

\item{...}{arguments from other functions}
}
\value{
the distribution graph
}
\description{
plot the distribution graph
}
\examples{
plot(bin_distribution(10, 0.2))
}
