% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-and-methods.R
\name{plot.bindis}
\alias{plot.bindis}
\title{plots binomial distribution}
\usage{
\method{plot}{bindis}(bindis)
}
\description{
graphs a barplot to display the probability histogram of a binomial distribution object
}
\examples{
> dis1 <- bin_distribution(trials = 5, prob = 0.5)
> plot(dis1)
}
