% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.bindis}
\alias{plot.bindis}
\title{plot.bindis}
\usage{
\method{plot}{bindis}(dis)
}
\arguments{
\item{dis}{(a bindis)}
}
\value{
a ggplot
}
\description{
given a distribution, plots it
}
\examples{
plot(bin_distribution(5,0.5))
}
