% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{plot.bindis}
\alias{plot.bindis}
\title{plot.bindis()}
\usage{
\method{plot}{bindis}(x, ...)
}
\description{
graph the cumulative distribution in an object "bindis"
}
\examples{
dis1 <- bin_distribution(trials = 5, prob = 0.5)
plot(dis1)

}
