% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot.bindis}
\alias{plot.bindis}
\title{Plot binomial distribution}
\usage{
\method{plot}{bindis}(dist)
}
\arguments{
\item{dist}{a bindist variable}
}
\description{
Plots a barplot of the binomial distribution
}
\examples{
dis1 <- bin_distribution(trials = 5, prob = 0.5)
plot(dis1)
}
