% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-variable.R
\name{print.binvar}
\alias{print.binvar}
\title{binomial variable printing}
\usage{
\method{print}{binvar}(x, ...)
}
\arguments{
\item{x}{a list with the class of binvar}

\item{...}{other inputs passed from other functions}
}
\description{
print out a binomial variable
}
