% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_var_and_related.R
\name{print.binvar}
\alias{print.binvar}
\title{title print.binvar
description prints trials and probability of a particular random variable
param var an object of class "binvar"
return concatenated strings}
\usage{
\method{print}{binvar}(var)
}
\description{
title print.binvar
description prints trials and probability of a particular random variable
param var an object of class "binvar"
return concatenated strings
}
