% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.binvar}
\alias{print.binvar}
\title{print.binvar}
\usage{
\method{print}{binvar}(bin)
}
\arguments{
\item{bin}{a binvar object}
}
\value{
nothing, just prints
}
\description{
makes a binomial random variable readable output
}
\examples{
print(bin_variable(5,0.5))
}
