% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{print.binvar}
\alias{print.binvar}
\title{Print binomial random variable}
\usage{
\method{print}{binvar}(bin)
}
\arguments{
\item{bin}{binvar that represents a binomial random variable}
}
\description{
The print representation of a binomial random variable
}
\examples{
bin1 <- bin_variable(trials = 10, p = 0.3)
bin1
}
