% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.summary.binvar}
\alias{print.summary.binvar}
\title{print.summary.binvar}
\usage{
\method{print}{summary.binvar}(binsum)
}
\arguments{
\item{binsum}{a summary.binvar object}
}
\value{
nothing, just prints
}
\description{
makes a binomial random variable's summary readable output
}
\examples{
print(summary(bin_variable(5,0.5)))
}
