% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{print.summary.binvar}
\alias{print.summary.binvar}
\title{print.summary.binvar()}
\usage{
\method{print}{summary.binvar}(x, ...)
}
\description{
create a method function to nicely print the summary measures
}
\examples{
binsum1 <- summary(bin1)
binsum1

}
