% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{print.summary.binvar}
\alias{print.summary.binvar}
\title{Print binomial random variable summary}
\usage{
\method{print}{summary.binvar}(sum_bin_var)
}
\arguments{
\item{sum_bin_var}{summary.binvar that represents a binomial random variable}
}
\description{
The print representation of a binomial random variable summart
}
\examples{
bin1 <- bin_variable(trials = 10, p = 0.3)
bin_sum <- summary(bin1)
bin_sum
}
