% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-variable.R
\name{summary.binvar}
\alias{summary.binvar}
\title{binomial variable summary}
\usage{
\method{summary}{binvar}(x, ...)
}
\arguments{
\item{x}{a list with the class of binvar}

\item{...}{other inputs passed from other functions}
}
\value{
a list consisting of all the basis statistic information about the distribution
}
\description{
summary a binomial variable
}
