% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_var_and_related.R
\name{summary.binvar}
\alias{summary.binvar}
\title{title summary.binvar
description creates a list with statistical calculations of a given random variable (mean, mode, etc.)
param var an object of class "binvar"
return list of classes "sumvar" and "list"}
\usage{
\method{summary}{binvar}(var)
}
\description{
title summary.binvar
description creates a list with statistical calculations of a given random variable (mean, mode, etc.)
param var an object of class "binvar"
return list of classes "sumvar" and "list"
}
