% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.binvar}
\alias{summary.binvar}
\title{summary.binvar}
\usage{
\method{summary}{binvar}(bin)
}
\arguments{
\item{bin}{a binvar object}
}
\value{
the binvar's different measures
}
\description{
makes a binomial random variable's summary
}
\examples{
summary(bin_variable(5,0.5))
}
