% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{summary.binvar}
\alias{summary.binvar}
\title{summary.binvar()}
\usage{
\method{summary}{binvar}(x, ...)
}
\description{
create a method function to return a list of the summary output
}
\examples{
bin1 <- bin_variable(trials = 10, p = 0.3)
binsum1 <- summary(bin1)

}
