% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{summary.binvar}
\alias{summary.binvar}
\title{Binomial random variable summary}
\usage{
\method{summary}{binvar}(bin_var)
}
\arguments{
\item{bin_var}{binvar that represents a binomial random variable}
}
\description{
computes summary of the binomial random variable passed in.
}
\examples{
bin1 <- bin_variable(trials = 10, p = 0.3)
summary(bin1)
}
